api.c: Split out more functions
More X509 function testing.
X509 store function testing.
X509 lookup function testing.
