+++ tests/api.c
@@ -232,6 +232,13 @@
 #include <tests/api/test_ossl_x509_ext.h>
 #include <tests/api/test_ossl_x509_name.h>
 #include <tests/api/test_ossl_x509_pk.h>
+#include <tests/api/test_ossl_x509_vp.h>
+#include <tests/api/test_ossl_x509_io.h>
+#include <tests/api/test_ossl_x509_crypto.h>
+#include <tests/api/test_ossl_x509_acert.h>
+#include <tests/api/test_ossl_x509_info.h>
+#include <tests/api/test_ossl_x509_str.h>
+#include <tests/api/test_ossl_x509_lu.h>
 #include <tests/api/test_ossl_pem.h>
 #include <tests/api/test_tls13.h>
 
@@ -13907,601 +13914,6 @@ static int test_wolfSSL_URI(void)
 }
 
 
-static int test_wolfSSL_TBS(void)
-{
-    EXPECT_DECLS;
-#if !defined(NO_CERTS) && !defined(NO_RSA) && !defined(NO_FILESYSTEM) \
-    && defined(OPENSSL_EXTRA)
-    WOLFSSL_X509* x509 = NULL;
-    const unsigned char* tbs;
-    int tbsSz;
-
-    ExpectNotNull(x509 = wolfSSL_X509_new());
-    ExpectNull(tbs = wolfSSL_X509_get_tbs(x509, &tbsSz));
-    wolfSSL_X509_free(x509);
-    x509 = NULL;
-
-    ExpectNotNull(x509 = wolfSSL_X509_load_certificate_file(caCertFile,
-        WOLFSSL_FILETYPE_PEM));
-
-    ExpectNull(tbs = wolfSSL_X509_get_tbs(NULL, &tbsSz));
-    ExpectNull(tbs = wolfSSL_X509_get_tbs(x509, NULL));
-    ExpectNotNull(tbs = wolfSSL_X509_get_tbs(x509, &tbsSz));
-    ExpectIntEQ(tbsSz, 1003);
-
-    wolfSSL_FreeX509(x509);
-#endif
-    return EXPECT_RESULT();
-}
-
-static int test_wolfSSL_X509_verify(void)
-{
-    EXPECT_DECLS;
-#if !defined(NO_CERTS) && !defined(NO_RSA) && !defined(NO_FILESYSTEM) && \
-    defined(OPENSSL_EXTRA)
-    WOLFSSL_X509* ca = NULL;
-    WOLFSSL_X509* serv = NULL;
-    WOLFSSL_EVP_PKEY* pkey = NULL;
-    unsigned char buf[2048];
-    const unsigned char* pt = NULL;
-    int bufSz = 0;
-
-    ExpectNotNull(ca = wolfSSL_X509_load_certificate_file(caCertFile,
-        WOLFSSL_FILETYPE_PEM));
-
-    ExpectIntNE(wolfSSL_X509_get_pubkey_buffer(NULL, buf, NULL),
-        WOLFSSL_SUCCESS);
-    ExpectIntNE(wolfSSL_X509_get_pubkey_buffer(NULL, buf, &bufSz),
-        WOLFSSL_SUCCESS);
-    ExpectIntEQ(wolfSSL_X509_get_pubkey_buffer(ca, NULL, &bufSz),
-        WOLFSSL_SUCCESS);
-    ExpectIntEQ(bufSz, 294);
-
-    bufSz--;
-    ExpectIntNE(wolfSSL_X509_get_pubkey_buffer(ca, buf, &bufSz),
-        WOLFSSL_SUCCESS);
-    bufSz = 2048;
-    ExpectIntEQ(wolfSSL_X509_get_pubkey_buffer(ca, buf, &bufSz),
-        WOLFSSL_SUCCESS);
-    ExpectIntEQ(wolfSSL_X509_get_pubkey_type(NULL), WC_NO_ERR_TRACE(WOLFSSL_FAILURE));
-    ExpectIntEQ(wolfSSL_X509_get_pubkey_type(ca), RSAk);
-
-
-    ExpectNotNull(serv = wolfSSL_X509_load_certificate_file(svrCertFile,
-        WOLFSSL_FILETYPE_PEM));
-
-    /* success case */
-    pt = buf;
-    ExpectNotNull(pkey = wolfSSL_d2i_PUBKEY(NULL, &pt, bufSz));
-
-    ExpectIntEQ(i2d_PUBKEY(pkey, NULL), bufSz);
-
-    ExpectIntEQ(wolfSSL_X509_verify(serv, pkey), WOLFSSL_SUCCESS);
-    wolfSSL_EVP_PKEY_free(pkey);
-    pkey = NULL;
-
-    /* fail case */
-    bufSz = 2048;
-    ExpectIntEQ(wolfSSL_X509_get_pubkey_buffer(serv, buf, &bufSz),
-        WOLFSSL_SUCCESS);
-    pt = buf;
-    ExpectNotNull(pkey = wolfSSL_d2i_PUBKEY(NULL, &pt, bufSz));
-    ExpectIntEQ(wolfSSL_X509_verify(serv, pkey), WC_NO_ERR_TRACE(WOLFSSL_FAILURE));
-
-    ExpectIntEQ(wolfSSL_X509_verify(NULL, pkey), WC_NO_ERR_TRACE(WOLFSSL_FATAL_ERROR));
-    ExpectIntEQ(wolfSSL_X509_verify(serv, NULL), WC_NO_ERR_TRACE(WOLFSSL_FATAL_ERROR));
-
-#ifndef NO_WOLFSSL_STUB
-    ExpectNull(wolfSSL_X509_get0_pubkey_bitstr(NULL));
-    ExpectNull(wolfSSL_X509_get0_pubkey_bitstr(serv));
-#endif
-
-    wolfSSL_EVP_PKEY_free(pkey);
-
-    wolfSSL_FreeX509(ca);
-    wolfSSL_FreeX509(serv);
-#endif
-    return EXPECT_RESULT();
-}
-
-#if defined(WOLFSSL_ACERT) && !defined(NO_CERTS) && !defined(NO_RSA) && \
-    defined(WC_RSA_PSS) && !defined(NO_FILESYSTEM) && defined(OPENSSL_EXTRA)
-/* Given acert file and its pubkey file, read them and then
- * attempt to verify signed acert.
- *
- * If expect_pass is true, then verification should pass.
- * If expect_pass is false, then verification should fail.
- * */
-static int do_acert_verify_test(const char * acert_file,
-                                const char * pkey_file,
-                                size_t       expect_pass)
-{
-    X509_ACERT * x509 = NULL;
-    EVP_PKEY *   pkey = NULL;
-    BIO *        bp = NULL;
-    int          verify_rc = 0;
-
-    /* First read the attribute certificate. */
-    bp = BIO_new_file(acert_file, "r");
-    if (bp == NULL) {
-        return -1;
-    }
-
-    x509 = PEM_read_bio_X509_ACERT(bp, NULL, NULL, NULL);
-    BIO_free(bp);
-    bp = NULL;
-
-    if (x509 == NULL) {
-        return -1;
-    }
-
-    /* Next read the associated pub key. */
-    bp = BIO_new_file(pkey_file, "r");
-
-    if (bp == NULL) {
-        X509_ACERT_free(x509);
-        x509 = NULL;
-        return -1;
-    }
-
-    pkey = PEM_read_bio_PUBKEY(bp, &pkey, NULL, NULL);
-    BIO_free(bp);
-    bp = NULL;
-
-    if (pkey == NULL) {
-        X509_ACERT_free(x509);
-        x509 = NULL;
-        return -1;
-    }
-
-    /* Finally, do verification. */
-    verify_rc = X509_ACERT_verify(x509, pkey);
-
-    X509_ACERT_free(x509);
-    x509 = NULL;
-
-    EVP_PKEY_free(pkey);
-    pkey = NULL;
-
-    if (expect_pass && verify_rc != 1) {
-        return -1;
-    }
-
-    if (!expect_pass && verify_rc == 1) {
-        return -1;
-    }
-
-    return 0;
-}
-#endif
-
-static int test_wolfSSL_X509_ACERT_verify(void)
-{
-    EXPECT_DECLS;
-#if defined(WOLFSSL_ACERT) && !defined(NO_CERTS) && !defined(NO_RSA) && \
-    defined(WC_RSA_PSS) && !defined(NO_FILESYSTEM) && defined(OPENSSL_EXTRA)
-    /* Walk over list of signed ACERTs and their pubkeys.
-     * All should load and pass verification. */
-    const char * acerts[4] = {"certs/acert/acert.pem",
-                              "certs/acert/acert_ietf.pem",
-                              "certs/acert/rsa_pss/acert.pem",
-                              "certs/acert/rsa_pss/acert_ietf.pem"};
-    const char * pkeys[4] =  {"certs/acert/acert_pubkey.pem",
-                              "certs/acert/acert_ietf_pubkey.pem",
-                              "certs/acert/rsa_pss/acert_pubkey.pem",
-                              "certs/acert/rsa_pss/acert_ietf_pubkey.pem"};
-    int    rc = 0;
-    size_t i = 0;
-    size_t j = 0;
-
-    for (i = 0; i < 4; ++i) {
-        for (j = i; j < 4; ++j) {
-            rc = do_acert_verify_test(acerts[i], pkeys[j], i == j);
-
-            if (rc) {
-                fprintf(stderr, "error: %s: i = %zu, j = %zu, rc = %d\n",
-                        "do_acert_verify_test", i, j, rc);
-                break;
-            }
-        }
-
-        if (rc) { break; }
-    }
-
-    ExpectIntEQ(rc, 0);
-#endif
-    return EXPECT_RESULT();
-}
-
-static int test_wolfSSL_X509_ACERT_misc_api(void)
-{
-    EXPECT_DECLS;
-#if defined(WOLFSSL_ACERT) && !defined(NO_CERTS) && !defined(NO_RSA) && \
-    !defined(NO_FILESYSTEM) && defined(OPENSSL_EXTRA)
-    const char * acerts[4] = {"certs/acert/acert.pem",
-                              "certs/acert/acert_ietf.pem",
-                              "certs/acert/rsa_pss/acert.pem",
-                              "certs/acert/rsa_pss/acert_ietf.pem"};
-    int          rc = 0;
-    X509_ACERT * x509 = NULL;
-    BIO *        bp = NULL;
-    long         ver_long = 0;
-    int          ver = 0;
-    int          nid = 0;
-    const byte * raw_attr = NULL;
-    word32       attr_len = 0;
-    size_t       i = 0;
-    int          buf_len = 0;
-    byte         ietf_serial[] = {0x03, 0xb5, 0x90, 0x59, 0x02,
-                                  0xa2, 0xaa, 0xb5, 0x40, 0x21,
-                                  0x44, 0xb8, 0x2c, 0x4f, 0xd9,
-                                  0x80, 0x1b, 0x5f, 0x57, 0xc2};
-
-    for (i = 0; i < 4; ++i) {
-        const char * acert_file = acerts[i];
-        int          is_rsa_pss = 0;
-        int          is_ietf_acert = 0;
-        byte         serial[64];
-        int          serial_len = sizeof(serial);
-
-        XMEMSET(serial, 0, sizeof(serial));
-
-        is_rsa_pss = XSTRSTR(acert_file, "rsa_pss") != NULL ? 1 : 0;
-        is_ietf_acert = XSTRSTR(acert_file, "ietf.pem") != NULL ? 1 : 0;
-
-        /* First read the attribute certificate. */
-        bp = BIO_new_file(acert_file, "r");
-        ExpectNotNull(bp);
-
-        x509 = PEM_read_bio_X509_ACERT(bp, NULL, NULL, NULL);
-        ExpectNotNull(x509);
-
-        /* We're done with the bio for now. */
-        if (bp != NULL) {
-            BIO_free(bp);
-            bp = NULL;
-        }
-
-        /* Check version and signature NID. */
-        ver_long = X509_ACERT_get_version(x509);
-        ExpectIntEQ(ver_long, 1);
-
-        ver = wolfSSL_X509_ACERT_version(x509);
-        ExpectIntEQ(ver, 2);
-
-        nid = X509_ACERT_get_signature_nid(x509);
-
-        if (is_rsa_pss) {
-            ExpectIntEQ(nid, NID_rsassaPss);
-        }
-        else {
-            ExpectIntEQ(nid, NID_sha256WithRSAEncryption);
-        }
-
-        /* Get the serial number buffer.
-         * The ietf acert example has a 20 byte serial number. */
-        rc = wolfSSL_X509_ACERT_get_serial_number(x509, serial, &serial_len);
-        ExpectIntEQ(rc, SSL_SUCCESS);
-
-        if (is_ietf_acert) {
-            ExpectIntEQ(serial_len, 20);
-            ExpectIntEQ(XMEMCMP(serial, ietf_serial, sizeof(ietf_serial)), 0);
-        }
-        else {
-            ExpectIntEQ(serial_len, 1);
-            ExpectTrue(serial[0] == 0x01);
-        }
-
-        /* Repeat the same but with null serial buffer. This is ok. */
-        rc = wolfSSL_X509_ACERT_get_serial_number(x509, NULL, &serial_len);
-        ExpectIntEQ(rc, SSL_SUCCESS);
-
-        if (is_ietf_acert) {
-            ExpectIntEQ(serial_len, 20);
-        }
-        else {
-            ExpectIntEQ(serial_len, 1);
-            ExpectTrue(serial[0] == 0x01);
-        }
-
-        /* Get the attributes buffer. */
-        rc = wolfSSL_X509_ACERT_get_attr_buf(x509, &raw_attr, &attr_len);
-        ExpectIntEQ(rc, SSL_SUCCESS);
-
-        if (is_ietf_acert) {
-            /* This cert has a 65 byte attributes field. */
-            ExpectNotNull(raw_attr);
-            ExpectIntEQ(attr_len, 65);
-        }
-        else {
-            /* This cert has a 237 byte attributes field. */
-            ExpectNotNull(raw_attr);
-            ExpectIntEQ(attr_len, 237);
-        }
-
-        /* Test printing acert to memory bio. */
-        ExpectNotNull(bp = BIO_new(BIO_s_mem()));
-        rc = X509_ACERT_print(bp, x509);
-        ExpectIntEQ(rc, SSL_SUCCESS);
-
-        /* Now do a bunch of invalid stuff with partially valid inputs. */
-        rc = wolfSSL_X509_ACERT_get_attr_buf(x509, &raw_attr, NULL);
-        ExpectIntEQ(rc, BAD_FUNC_ARG);
-
-        rc = wolfSSL_X509_ACERT_get_attr_buf(x509, NULL, &attr_len);
-        ExpectIntEQ(rc, BAD_FUNC_ARG);
-
-        rc = wolfSSL_X509_ACERT_get_attr_buf(NULL, &raw_attr, &attr_len);
-        ExpectIntEQ(rc, BAD_FUNC_ARG);
-
-        ver_long = X509_ACERT_get_version(NULL);
-        ExpectIntEQ(ver_long, 0);
-
-        ver = wolfSSL_X509_ACERT_version(NULL);
-        ExpectIntEQ(ver, 0);
-
-        rc = wolfSSL_X509_ACERT_get_signature(x509, NULL, NULL);
-        ExpectIntEQ(rc, WOLFSSL_FATAL_ERROR);
-
-        rc = wolfSSL_X509_ACERT_get_signature(x509, NULL, &buf_len);
-        ExpectIntEQ(rc, SSL_SUCCESS);
-        ExpectIntEQ(buf_len, 256);
-
-        rc = wolfSSL_X509_ACERT_get_serial_number(x509, serial, NULL);
-        ExpectIntEQ(rc, BAD_FUNC_ARG);
-
-        rc = X509_ACERT_print(bp, NULL);
-        ExpectIntEQ(rc, WOLFSSL_FAILURE);
-
-        rc = X509_ACERT_print(NULL, x509);
-        ExpectIntEQ(rc, WOLFSSL_FAILURE);
-
-        /* Finally free the acert and bio, we're done with them. */
-        if (x509 != NULL) {
-            X509_ACERT_free(x509);
-            x509 = NULL;
-        }
-
-        if (bp != NULL) {
-            BIO_free(bp);
-            bp = NULL;
-        }
-    }
-#endif
-    return EXPECT_RESULT();
-}
-
-static int test_wolfSSL_X509_ACERT_buffer(void)
-{
-    EXPECT_DECLS;
-#if defined(WOLFSSL_ACERT) && !defined(NO_CERTS) && \
-    !defined(NO_RSA) && defined(WC_RSA_PSS) && \
-    (defined(OPENSSL_EXTRA_X509_SMALL) || defined(OPENSSL_EXTRA))
-    const byte acert_ietf[] = \
-    "-----BEGIN ATTRIBUTE CERTIFICATE-----\n"
-    "MIICPTCCASUCAQEwN6AWMBGkDzANMQswCQYDVQQDDAJDQQIBAqEdpBswGTEXMBUG\n"
-    "A1UEAwwOc2VydmVyLmV4YW1wbGWgLTArpCkwJzElMCMGA1UEAwwcQXR0cmlidXRl\n"
-    "IENlcnRpZmljYXRlIElzc3VlcjANBgkqhkiG9w0BAQsFAAIUA7WQWQKiqrVAIUS4\n"
-    "LE/ZgBtfV8IwIhgPMjAyMTA2MTUxMjM1MDBaGA8yMDMxMDYxMzEyMzUwMFowQTAj\n"
-    "BggrBgEFBQcKBDEXMBWgCYYHVGVzdHZhbDAIDAZncm91cDEwGgYDVQRIMRMwEaEP\n"
-    "gw1hZG1pbmlzdHJhdG9yMCwwHwYDVR0jBBgwFoAUYm7JaGdsZLtTgt0tqoCK2MrI\n"
-    "i10wCQYDVR04BAIFADANBgkqhkiG9w0BAQsFAAOCAQEAlIOJ2Dj3TEUj6BIv6vUs\n"
-    "GqFWms05i+d10XSzWrunlUTQPoJcUjYkifOWp/7RpZ2XnRl+6hH+nIbmwSmXWwBn\n"
-    "ERw2bQMmw/""/nWuN4Qv9t7ltuovWC0pJX6VMT1IRTuTV4SxuZpFL37vkmnFlPBlb+\n"
-    "mn3ESSxLTjThWFIq1tip4IaxE/i5Uh32GlJglatFHM1PCGoJtyLtYb6KHDlvknw6\n"
-    "coDyjIcj0FZwtQw41jLwxI8jWNmrpt978wdpprB/URrRs+m02HmeQoiHFi/qvdv8\n"
-    "d+5vHf3Pi/ulhz/+dvr0p1vEQSoFnYxLXuty2p5m3PJPZCFmT3gURgmgR3BN9d7A\n"
-    "Bw==\n"
-    "-----END ATTRIBUTE CERTIFICATE-----\n";
-    X509_ACERT * x509 = NULL;
-    int          rc = 0;
-    byte         ietf_serial[] = {0x03, 0xb5, 0x90, 0x59, 0x02,
-                                  0xa2, 0xaa, 0xb5, 0x40, 0x21,
-                                  0x44, 0xb8, 0x2c, 0x4f, 0xd9,
-                                  0x80, 0x1b, 0x5f, 0x57, 0xc2};
-    byte         serial[64];
-    int          serial_len = sizeof(serial);
-    const byte * raw_attr = NULL;
-    word32       attr_len = 0;
-
-    x509 = wolfSSL_X509_ACERT_load_certificate_buffer_ex(acert_ietf,
-                                                         sizeof(acert_ietf),
-                                                         WOLFSSL_FILETYPE_PEM,
-                                                         HEAP_HINT);
-
-    rc = wolfSSL_X509_ACERT_get_serial_number(x509, serial, &serial_len);
-    ExpectIntEQ(rc, SSL_SUCCESS);
-
-    ExpectIntEQ(serial_len, 20);
-    ExpectIntEQ(XMEMCMP(serial, ietf_serial, sizeof(ietf_serial)), 0);
-
-    /* Get the attributes buffer. */
-    rc = wolfSSL_X509_ACERT_get_attr_buf(x509, &raw_attr, &attr_len);
-    ExpectIntEQ(rc, SSL_SUCCESS);
-
-    /* This cert has a 65 byte attributes field. */
-    ExpectNotNull(raw_attr);
-    ExpectIntEQ(attr_len, 65);
-
-    ExpectNotNull(x509);
-
-    if (x509 != NULL) {
-        wolfSSL_X509_ACERT_free(x509);
-        x509 = NULL;
-    }
-#endif
-    return EXPECT_RESULT();
-}
-
-/* note: when ACERT generation and signing are implemented,
- * this test will be filled out appropriately.
- * */
-static int test_wolfSSL_X509_ACERT_new_and_sign(void)
-{
-    EXPECT_DECLS;
-#if defined(WOLFSSL_ACERT) && !defined(NO_CERTS) && \
-    !defined(NO_RSA) && defined(WC_RSA_PSS) && \
-    (defined(OPENSSL_EXTRA_X509_SMALL) || defined(OPENSSL_EXTRA))
-    X509_ACERT * x509 = NULL;
-    int          rc = 0;
-
-    x509 = X509_ACERT_new();
-    ExpectNotNull(x509);
-
-    if (x509 != NULL) {
-        wolfSSL_X509_ACERT_free(x509);
-        x509 = NULL;
-    }
-
-    /* Same but with static memory hint. */
-    x509 = wolfSSL_X509_ACERT_new_ex(HEAP_HINT);
-    ExpectNotNull(x509);
-
-    #ifndef NO_WOLFSSL_STUB
-    /* ACERT sign not implemented yet. */
-    if (x509 != NULL) {
-        rc = wolfSSL_X509_ACERT_sign(x509, NULL, NULL);
-        ExpectIntEQ(rc, WOLFSSL_NOT_IMPLEMENTED);
-    }
-    #else
-    (void) rc;
-    #endif /* NO_WOLFSSL_STUB */
-
-    if (x509 != NULL) {
-        wolfSSL_X509_ACERT_free(x509);
-        x509 = NULL;
-    }
-
-#endif
-    return EXPECT_RESULT();
-}
-
-/* Test ACERT support, but with ASN functions only.
- *
- * This example acert_ietf has both Holder IssuerSerial
- * and Holder entityName fields.
- * */
-static int test_wolfSSL_X509_ACERT_asn(void)
-{
-    EXPECT_DECLS;
-#if defined(WOLFSSL_ACERT) && !defined(NO_CERTS)
-    const byte     acert_ietf[] = \
-    "-----BEGIN ATTRIBUTE CERTIFICATE-----\n"
-    "MIICPTCCASUCAQEwN6AWMBGkDzANMQswCQYDVQQDDAJDQQIBAqEdpBswGTEXMBUG\n"
-    "A1UEAwwOc2VydmVyLmV4YW1wbGWgLTArpCkwJzElMCMGA1UEAwwcQXR0cmlidXRl\n"
-    "IENlcnRpZmljYXRlIElzc3VlcjANBgkqhkiG9w0BAQsFAAIUA7WQWQKiqrVAIUS4\n"
-    "LE/ZgBtfV8IwIhgPMjAyMTA2MTUxMjM1MDBaGA8yMDMxMDYxMzEyMzUwMFowQTAj\n"
-    "BggrBgEFBQcKBDEXMBWgCYYHVGVzdHZhbDAIDAZncm91cDEwGgYDVQRIMRMwEaEP\n"
-    "gw1hZG1pbmlzdHJhdG9yMCwwHwYDVR0jBBgwFoAUYm7JaGdsZLtTgt0tqoCK2MrI\n"
-    "i10wCQYDVR04BAIFADANBgkqhkiG9w0BAQsFAAOCAQEAlIOJ2Dj3TEUj6BIv6vUs\n"
-    "GqFWms05i+d10XSzWrunlUTQPoJcUjYkifOWp/7RpZ2XnRl+6hH+nIbmwSmXWwBn\n"
-    "ERw2bQMmw/""/nWuN4Qv9t7ltuovWC0pJX6VMT1IRTuTV4SxuZpFL37vkmnFlPBlb+\n"
-    "mn3ESSxLTjThWFIq1tip4IaxE/i5Uh32GlJglatFHM1PCGoJtyLtYb6KHDlvknw6\n"
-    "coDyjIcj0FZwtQw41jLwxI8jWNmrpt978wdpprB/URrRs+m02HmeQoiHFi/qvdv8\n"
-    "d+5vHf3Pi/ulhz/+dvr0p1vEQSoFnYxLXuty2p5m3PJPZCFmT3gURgmgR3BN9d7A\n"
-    "Bw==\n"
-    "-----END ATTRIBUTE CERTIFICATE-----\n";
-    int            rc = 0;
-    int            n_diff = 0;
-    byte           ietf_serial[] =      {0x03, 0xb5, 0x90, 0x59, 0x02,
-                                         0xa2, 0xaa, 0xb5, 0x40, 0x21,
-                                         0x44, 0xb8, 0x2c, 0x4f, 0xd9,
-                                         0x80, 0x1b, 0x5f, 0x57, 0xc2};
-    byte           holderIssuerName[] = {0x31, 0x0b, 0x30, 0x09, 0x06,
-                                         0x03, 0x55, 0x04, 0x03, 0x0c,
-                                         0x02, 0x43, 0x41};
-    byte           holderEntityName[] = {0x31, 0x17, 0x30, 0x15, 0x06,
-                                         0x03, 0x55, 0x04, 0x03, 0x0c,
-                                         0x0e, 0x73, 0x65, 0x72, 0x76,
-                                         0x65, 0x72, 0x2e, 0x65, 0x78,
-                                         0x61, 0x6d, 0x70, 0x6c, 0x65};
-    DerBuffer *    der = NULL;
-    WC_DECLARE_VAR(acert, DecodedAcert, 1, 0);
-
-    rc = wc_PemToDer(acert_ietf, sizeof(acert_ietf), ACERT_TYPE, &der,
-                     HEAP_HINT, NULL, NULL);
-
-    ExpectIntEQ(rc, 0);
-    ExpectNotNull(der);
-
-    if (der != NULL) {
-        ExpectNotNull(der->buffer);
-    }
-
-#ifdef WOLFSSL_SMALL_STACK
-    acert = (DecodedAcert*)XMALLOC(sizeof(DecodedAcert), HEAP_HINT,
-                                   DYNAMIC_TYPE_DCERT);
-    ExpectNotNull(acert);
-#else
-    XMEMSET(acert, 0, sizeof(DecodedAcert));
-#endif
-
-    if (der != NULL && der->buffer != NULL
-#ifdef WOLFSSL_SMALL_STACK
-        && acert != NULL
-#endif
-    ) {
-        wc_InitDecodedAcert(acert, der->buffer, der->length, HEAP_HINT);
-        rc = wc_ParseX509Acert(acert, VERIFY_SKIP_DATE);
-        ExpectIntEQ(rc, 0);
-
-        ExpectIntEQ(acert->serialSz, 20);
-        ExpectIntEQ(XMEMCMP(acert->serial, ietf_serial, sizeof(ietf_serial)),
-                    0);
-
-        /* This cert has a 65 byte attributes field. */
-        ExpectNotNull(acert->rawAttr);
-        ExpectIntEQ(acert->rawAttrLen, 65);
-
-        ExpectNotNull(acert->holderIssuerName);
-        ExpectNotNull(acert->holderEntityName);
-
-        if ((acert->holderIssuerName != NULL) &&
-            (acert->holderEntityName != NULL)) {
-            ExpectNotNull(acert->holderEntityName->name);
-            ExpectNotNull(acert->holderIssuerName->name);
-        }
-
-        if ((acert->holderIssuerName != NULL) &&
-            (acert->holderEntityName != NULL) &&
-            (acert->holderIssuerName->name != NULL) &&
-            (acert->holderEntityName->name != NULL)) {
-            ExpectIntEQ(acert->holderIssuerName->len,
-                        sizeof(holderIssuerName));
-            ExpectIntEQ(acert->holderEntityName->len,
-                        sizeof(holderEntityName));
-
-            ExpectIntEQ(acert->holderIssuerName->type, ASN_DIR_TYPE);
-            ExpectIntEQ(acert->holderEntityName->type, ASN_DIR_TYPE);
-
-            n_diff = XMEMCMP(acert->holderIssuerName->name, holderIssuerName,
-                             sizeof(holderIssuerName));
-            ExpectIntEQ(n_diff, 0);
-
-            n_diff = XMEMCMP(acert->holderEntityName->name, holderEntityName,
-                             sizeof(holderEntityName));
-            ExpectIntEQ(n_diff, 0);
-        }
-
-        wc_FreeDecodedAcert(acert);
-    }
-
-#ifdef WOLFSSL_SMALL_STACK
-    if (acert != NULL) {
-        XFREE(acert, HEAP_HINT, DYNAMIC_TYPE_DCERT);
-        acert = NULL;
-    }
-#endif
-
-    if (der != NULL) {
-        wc_FreeDer(&der);
-        der = NULL;
-    }
-
-#endif
-    return EXPECT_RESULT();
-}
-
 #if !defined(NO_DH) && !defined(NO_AES) && defined(WOLFSSL_CERT_GEN) && \
          defined(HAVE_SSL_MEMIO_TESTS_DEPENDENCIES) && \
          defined(OPENSSL_EXTRA) && !defined(NO_ASN_TIME)
@@ -15161,140 +14573,6 @@ static int test_wolfSSL_lhash(void)
 #endif
     return EXPECT_RESULT();
 }
-
-#ifndef NO_BIO
-static int test_wolfSSL_X509_INFO_multiple_info(void)
-{
-    EXPECT_DECLS;
-#if defined(OPENSSL_ALL) && !defined(NO_RSA)
-    STACK_OF(X509_INFO) *info_stack = NULL;
-    X509_INFO *info = NULL;
-    int len;
-    int i;
-    const char* files[] = {
-        cliCertFile,
-        cliKeyFile,
-        /* This needs to be the order as svrCertFile contains the
-         * intermediate cert as well. */
-        svrKeyFile,
-        svrCertFile,
-        NULL,
-    };
-    const char** curFile;
-    BIO *fileBIO = NULL;
-    BIO *concatBIO = NULL;
-    byte tmp[FOURK_BUF];
-
-    /* concatenate the cert and the key file to force PEM_X509_INFO_read_bio
-     * to group objects together. */
-    ExpectNotNull(concatBIO = BIO_new(BIO_s_mem()));
-    for (curFile = files; EXPECT_SUCCESS() && *curFile != NULL; curFile++) {
-        int fileLen = 0;
-        ExpectNotNull(fileBIO = BIO_new_file(*curFile, "rb"));
-        ExpectIntGT(fileLen = wolfSSL_BIO_get_len(fileBIO), 0);
-        if (EXPECT_SUCCESS()) {
-            while ((len = BIO_read(fileBIO, tmp, sizeof(tmp))) > 0) {
-                ExpectIntEQ(BIO_write(concatBIO, tmp, len), len);
-                fileLen -= len;
-                if (EXPECT_FAIL())
-                    break;
-            }
-            /* Make sure we read the entire file */
-            ExpectIntEQ(fileLen, 0);
-        }
-        BIO_free(fileBIO);
-        fileBIO = NULL;
-    }
-
-    ExpectNotNull(info_stack = PEM_X509_INFO_read_bio(concatBIO, NULL, NULL,
-        NULL));
-    ExpectIntEQ(sk_X509_INFO_num(info_stack), 3);
-    for (i = 0; i < sk_X509_INFO_num(info_stack); i++) {
-        ExpectNotNull(info = sk_X509_INFO_value(info_stack, i));
-        ExpectNotNull(info->x509);
-        ExpectNull(info->crl);
-        if (i != 2) {
-            ExpectNotNull(info->x_pkey);
-            ExpectIntEQ(X509_check_private_key(info->x509,
-                                               info->x_pkey->dec_pkey), 1);
-        }
-        else {
-            ExpectNull(info->x_pkey);
-        }
-    }
-
-    sk_X509_INFO_pop_free(info_stack, X509_INFO_free);
-    BIO_free(concatBIO);
-#endif
-    return EXPECT_RESULT();
-}
-#endif
-
-#ifndef NO_BIO
-static int test_wolfSSL_X509_INFO(void)
-{
-    EXPECT_DECLS;
-#if defined(OPENSSL_ALL) && !defined(NO_RSA)
-    STACK_OF(X509_INFO) *info_stack = NULL;
-    X509_INFO *info = NULL;
-    BIO *cert = NULL;
-    int i;
-    /* PEM in hex format to avoid null terminator */
-    byte data[] = {
-        0x2d, 0x2d, 0x2d, 0x2d, 0x2d, 0x2d, 0x2d, 0x2d, 0x2d, 0x42, 0x45, 0x47,
-        0x49, 0x4e, 0x20, 0x43, 0x45, 0x52, 0x54, 0x63, 0x2d, 0x2d, 0x2d, 0x2d,
-        0x2d, 0x0a, 0x4d, 0x49, 0x49, 0x44, 0x4d, 0x54, 0x42, 0x75, 0x51, 0x3d,
-        0x0a, 0x2d, 0x2d, 0x2d, 0x2d, 0x2d, 0x45, 0x4e, 0x44, 0x20, 0x2d, 0x2d,
-        0x2d, 0x2d, 0x2d
-    };
-    /* PEM in hex format to avoid null terminator */
-    byte data2[] = {
-        0x41, 0x53, 0x4e, 0x31, 0x20, 0x4f, 0x49, 0x44, 0x3a, 0x20, 0x70, 0x72,
-        0x69, 0x6d, 0x65, 0x32, 0x35, 0x36, 0x76, 0x31, 0x0a, 0x2d, 0x2d, 0x2d,
-        0x2d, 0x2d, 0x42, 0x45, 0x47, 0x49, 0x4e, 0x20, 0x45, 0x43, 0x20, 0x50,
-        0x41, 0x52, 0x41, 0x4d, 0x45, 0x54, 0x45, 0x52, 0x53, 0x2d, 0x2d, 0x2d,
-        0x2d, 0x43, 0x65, 0x72, 0x74, 0x69, 0x2d, 0x0a, 0x42, 0x67, 0x67, 0x71,
-        0x68, 0x6b, 0x6a, 0x4f, 0x50, 0x51, 0x4d, 0x42, 0x42, 0x77, 0x3d, 0x3d,
-        0x0a, 0x2d, 0x2d, 0x2d, 0x2d, 0x2d
-    };
-
-    ExpectNotNull(cert = BIO_new_file(cliCertFileExt, "rb"));
-    ExpectNotNull(info_stack = PEM_X509_INFO_read_bio(cert, NULL, NULL, NULL));
-    for (i = 0; i < sk_X509_INFO_num(info_stack); i++) {
-        ExpectNotNull(info = sk_X509_INFO_value(info_stack, i));
-        ExpectNotNull(info->x509);
-        ExpectNull(info->crl);
-        ExpectNull(info->x_pkey);
-    }
-    sk_X509_INFO_pop_free(info_stack, X509_INFO_free);
-    info_stack = NULL;
-    BIO_free(cert);
-    cert = NULL;
-
-    ExpectNotNull(cert = BIO_new_file(cliCertFileExt, "rb"));
-    ExpectNotNull(info_stack = PEM_X509_INFO_read_bio(cert, NULL, NULL, NULL));
-    sk_X509_INFO_pop_free(info_stack, X509_INFO_free);
-    info_stack = NULL;
-    BIO_free(cert);
-    cert = NULL;
-
-    /* This case should fail due to invalid input. */
-    ExpectNotNull(cert = BIO_new(BIO_s_mem()));
-    ExpectIntEQ(BIO_write(cert, data, sizeof(data)), sizeof(data));
-    ExpectNull(info_stack = PEM_X509_INFO_read_bio(cert, NULL, NULL, NULL));
-    sk_X509_INFO_pop_free(info_stack, X509_INFO_free);
-    info_stack = NULL;
-    BIO_free(cert);
-    cert = NULL;
-    ExpectNotNull(cert = BIO_new(BIO_s_mem()));
-    ExpectIntEQ(BIO_write(cert, data2, sizeof(data2)), sizeof(data2));
-    ExpectNull(info_stack = PEM_X509_INFO_read_bio(cert, NULL, NULL, NULL));
-    sk_X509_INFO_pop_free(info_stack, X509_INFO_free);
-    BIO_free(cert);
-#endif
-    return EXPECT_RESULT();
-}
-#endif
 
 static int test_wc_PemToDer(void)
 {
@@ -16350,44 +15628,6 @@ static int test_wolfSSL_certs(void)
 #endif /* OPENSSL_EXTRA && !NO_CERTS */
     return EXPECT_RESULT();
 }
-
-static int test_wolfSSL_X509_check_private_key(void)
-{
-    EXPECT_DECLS;
-#if defined(OPENSSL_EXTRA) && !defined(NO_CERTS) && !defined(NO_RSA) && \
-        defined(USE_CERT_BUFFERS_2048) && !defined(NO_CHECK_PRIVATE_KEY) && \
-        !defined(NO_FILESYSTEM)
-    X509*  x509 = NULL;
-    EVP_PKEY* pkey = NULL;
-    const byte* key;
-
-    /* Check with correct key */
-    ExpectNotNull((x509 = X509_load_certificate_file(cliCertFile,
-        SSL_FILETYPE_PEM)));
-    key = client_key_der_2048;
-    ExpectNotNull(d2i_PrivateKey(EVP_PKEY_RSA, &pkey, &key,
-        (long)sizeof_client_key_der_2048));
-    ExpectIntEQ(X509_check_private_key(x509, pkey), 1);
-    EVP_PKEY_free(pkey);
-    pkey = NULL;
-
-    /* Check with wrong key */
-    key = server_key_der_2048;
-    ExpectNotNull(d2i_PrivateKey(EVP_PKEY_RSA, &pkey, &key,
-        (long)sizeof_server_key_der_2048));
-    ExpectIntEQ(X509_check_private_key(x509, pkey), 0);
-
-    /* test for incorrect parameter */
-    ExpectIntEQ(X509_check_private_key(NULL, pkey), 0);
-    ExpectIntEQ(X509_check_private_key(x509, NULL), 0);
-    ExpectIntEQ(X509_check_private_key(NULL, NULL), 0);
-
-    EVP_PKEY_free(pkey);
-    X509_free(x509);
-#endif
-    return EXPECT_RESULT();
-}
-
 
 static int test_wolfSSL_private_keys(void)
 {
@@ -17853,303 +17093,6 @@ static int test_wolfSSL_ERR_peek_last_error_line(void)
 }
 #endif /* !NO_WOLFSSL_CLIENT && !NO_WOLFSSL_SERVER */
 
-#if defined(OPENSSL_EXTRA) && !defined(NO_CERTS) && \
-       !defined(NO_FILESYSTEM) && !defined(NO_RSA)
-static int verify_cb(int ok, X509_STORE_CTX *ctx)
-{
-    (void) ok;
-    (void) ctx;
-    fprintf(stderr, "ENTER verify_cb\n");
-    return SSL_SUCCESS;
-}
-#endif
-
-static int test_wolfSSL_X509_LOOKUP_ctrl_hash_dir(void)
-{
-    EXPECT_DECLS;
-#if defined(OPENSSL_ALL) && !defined(NO_FILESYSTEM) && !defined(NO_WOLFSSL_DIR)
-    const int  MAX_DIR = 4;
-    const char paths[][32] = {
-                             "./certs/ed25519",
-                             "./certs/ecc",
-                             "./certs/crl",
-                             "./certs/",
-                            };
-
-    char CertCrl_path[MAX_FILENAME_SZ];
-    char *p;
-    X509_STORE* str = NULL;
-    X509_LOOKUP* lookup = NULL;
-    WOLFSSL_STACK* sk = NULL;
-    int len, total_len, i;
-
-    (void)sk;
-
-    XMEMSET(CertCrl_path, 0, MAX_FILENAME_SZ);
-
-    /* illegal string */
-    ExpectNotNull((str = wolfSSL_X509_STORE_new()));
-    ExpectNotNull(lookup = X509_STORE_add_lookup(str, X509_LOOKUP_file()));
-    ExpectIntEQ(X509_LOOKUP_ctrl(lookup, X509_L_ADD_DIR, "",
-                                    SSL_FILETYPE_PEM, NULL), 0);
-    ExpectIntEQ(X509_LOOKUP_ctrl(lookup, X509_L_ADD_STORE, "",
-        SSL_FILETYPE_PEM, NULL), WOLFSSL_NOT_IMPLEMENTED);
-    ExpectIntEQ(X509_LOOKUP_ctrl(lookup, X509_L_LOAD_STORE, "",
-        SSL_FILETYPE_PEM, NULL), WOLFSSL_NOT_IMPLEMENTED);
-    ExpectIntEQ(X509_LOOKUP_ctrl(lookup, 0, "",
-        SSL_FILETYPE_PEM, NULL), WOLFSSL_FAILURE);
-
-    /* free store */
-    X509_STORE_free(str);
-    str = NULL;
-
-    /* short folder string */
-    ExpectNotNull((str = wolfSSL_X509_STORE_new()));
-    ExpectNotNull(lookup = X509_STORE_add_lookup(str, X509_LOOKUP_file()));
-    ExpectIntEQ(X509_LOOKUP_ctrl(lookup, X509_L_ADD_DIR, "./",
-                                    SSL_FILETYPE_PEM,NULL), 1);
-    #if defined(WOLFSSL_INT_H)
-    /* only available when including internal.h */
-    ExpectNotNull(sk = lookup->dirs->dir_entry);
-    #endif
-    /* free store */
-    X509_STORE_free(str);
-    str = NULL;
-
-    /* typical function check */
-    p = &CertCrl_path[0];
-    total_len = 0;
-
-    for (i = MAX_DIR - 1; i>=0 && total_len < MAX_FILENAME_SZ; i--) {
-        len = (int)XSTRLEN((const char*)&paths[i]);
-        total_len += len;
-        XSTRNCPY(p, paths[i], MAX_FILENAME_SZ - total_len);
-        p += len;
-        if (i != 0) *(p++) = SEPARATOR_CHAR;
-    }
-
-    ExpectNotNull((str = wolfSSL_X509_STORE_new()));
-    ExpectNotNull(lookup = X509_STORE_add_lookup(str, X509_LOOKUP_file()));
-    ExpectIntEQ(X509_LOOKUP_ctrl(lookup, X509_L_ADD_DIR, CertCrl_path,
-                                    SSL_FILETYPE_PEM,NULL), 1);
-    #if defined(WOLFSSL_INT_H)
-    /* only available when including internal.h */
-    ExpectNotNull(sk = lookup->dirs->dir_entry);
-    #endif
-
-    X509_STORE_free(str);
-#endif
-    return EXPECT_RESULT();
-}
-
-static int test_wolfSSL_X509_LOOKUP_ctrl_file(void)
-{
-    EXPECT_DECLS;
-#if defined(OPENSSL_ALL) && !defined(NO_CERTS) && \
-    !defined(NO_FILESYSTEM) && !defined(NO_RSA) && \
-    defined(WOLFSSL_SIGNER_DER_CERT)
-    X509_STORE_CTX* ctx = NULL;
-    X509_STORE* str = NULL;
-    X509_LOOKUP* lookup = NULL;
-
-    X509* cert1 = NULL;
-    X509* x509Ca = NULL;
-    X509* x509Svr = NULL;
-    X509* issuer = NULL;
-
-    WOLFSSL_STACK* sk = NULL;
-    X509_NAME* caName = NULL;
-    X509_NAME* issuerName = NULL;
-
-    XFILE file1 = XBADFILE;
-    int i;
-    int cert_count = 0;
-    int cmp;
-
-    char der[] = "certs/ca-cert.der";
-
-#ifdef HAVE_CRL
-    char pem[][100] = {
-        "./certs/crl/crl.pem",
-        "./certs/crl/crl2.pem",
-        "./certs/crl/caEccCrl.pem",
-        "./certs/crl/eccCliCRL.pem",
-        "./certs/crl/eccSrvCRL.pem",
-        ""
-    };
-#endif
-    ExpectTrue((file1 = XFOPEN("./certs/ca-cert.pem", "rb")) != XBADFILE);
-    ExpectNotNull(cert1 = wolfSSL_PEM_read_X509(file1, NULL, NULL, NULL));
-    if (file1 != XBADFILE)
-        XFCLOSE(file1);
-
-    ExpectNotNull(ctx = X509_STORE_CTX_new());
-    ExpectNotNull((str = wolfSSL_X509_STORE_new()));
-    ExpectNotNull(lookup = X509_STORE_add_lookup(str, X509_LOOKUP_file()));
-    ExpectIntEQ(wolfSSL_X509_load_cert_crl_file(NULL, NULL,
-        WOLFSSL_FILETYPE_PEM), 0);
-    ExpectIntEQ(wolfSSL_X509_load_cert_crl_file(lookup, NULL,
-        WOLFSSL_FILETYPE_PEM), 0);
-    ExpectIntEQ(wolfSSL_X509_load_cert_crl_file(NULL, caCertFile,
-        WOLFSSL_FILETYPE_PEM), 0);
-    ExpectIntEQ(wolfSSL_X509_load_cert_crl_file(NULL, der       ,
-        WOLFSSL_FILETYPE_PEM), 0);
-    ExpectIntEQ(X509_LOOKUP_ctrl(lookup, X509_L_FILE_LOAD, caCertFile,
-                                    SSL_FILETYPE_PEM,NULL), 1);
-    ExpectNotNull(sk = wolfSSL_CertManagerGetCerts(str->cm));
-    ExpectIntEQ((cert_count = sk_X509_num(sk)), 1);
-
-    /* check if CA cert is loaded into the store */
-    for (i = 0; i < cert_count; i++) {
-        x509Ca = sk_X509_value(sk, i);
-        ExpectIntEQ(0, wolfSSL_X509_cmp(x509Ca, cert1));
-    }
-
-    ExpectNotNull((x509Svr =
-            wolfSSL_X509_load_certificate_file(svrCertFile, SSL_FILETYPE_PEM)));
-
-    ExpectIntEQ(X509_STORE_CTX_init(ctx, str, x509Svr, NULL), SSL_SUCCESS);
-
-    ExpectNull(X509_STORE_CTX_get0_current_issuer(NULL));
-    issuer = X509_STORE_CTX_get0_current_issuer(ctx);
-    ExpectNull(issuer);
-
-    ExpectIntEQ(X509_verify_cert(ctx), 1);
-
-    issuer = X509_STORE_CTX_get0_current_issuer(ctx);
-    ExpectNotNull(issuer);
-    caName = X509_get_subject_name(x509Ca);
-    ExpectNotNull(caName);
-    issuerName = X509_get_subject_name(issuer);
-    ExpectNotNull(issuerName);
-    cmp = X509_NAME_cmp(caName, issuerName);
-    ExpectIntEQ(cmp, 0);
-
-    /* load der format */
-    issuer = NULL;
-    X509_STORE_CTX_free(ctx);
-    ctx = NULL;
-    X509_STORE_free(str);
-    str = NULL;
-    sk_X509_pop_free(sk, NULL);
-    sk = NULL;
-    X509_free(x509Svr);
-    x509Svr = NULL;
-
-    ExpectNotNull((str = wolfSSL_X509_STORE_new()));
-    ExpectNotNull(lookup = X509_STORE_add_lookup(str, X509_LOOKUP_file()));
-    ExpectIntEQ(X509_LOOKUP_ctrl(lookup, X509_L_FILE_LOAD, der,
-                                    SSL_FILETYPE_ASN1,NULL), 1);
-    ExpectNotNull(sk = wolfSSL_CertManagerGetCerts(str->cm));
-    ExpectIntEQ((cert_count = sk_X509_num(sk)), 1);
-    /* check if CA cert is loaded into the store */
-    for (i = 0; i < cert_count; i++) {
-        x509Ca = sk_X509_value(sk, i);
-        ExpectIntEQ(0, wolfSSL_X509_cmp(x509Ca, cert1));
-    }
-
-    X509_STORE_free(str);
-    str = NULL;
-    sk_X509_pop_free(sk, NULL);
-    sk = NULL;
-    X509_free(cert1);
-    cert1 = NULL;
-
-#ifdef HAVE_CRL
-    ExpectNotNull(str = wolfSSL_X509_STORE_new());
-    ExpectNotNull(lookup = X509_STORE_add_lookup(str, X509_LOOKUP_file()));
-    ExpectIntEQ(X509_LOOKUP_ctrl(lookup, X509_L_FILE_LOAD, caCertFile,
-                                                    SSL_FILETYPE_PEM,NULL), 1);
-    ExpectIntEQ(X509_LOOKUP_ctrl(lookup, X509_L_FILE_LOAD,
-                                "certs/server-revoked-cert.pem",
-                                 SSL_FILETYPE_PEM,NULL), 1);
-    if (str) {
-        ExpectIntEQ(wolfSSL_CertManagerVerify(str->cm, svrCertFile,
-                    WOLFSSL_FILETYPE_PEM), 1);
-        /* since store hasn't yet known the revoked cert*/
-        ExpectIntEQ(wolfSSL_CertManagerVerify(str->cm,
-                    "certs/server-revoked-cert.pem",
-                    WOLFSSL_FILETYPE_PEM), 1);
-    }
-    for (i = 0; pem[i][0] != '\0'; i++)
-    {
-        ExpectIntEQ(X509_LOOKUP_ctrl(lookup, X509_L_FILE_LOAD, pem[i],
-                                        SSL_FILETYPE_PEM, NULL), 1);
-    }
-
-    if (str) {
-        /* since store knows crl list */
-        ExpectIntEQ(wolfSSL_CertManagerVerify(str->cm,
-                    "certs/server-revoked-cert.pem",
-                    WOLFSSL_FILETYPE_PEM ), WC_NO_ERR_TRACE(CRL_CERT_REVOKED));
-    }
-
-    ExpectIntEQ(X509_LOOKUP_ctrl(NULL, 0, NULL, 0, NULL), 0);
-    X509_STORE_free(str);
-#endif
-#endif
-    return EXPECT_RESULT();
-}
-
-static int test_wolfSSL_X509_STORE_CTX_trusted_stack_cleanup(void)
-{
-    int res = TEST_SKIPPED;
-#if defined(OPENSSL_EXTRA)
-    X509_STORE_CTX_cleanup(NULL);
-    X509_STORE_CTX_trusted_stack(NULL, NULL);
-
-    res = TEST_SUCCESS;
-#endif
-    return res;
-}
-
-static int test_wolfSSL_X509_STORE_CTX_get_issuer(void)
-{
-    EXPECT_DECLS;
-#if defined(OPENSSL_EXTRA) && !defined(NO_RSA) && !defined(NO_FILESYSTEM)
-    X509_STORE_CTX* ctx = NULL;
-    X509_STORE* str = NULL;
-    X509* x509Ca = NULL;
-    X509* x509Svr = NULL;
-    X509* issuer = NULL;
-    X509_NAME* caName = NULL;
-    X509_NAME* issuerName = NULL;
-
-    ExpectNotNull(ctx = X509_STORE_CTX_new());
-    ExpectNotNull((str = wolfSSL_X509_STORE_new()));
-    ExpectNotNull((x509Ca =
-            wolfSSL_X509_load_certificate_file(caCertFile, SSL_FILETYPE_PEM)));
-    ExpectIntEQ(X509_STORE_add_cert(str, x509Ca), SSL_SUCCESS);
-    ExpectNotNull((x509Svr =
-            wolfSSL_X509_load_certificate_file(svrCertFile, SSL_FILETYPE_PEM)));
-
-    ExpectIntEQ(X509_STORE_CTX_init(ctx, str, x509Svr, NULL), SSL_SUCCESS);
-
-    /* Issuer0 is not set until chain is built for verification */
-    ExpectNull(X509_STORE_CTX_get0_current_issuer(NULL));
-    ExpectNull(issuer = X509_STORE_CTX_get0_current_issuer(ctx));
-
-    /* Issuer1 will use the store to make a new issuer */
-    ExpectIntEQ(X509_STORE_CTX_get1_issuer(&issuer, ctx, x509Svr), 1);
-    ExpectNotNull(issuer);
-    X509_free(issuer);
-
-    ExpectIntEQ(X509_verify_cert(ctx), 1);
-    ExpectNotNull(issuer = X509_STORE_CTX_get0_current_issuer(ctx));
-    ExpectNotNull(caName = X509_get_subject_name(x509Ca));
-    ExpectNotNull(issuerName = X509_get_subject_name(issuer));
-#ifdef WOLFSSL_SIGNER_DER_CERT
-    ExpectIntEQ(X509_NAME_cmp(caName, issuerName), 0);
-#endif
-
-    X509_STORE_CTX_free(ctx);
-    X509_free(x509Svr);
-    X509_STORE_free(str);
-    X509_free(x509Ca);
-#endif
-    return EXPECT_RESULT();
-}
-
 static int test_wolfSSL_PKCS7_certs(void)
 {
     EXPECT_DECLS;
@@ -18221,871 +17164,6 @@ static int test_wolfSSL_PKCS7_certs(void)
     }
 #endif /* defined(OPENSSL_ALL) && !defined(NO_CERTS) && \
          !defined(NO_FILESYSTEM) && !defined(NO_RSA) && defined(HAVE_PKCS7) */
-    return EXPECT_RESULT();
-}
-
-static int test_wolfSSL_X509_STORE_CTX(void)
-{
-    EXPECT_DECLS;
-#if defined(OPENSSL_EXTRA) && !defined(NO_CERTS) && \
-   !defined(NO_FILESYSTEM) && !defined(NO_RSA)
-    X509_STORE_CTX* ctx = NULL;
-    X509_STORE* str = NULL;
-    X509* x509 = NULL;
-#ifdef OPENSSL_ALL
-    X509* x5092 = NULL;
-    STACK_OF(X509) *sk = NULL;
-    STACK_OF(X509) *sk2 = NULL;
-    STACK_OF(X509) *sk3 = NULL;
-#endif
-
-    ExpectNotNull(ctx = X509_STORE_CTX_new());
-    ExpectNotNull((str = wolfSSL_X509_STORE_new()));
-    ExpectNotNull((x509 =
-                wolfSSL_X509_load_certificate_file(svrCertFile, SSL_FILETYPE_PEM)));
-    ExpectIntEQ(X509_STORE_add_cert(str, x509), SSL_SUCCESS);
-#ifdef OPENSSL_ALL
-    /* sk_X509_new only in OPENSSL_ALL */
-    sk = sk_X509_new_null();
-    ExpectNotNull(sk);
-    ExpectIntEQ(X509_STORE_CTX_init(ctx, str, x509, sk), SSL_SUCCESS);
-#else
-    ExpectIntEQ(X509_STORE_CTX_init(ctx, str, x509, NULL), SSL_SUCCESS);
-#endif
-    ExpectIntEQ(SSL_get_ex_data_X509_STORE_CTX_idx(), 0);
-    X509_STORE_CTX_set_error(ctx, -5);
-    X509_STORE_CTX_set_error(NULL, -5);
-
-    X509_STORE_CTX_free(ctx);
-    ctx = NULL;
-#ifdef OPENSSL_ALL
-    sk_X509_pop_free(sk, NULL);
-    sk = NULL;
-#endif
-    X509_STORE_free(str);
-    str = NULL;
-    X509_free(x509);
-    x509 = NULL;
-
-    ExpectNotNull(ctx = X509_STORE_CTX_new());
-    X509_STORE_CTX_set_verify_cb(ctx, verify_cb);
-    X509_STORE_CTX_free(ctx);
-    ctx = NULL;
-
-#ifdef OPENSSL_ALL
-    /* test X509_STORE_CTX_get(1)_chain */
-    ExpectNotNull((x509 = X509_load_certificate_file(svrCertFile,
-                                                     SSL_FILETYPE_PEM)));
-    ExpectNotNull((x5092 = X509_load_certificate_file(cliCertFile,
-                                                     SSL_FILETYPE_PEM)));
-    ExpectNotNull((sk = sk_X509_new_null()));
-    ExpectIntEQ(sk_X509_push(sk, x509), 1);
-    if (EXPECT_FAIL()) {
-        X509_free(x509);
-        x509 = NULL;
-    }
-    ExpectNotNull((str = X509_STORE_new()));
-    ExpectNotNull((ctx = X509_STORE_CTX_new()));
-    ExpectIntEQ(X509_STORE_CTX_init(ctx, str, x5092, sk), 1);
-    ExpectNull((sk2 = X509_STORE_CTX_get_chain(NULL)));
-    ExpectNull((sk2 = X509_STORE_CTX_get_chain(ctx)));
-    ExpectNull((sk3 = X509_STORE_CTX_get1_chain(NULL)));
-    ExpectNull((sk3 = X509_STORE_CTX_get1_chain(ctx)));
-    X509_STORE_CTX_free(ctx);
-    ctx = NULL;
-    X509_STORE_free(str);
-    str = NULL;
-    /* CTX certs not freed yet */
-    X509_free(x5092);
-    x5092 = NULL;
-    sk_X509_pop_free(sk, NULL);
-    sk = NULL;
-    /* sk3 is dup so free here */
-    sk_X509_pop_free(sk3, NULL);
-    sk3 = NULL;
-#endif
-
-    /* test X509_STORE_CTX_get/set_ex_data */
-    {
-        int i = 0, tmpData = 5;
-        void* tmpDataRet;
-        ExpectNotNull(ctx = X509_STORE_CTX_new());
-    #ifdef HAVE_EX_DATA
-        for (i = 0; i < MAX_EX_DATA; i++) {
-            ExpectIntEQ(X509_STORE_CTX_set_ex_data(ctx, i, &tmpData),
-                        WOLFSSL_SUCCESS);
-            tmpDataRet = (int*)X509_STORE_CTX_get_ex_data(ctx, i);
-            ExpectNotNull(tmpDataRet);
-            ExpectIntEQ(tmpData, *(int*)tmpDataRet);
-        }
-    #else
-        ExpectIntEQ(X509_STORE_CTX_set_ex_data(ctx, i, &tmpData),
-                    WC_NO_ERR_TRACE(WOLFSSL_FAILURE));
-        tmpDataRet = (int*)X509_STORE_CTX_get_ex_data(ctx, i);
-        ExpectNull(tmpDataRet);
-    #endif
-        X509_STORE_CTX_free(ctx);
-        ctx = NULL;
-    }
-
-    /* test X509_STORE_get/set_ex_data */
-    {
-        int i = 0, tmpData = 99;
-        void* tmpDataRet;
-        ExpectNotNull(str = X509_STORE_new());
-    #ifdef HAVE_EX_DATA
-        for (i = 0; i < MAX_EX_DATA; i++) {
-            ExpectIntEQ(X509_STORE_set_ex_data(str, i, &tmpData),
-                        WOLFSSL_SUCCESS);
-            tmpDataRet = (int*)X509_STORE_get_ex_data(str, i);
-            ExpectNotNull(tmpDataRet);
-            ExpectIntEQ(tmpData, *(int*)tmpDataRet);
-        }
-    #else
-        ExpectIntEQ(X509_STORE_set_ex_data(str, i, &tmpData),
-                    WC_NO_ERR_TRACE(WOLFSSL_FAILURE));
-        tmpDataRet = (int*)X509_STORE_get_ex_data(str, i);
-        ExpectNull(tmpDataRet);
-    #endif
-        X509_STORE_free(str);
-        str = NULL;
-    }
-
-#endif /* defined(OPENSSL_EXTRA) && !defined(NO_CERTS) && \
-        * !defined(NO_FILESYSTEM) && !defined(NO_RSA) */
-
-    return EXPECT_RESULT();
-}
-
-#if defined(OPENSSL_EXTRA) && !defined(NO_CERTS) && \
-   !defined(NO_FILESYSTEM) && !defined(NO_RSA)
-
-typedef struct {
-    const char *caFile;
-    const char *caIntFile;
-    const char *caInt2File;
-    const char *leafFile;
-    X509 *x509Ca;
-    X509 *x509CaInt;
-    X509 *x509CaInt2;
-    X509 *x509Leaf;
-    STACK_OF(X509)* expectedChain;
-} X509_STORE_test_data;
-
-static X509 * test_wolfSSL_X509_STORE_CTX_ex_helper(const char *file)
-{
-    XFILE fp = XBADFILE;
-    X509 *x = NULL;
-
-    fp = XFOPEN(file, "rb");
-    if (fp == NULL) {
-        return NULL;
-    }
-    x = PEM_read_X509(fp, 0, 0, 0);
-    XFCLOSE(fp);
-
-    return x;
-}
-
-static int test_wolfSSL_X509_STORE_CTX_ex1(X509_STORE_test_data *testData)
-{
-    EXPECT_DECLS;
-    X509_STORE* store = NULL;
-    X509_STORE_CTX* ctx = NULL;
-    STACK_OF(X509)* chain = NULL;
-    int i = 0;
-
-    /* Test case 1, add X509 certs to store and verify */
-    ExpectNotNull(store = X509_STORE_new());
-    ExpectIntEQ(X509_STORE_add_cert(store, testData->x509Ca), 1);
-    ExpectIntEQ(X509_STORE_add_cert(store, testData->x509CaInt), 1);
-    ExpectIntEQ(X509_STORE_add_cert(store, testData->x509CaInt2), 1);
-    ExpectNotNull(ctx = X509_STORE_CTX_new());
-    ExpectIntEQ(X509_STORE_CTX_init(ctx, store, testData->x509Leaf, NULL), 1);
-    ExpectIntEQ(X509_verify_cert(ctx), 1);
-    ExpectNotNull(chain = X509_STORE_CTX_get_chain(ctx));
-    ExpectIntEQ(sk_X509_num(chain), sk_X509_num(testData->expectedChain));
-    for (i = 0; i < sk_X509_num(chain); i++) {
-        ExpectIntEQ(X509_cmp(sk_X509_value(chain, i),
-                             sk_X509_value(testData->expectedChain, i)), 0);
-    }
-    X509_STORE_CTX_free(ctx);
-    X509_STORE_free(store);
-    return EXPECT_RESULT();
-}
-
-static int test_wolfSSL_X509_STORE_CTX_ex2(X509_STORE_test_data *testData)
-{
-    EXPECT_DECLS;
-    X509_STORE* store = NULL;
-    X509_STORE_CTX* ctx = NULL;
-    STACK_OF(X509)* chain = NULL;
-    int i = 0;
-
-    /* Test case 2, add certs by filename to store and verify */
-    ExpectNotNull(store = X509_STORE_new());
-    ExpectIntEQ(X509_STORE_load_locations(
-        store, testData->caFile, NULL), 1);
-    ExpectIntEQ(X509_STORE_load_locations(
-        store, testData->caIntFile, NULL), 1);
-    ExpectIntEQ(X509_STORE_load_locations(
-        store, testData->caInt2File, NULL), 1);
-    ExpectNotNull(ctx = X509_STORE_CTX_new());
-    ExpectIntEQ(X509_STORE_CTX_init(ctx, store, testData->x509Leaf, NULL), 1);
-    ExpectIntEQ(X509_verify_cert(ctx), 1);
-    ExpectNotNull(chain = X509_STORE_CTX_get_chain(ctx));
-    ExpectIntEQ(sk_X509_num(chain), sk_X509_num(testData->expectedChain));
-    for (i = 0; i < sk_X509_num(chain); i++) {
-        ExpectIntEQ(X509_cmp(sk_X509_value(chain, i),
-                             sk_X509_value(testData->expectedChain, i)), 0);
-    }
-    X509_STORE_CTX_free(ctx);
-    X509_STORE_free(store);
-    return EXPECT_RESULT();
-}
-
-static int test_wolfSSL_X509_STORE_CTX_ex3(X509_STORE_test_data *testData)
-{
-    EXPECT_DECLS;
-    X509_STORE* store = NULL;
-    X509_STORE_CTX* ctx = NULL;
-    STACK_OF(X509)* chain = NULL;
-    int i = 0;
-
-    /* Test case 3, mix and match X509 with files */
-    ExpectNotNull(store = X509_STORE_new());
-    ExpectIntEQ(X509_STORE_add_cert(store, testData->x509CaInt), 1);
-    ExpectIntEQ(X509_STORE_add_cert(store, testData->x509CaInt2), 1);
-    ExpectIntEQ(X509_STORE_load_locations(
-        store, testData->caFile, NULL), 1);
-    ExpectNotNull(ctx = X509_STORE_CTX_new());
-    ExpectIntEQ(X509_STORE_CTX_init(ctx, store, testData->x509Leaf, NULL), 1);
-    ExpectIntEQ(X509_verify_cert(ctx), 1);
-    ExpectNotNull(chain = X509_STORE_CTX_get_chain(ctx));
-    ExpectIntEQ(sk_X509_num(chain), sk_X509_num(testData->expectedChain));
-    for (i = 0; i < sk_X509_num(chain); i++) {
-        ExpectIntEQ(X509_cmp(sk_X509_value(chain, i),
-                             sk_X509_value(testData->expectedChain, i)), 0);
-    }
-    X509_STORE_CTX_free(ctx);
-    X509_STORE_free(store);
-    return EXPECT_RESULT();
-}
-
-static int test_wolfSSL_X509_STORE_CTX_ex4(X509_STORE_test_data *testData)
-{
-    EXPECT_DECLS;
-    X509_STORE* store = NULL;
-    X509_STORE_CTX* ctx = NULL;
-    STACK_OF(X509)* chain = NULL;
-    STACK_OF(X509)* inter = NULL;
-    int i = 0;
-
-    /* Test case 4, CA loaded by file, intermediates passed on init */
-    ExpectNotNull(store = X509_STORE_new());
-    ExpectIntEQ(X509_STORE_load_locations(
-        store, testData->caFile, NULL), 1);
-    ExpectNotNull(inter = sk_X509_new_null());
-    ExpectIntGE(sk_X509_push(inter, testData->x509CaInt), 1);
-    ExpectIntGE(sk_X509_push(inter, testData->x509CaInt2), 1);
-    ExpectNotNull(ctx = X509_STORE_CTX_new());
-    ExpectIntEQ(X509_STORE_CTX_init(ctx, store, testData->x509Leaf, inter), 1);
-    ExpectIntEQ(X509_verify_cert(ctx), 1);
-    ExpectNotNull(chain = X509_STORE_CTX_get_chain(ctx));
-    ExpectIntEQ(sk_X509_num(chain), sk_X509_num(testData->expectedChain));
-    for (i = 0; i < sk_X509_num(chain); i++) {
-        ExpectIntEQ(X509_cmp(sk_X509_value(chain, i),
-                             sk_X509_value(testData->expectedChain, i)), 0);
-    }
-    X509_STORE_CTX_free(ctx);
-    X509_STORE_free(store);
-    sk_X509_free(inter);
-    return EXPECT_RESULT();
-}
-
-static int test_wolfSSL_X509_STORE_CTX_ex5(X509_STORE_test_data *testData)
-{
-    EXPECT_DECLS;
-    X509_STORE* store = NULL;
-    X509_STORE_CTX* ctx = NULL;
-    STACK_OF(X509)* chain = NULL;
-    STACK_OF(X509)* trusted = NULL;
-    int i = 0;
-
-    /* Test case 5, manually set trusted stack */
-    ExpectNotNull(store = X509_STORE_new());
-    ExpectNotNull(trusted = sk_X509_new_null());
-    ExpectIntGE(sk_X509_push(trusted, testData->x509Ca), 1);
-    ExpectIntGE(sk_X509_push(trusted, testData->x509CaInt), 1);
-    ExpectIntGE(sk_X509_push(trusted, testData->x509CaInt2), 1);
-    ExpectNotNull(ctx = X509_STORE_CTX_new());
-    ExpectIntEQ(X509_STORE_CTX_init(ctx, store, testData->x509Leaf, NULL), 1);
-    X509_STORE_CTX_trusted_stack(ctx, trusted);
-    ExpectIntEQ(X509_verify_cert(ctx), 1);
-    ExpectNotNull(chain = X509_STORE_CTX_get_chain(ctx));
-    ExpectIntEQ(sk_X509_num(chain), sk_X509_num(testData->expectedChain));
-    for (i = 0; i < sk_X509_num(chain); i++) {
-        ExpectIntEQ(X509_cmp(sk_X509_value(chain, i),
-                             sk_X509_value(testData->expectedChain, i)), 0);
-    }
-    X509_STORE_CTX_free(ctx);
-    X509_STORE_free(store);
-    sk_X509_free(trusted);
-    return EXPECT_RESULT();
-}
-
-static int test_wolfSSL_X509_STORE_CTX_ex6(X509_STORE_test_data *testData)
-{
-    EXPECT_DECLS;
-    X509_STORE* store = NULL;
-    X509_STORE_CTX* ctx = NULL;
-    STACK_OF(X509)* chain = NULL;
-    STACK_OF(X509)* trusted = NULL;
-    STACK_OF(X509)* inter = NULL;
-    int i = 0;
-
-    /* Test case 6, manually set trusted stack will be unified with
-     * any intermediates provided on init */
-    ExpectNotNull(store = X509_STORE_new());
-    ExpectNotNull(trusted = sk_X509_new_null());
-    ExpectNotNull(inter = sk_X509_new_null());
-    ExpectIntGE(sk_X509_push(trusted, testData->x509Ca), 1);
-    ExpectIntGE(sk_X509_push(inter, testData->x509CaInt), 1);
-    ExpectIntGE(sk_X509_push(inter, testData->x509CaInt2), 1);
-    ExpectNotNull(ctx = X509_STORE_CTX_new());
-    ExpectIntEQ(X509_STORE_CTX_init(ctx, store, testData->x509Leaf, inter), 1);
-    X509_STORE_CTX_trusted_stack(ctx, trusted);
-    ExpectIntEQ(X509_verify_cert(ctx), 1);
-    ExpectNotNull(chain = X509_STORE_CTX_get_chain(ctx));
-    ExpectIntEQ(sk_X509_num(chain), sk_X509_num(testData->expectedChain));
-    for (i = 0; i < sk_X509_num(chain); i++) {
-        ExpectIntEQ(X509_cmp(sk_X509_value(chain, i),
-                             sk_X509_value(testData->expectedChain, i)), 0);
-    }
-    X509_STORE_CTX_free(ctx);
-    X509_STORE_free(store);
-    sk_X509_free(trusted);
-    sk_X509_free(inter);
-    return EXPECT_RESULT();
-}
-
-static int test_wolfSSL_X509_STORE_CTX_ex7(X509_STORE_test_data *testData)
-{
-    EXPECT_DECLS;
-    X509_STORE* store = NULL;
-    X509_STORE_CTX* ctx = NULL;
-    STACK_OF(X509)* chain = NULL;
-    int i = 0;
-
-    /* Test case 7, certs added to store after ctx init are still used */
-    ExpectNotNull(store = X509_STORE_new());
-    ExpectNotNull(ctx = X509_STORE_CTX_new());
-    ExpectIntEQ(X509_STORE_CTX_init(ctx, store, testData->x509Leaf, NULL), 1);
-    ExpectIntNE(X509_verify_cert(ctx), 1);
-    ExpectIntEQ(X509_STORE_add_cert(store, testData->x509CaInt2), 1);
-    ExpectIntEQ(X509_STORE_add_cert(store, testData->x509CaInt), 1);
-    ExpectIntEQ(X509_STORE_add_cert(store, testData->x509Ca), 1);
-    ExpectIntEQ(X509_verify_cert(ctx), 1);
-    ExpectNotNull(chain = X509_STORE_CTX_get_chain(ctx));
-    ExpectIntEQ(sk_X509_num(chain), sk_X509_num(testData->expectedChain));
-    for (i = 0; i < sk_X509_num(chain); i++) {
-        ExpectIntEQ(X509_cmp(sk_X509_value(chain, i),
-                             sk_X509_value(testData->expectedChain, i)), 0);
-    }
-    X509_STORE_CTX_free(ctx);
-    X509_STORE_free(store);
-    return EXPECT_RESULT();
-}
-
-static int test_wolfSSL_X509_STORE_CTX_ex8(X509_STORE_test_data *testData)
-{
-    EXPECT_DECLS;
-    X509_STORE* store = NULL;
-    X509_STORE_CTX* ctx = NULL;
-    STACK_OF(X509)* chain = NULL;
-    int i = 0;
-
-    /* Test case 8, Only full chain verifies */
-    ExpectNotNull(store = X509_STORE_new());
-    ExpectNotNull(ctx = X509_STORE_CTX_new());
-    ExpectIntEQ(X509_STORE_CTX_init(ctx, store, testData->x509Leaf, NULL), 1);
-    ExpectIntNE(X509_verify_cert(ctx), 1);
-    ExpectIntEQ(X509_STORE_add_cert(store, testData->x509CaInt2), 1);
-    ExpectIntNE(X509_verify_cert(ctx), 1);
-    ExpectIntEQ(X509_STORE_add_cert(store, testData->x509CaInt), 1);
-    ExpectIntNE(X509_verify_cert(ctx), 1);
-    ExpectIntEQ(X509_STORE_add_cert(store, testData->x509Ca), 1);
-    ExpectIntEQ(X509_verify_cert(ctx), 1);
-    ExpectNotNull(chain = X509_STORE_CTX_get_chain(ctx));
-    ExpectIntEQ(sk_X509_num(chain), sk_X509_num(testData->expectedChain));
-    for (i = 0; i < sk_X509_num(chain); i++) {
-        ExpectIntEQ(X509_cmp(sk_X509_value(chain, i),
-                             sk_X509_value(testData->expectedChain, i)), 0);
-    }
-    X509_STORE_CTX_free(ctx);
-    X509_STORE_free(store);
-    return EXPECT_RESULT();
-}
-
-static int test_wolfSSL_X509_STORE_CTX_ex9(X509_STORE_test_data *testData)
-{
-    EXPECT_DECLS;
-    X509_STORE* store = NULL;
-    X509_STORE_CTX* ctx = NULL;
-    X509_STORE_CTX* ctx2 = NULL;
-    STACK_OF(X509)* trusted = NULL;
-
-    /* Test case 9, certs added to store should not be reflected in ctx that
-     * has been manually set with a trusted stack, but are reflected in ctx
-     * that has not set trusted stack */
-    ExpectNotNull(store = X509_STORE_new());
-    ExpectNotNull(ctx = X509_STORE_CTX_new());
-    ExpectNotNull(ctx2 = X509_STORE_CTX_new());
-    ExpectNotNull(trusted = sk_X509_new_null());
-    ExpectIntGE(sk_X509_push(trusted, testData->x509Ca), 1);
-    ExpectIntGE(sk_X509_push(trusted, testData->x509CaInt), 1);
-    ExpectIntGE(sk_X509_push(trusted, testData->x509CaInt2), 1);
-    ExpectIntEQ(X509_STORE_CTX_init(ctx, store, testData->x509Leaf, NULL), 1);
-    ExpectIntEQ(X509_STORE_CTX_init(ctx2, store, testData->x509Leaf, NULL), 1);
-    ExpectIntNE(X509_verify_cert(ctx), 1);
-    ExpectIntNE(X509_verify_cert(ctx2), 1);
-    X509_STORE_CTX_trusted_stack(ctx, trusted);
-    /* CTX1 should now verify */
-    ExpectIntEQ(X509_verify_cert(ctx), 1);
-    ExpectIntNE(X509_verify_cert(ctx2), 1);
-    ExpectIntEQ(X509_STORE_add_cert(store, testData->x509Ca), 1);
-    ExpectIntEQ(X509_STORE_add_cert(store, testData->x509CaInt), 1);
-    ExpectIntEQ(X509_STORE_add_cert(store, testData->x509CaInt2), 1);
-    /* CTX2 should now verify */
-    ExpectIntEQ(X509_verify_cert(ctx2), 1);
-    X509_STORE_CTX_free(ctx);
-    X509_STORE_CTX_free(ctx2);
-    X509_STORE_free(store);
-    sk_X509_free(trusted);
-    return EXPECT_RESULT();
-}
-
-static int test_wolfSSL_X509_STORE_CTX_ex10(X509_STORE_test_data *testData)
-{
-    EXPECT_DECLS;
-    X509_STORE* store = NULL;
-    X509_STORE_CTX* ctx = NULL;
-    STACK_OF(X509)* chain = NULL;
-
-    /* Test case 10, ensure partial chain flag works */
-    ExpectNotNull(store = X509_STORE_new());
-    ExpectIntEQ(X509_STORE_add_cert(store, testData->x509CaInt), 1);
-    ExpectIntEQ(X509_STORE_add_cert(store, testData->x509CaInt2), 1);
-    ExpectNotNull(ctx = X509_STORE_CTX_new());
-    ExpectIntEQ(X509_STORE_CTX_init(ctx, store, testData->x509Leaf, NULL), 1);
-    /* Fails because chain is incomplete */
-    ExpectIntNE(X509_verify_cert(ctx), 1);
-    ExpectIntEQ(X509_STORE_set_flags(store, X509_V_FLAG_PARTIAL_CHAIN), 1);
-    /* Partial chain now OK */
-    ExpectIntEQ(X509_verify_cert(ctx), 1);
-    ExpectNotNull(chain = X509_STORE_CTX_get_chain(ctx));
-    X509_STORE_CTX_free(ctx);
-    X509_STORE_free(store);
-    return EXPECT_RESULT();
-}
-
-static int test_wolfSSL_X509_STORE_CTX_ex11(X509_STORE_test_data *testData)
-{
-    EXPECT_DECLS;
-    X509_STORE* store = NULL;
-    X509_STORE_CTX* ctx = NULL;
-    STACK_OF(X509)* chain = NULL;
-
-    /* Test case 11, test partial chain flag on ctx itself */
-    ExpectNotNull(store = X509_STORE_new());
-    ExpectIntEQ(X509_STORE_add_cert(store, testData->x509CaInt), 1);
-    ExpectIntEQ(X509_STORE_add_cert(store, testData->x509CaInt2), 1);
-    ExpectNotNull(ctx = X509_STORE_CTX_new());
-    ExpectIntEQ(X509_STORE_CTX_init(ctx, store, testData->x509Leaf, NULL), 1);
-    /* Fails because chain is incomplete */
-    ExpectIntNE(X509_verify_cert(ctx), 1);
-    X509_STORE_CTX_set_flags(ctx, X509_V_FLAG_PARTIAL_CHAIN);
-    /* Partial chain now OK */
-    ExpectIntEQ(X509_verify_cert(ctx), 1);
-    ExpectNotNull(chain = X509_STORE_CTX_get_chain(ctx));
-    X509_STORE_CTX_free(ctx);
-    X509_STORE_free(store);
-    return EXPECT_RESULT();
-}
-
-static int test_wolfSSL_X509_STORE_CTX_ex12(void)
-{
-    EXPECT_DECLS;
-#ifdef HAVE_ECC
-    X509_STORE* store = NULL;
-    X509_STORE_CTX* ctx = NULL;
-    STACK_OF(X509)* chain = NULL;
-    X509* rootEccX509 = NULL;
-    X509* badAkiX509 = NULL;
-    X509* ca1X509 = NULL;
-
-    const char* intCARootECCFile    = "./certs/ca-ecc-cert.pem";
-    const char* intCA1ECCFile       = "./certs/intermediate/ca-int-ecc-cert.pem";
-    const char* intCABadAKIECCFile = "./certs/intermediate/ca-ecc-bad-aki.pem";
-
-    /* Test case 12, multiple CAs with the same SKI including 1 with intentionally
-       bad/unregistered AKI.  x509_verify_cert should still form a valid chain
-       using the valid CA, ignoring the bad CA. Developed from customer provided
-       reproducer. */
-
-    ExpectNotNull(store = X509_STORE_new());
-    ExpectNotNull(rootEccX509 = test_wolfSSL_X509_STORE_CTX_ex_helper(intCARootECCFile));
-    ExpectIntEQ(X509_STORE_add_cert(store, rootEccX509), 1);
-    ExpectNotNull(badAkiX509 = test_wolfSSL_X509_STORE_CTX_ex_helper(intCABadAKIECCFile));
-    ExpectNotNull(ctx = X509_STORE_CTX_new());
-    ExpectIntEQ(X509_STORE_CTX_init(ctx, store, badAkiX509, NULL), 1);
-    ExpectIntEQ(X509_verify_cert(ctx), 0);
-    X509_STORE_CTX_cleanup(ctx);
-
-    ExpectIntEQ(X509_STORE_add_cert(store, badAkiX509), 1);
-    ExpectNotNull(ca1X509 = test_wolfSSL_X509_STORE_CTX_ex_helper(intCA1ECCFile));
-    ExpectIntEQ(X509_STORE_CTX_init(ctx, store, ca1X509, NULL), 1);
-    ExpectIntEQ(X509_verify_cert(ctx), 1);
-    ExpectNotNull(chain = X509_STORE_CTX_get_chain(ctx));
-
-    X509_STORE_CTX_free(ctx);
-    X509_STORE_free(store);
-    X509_free(rootEccX509);
-    X509_free(badAkiX509);
-    X509_free(ca1X509);
-#endif
-    return EXPECT_RESULT();
-}
-#endif
-
-static int test_wolfSSL_X509_STORE_CTX_ex(void)
-{
-    EXPECT_DECLS;
-#if defined(OPENSSL_EXTRA) && !defined(NO_CERTS) && \
-   !defined(NO_FILESYSTEM) && !defined(NO_RSA)
-    X509_STORE_test_data testData;
-    XMEMSET((void *)&testData, 0, sizeof(X509_STORE_test_data));
-    testData.caFile =     "./certs/ca-cert.pem";
-    testData.caIntFile =  "./certs/intermediate/ca-int-cert.pem";
-    testData.caInt2File = "./certs/intermediate/ca-int2-cert.pem";
-    testData.leafFile =   "./certs/intermediate/server-chain.pem";
-
-    ExpectNotNull(testData.x509Ca = \
-                  test_wolfSSL_X509_STORE_CTX_ex_helper(testData.caFile));
-    ExpectNotNull(testData.x509CaInt = \
-                  test_wolfSSL_X509_STORE_CTX_ex_helper(testData.caIntFile));
-    ExpectNotNull(testData.x509CaInt2 = \
-                  test_wolfSSL_X509_STORE_CTX_ex_helper(testData.caInt2File));
-    ExpectNotNull(testData.x509Leaf = \
-                  test_wolfSSL_X509_STORE_CTX_ex_helper(testData.leafFile));
-    ExpectNotNull(testData.expectedChain = sk_X509_new_null());
-    ExpectIntGE(sk_X509_push(testData.expectedChain, testData.x509Leaf), 1);
-    ExpectIntGE(sk_X509_push(testData.expectedChain, testData.x509CaInt2), 1);
-    ExpectIntGE(sk_X509_push(testData.expectedChain, testData.x509CaInt), 1);
-    ExpectIntGE(sk_X509_push(testData.expectedChain, testData.x509Ca), 1);
-
-    ExpectIntEQ(test_wolfSSL_X509_STORE_CTX_ex1(&testData), 1);
-    ExpectIntEQ(test_wolfSSL_X509_STORE_CTX_ex2(&testData), 1);
-    ExpectIntEQ(test_wolfSSL_X509_STORE_CTX_ex3(&testData), 1);
-    ExpectIntEQ(test_wolfSSL_X509_STORE_CTX_ex4(&testData), 1);
-    ExpectIntEQ(test_wolfSSL_X509_STORE_CTX_ex5(&testData), 1);
-    ExpectIntEQ(test_wolfSSL_X509_STORE_CTX_ex6(&testData), 1);
-    ExpectIntEQ(test_wolfSSL_X509_STORE_CTX_ex7(&testData), 1);
-    ExpectIntEQ(test_wolfSSL_X509_STORE_CTX_ex8(&testData), 1);
-    ExpectIntEQ(test_wolfSSL_X509_STORE_CTX_ex9(&testData), 1);
-    ExpectIntEQ(test_wolfSSL_X509_STORE_CTX_ex10(&testData), 1);
-    ExpectIntEQ(test_wolfSSL_X509_STORE_CTX_ex11(&testData), 1);
-    test_wolfSSL_X509_STORE_CTX_ex12();
-
-    if(testData.x509Ca) {
-        X509_free(testData.x509Ca);
-    }
-    if(testData.x509CaInt) {
-        X509_free(testData.x509CaInt);
-    }
-    if(testData.x509CaInt2) {
-        X509_free(testData.x509CaInt2);
-    }
-    if(testData.x509Leaf) {
-        X509_free(testData.x509Leaf);
-    }
-    if (testData.expectedChain) {
-        sk_X509_free(testData.expectedChain);
-    }
-
-#endif /* defined(OPENSSL_EXTRA) && !defined(NO_CERTS) && \
-        * !defined(NO_FILESYSTEM) && !defined(NO_RSA) */
-
-    return EXPECT_RESULT();
-}
-
-
-#if defined(OPENSSL_EXTRA) && !defined(NO_RSA) && !defined(NO_FILESYSTEM)
-static int test_X509_STORE_untrusted_load_cert_to_stack(const char* filename,
-        STACK_OF(X509)* chain)
-{
-    EXPECT_DECLS;
-    XFILE fp = XBADFILE;
-    X509* cert = NULL;
-
-    ExpectTrue((fp = XFOPEN(filename, "rb"))
-            != XBADFILE);
-    ExpectNotNull(cert = PEM_read_X509(fp, 0, 0, 0 ));
-    if (fp != XBADFILE) {
-        XFCLOSE(fp);
-        fp = XBADFILE;
-    }
-    ExpectIntGT(sk_X509_push(chain, cert), 0);
-    if (EXPECT_FAIL())
-        X509_free(cert);
-
-    return EXPECT_RESULT();
-}
-
-#if defined(OPENSSL_ALL)
-
-static int last_errcode;
-static int last_errdepth;
-
-static int X509Callback(int ok, X509_STORE_CTX *ctx)
-{
-
-    if (!ok) {
-        last_errcode  = X509_STORE_CTX_get_error(ctx);
-        last_errdepth = X509_STORE_CTX_get_error_depth(ctx);
-    }
-    /* Always return OK to allow verification to continue.*/
-    return 1;
-}
-
-static int test_X509_STORE_InvalidCa(void)
-{
-    EXPECT_DECLS;
-    const char* filename = "./certs/intermediate/ca_false_intermediate/"
-                                                    "test_int_not_cacert.pem";
-    const char* srvfile = "./certs/intermediate/ca_false_intermediate/"
-                                            "test_sign_bynoca_srv.pem";
-    X509_STORE_CTX* ctx = NULL;
-    X509_STORE* str = NULL;
-    XFILE fp = XBADFILE;
-    X509* cert = NULL;
-    STACK_OF(X509)* untrusted = NULL;
-
-    last_errcode = 0;
-    last_errdepth = 0;
-
-    ExpectTrue((fp = XFOPEN(srvfile, "rb"))
-            != XBADFILE);
-    ExpectNotNull(cert = PEM_read_X509(fp, 0, 0, 0 ));
-    if (fp != XBADFILE) {
-        XFCLOSE(fp);
-        fp = XBADFILE;
-    }
-
-    ExpectNotNull(str = X509_STORE_new());
-    ExpectNotNull(ctx = X509_STORE_CTX_new());
-    ExpectNotNull(untrusted = sk_X509_new_null());
-
-    /* create cert chain stack */
-    ExpectIntEQ(test_X509_STORE_untrusted_load_cert_to_stack(filename,
-                untrusted), TEST_SUCCESS);
-
-    X509_STORE_set_verify_cb(str, X509Callback);
-
-    ExpectIntEQ(X509_STORE_load_locations(str,
-                "./certs/intermediate/ca_false_intermediate/test_ca.pem",
-                                                                    NULL), 1);
-
-    ExpectIntEQ(X509_STORE_CTX_init(ctx, str, cert, untrusted), 1);
-    ExpectIntEQ(X509_verify_cert(ctx), 1);
-    ExpectIntEQ(last_errcode, X509_V_ERR_INVALID_CA);
-
-    X509_free(cert);
-    X509_STORE_free(str);
-    X509_STORE_CTX_free(ctx);
-    sk_X509_pop_free(untrusted, NULL);
-
-    return EXPECT_RESULT();
-}
-#endif /* OPENSSL_ALL */
-
-
-
-static int test_X509_STORE_untrusted_certs(const char** filenames, int ret,
-        int err, int loadCA)
-{
-    EXPECT_DECLS;
-    X509_STORE_CTX* ctx = NULL;
-    X509_STORE* str = NULL;
-    XFILE fp = XBADFILE;
-    X509* cert = NULL;
-    STACK_OF(X509)* untrusted = NULL;
-
-    ExpectTrue((fp = XFOPEN("./certs/intermediate/server-int-cert.pem", "rb"))
-            != XBADFILE);
-    ExpectNotNull(cert = PEM_read_X509(fp, 0, 0, 0 ));
-    if (fp != XBADFILE) {
-        XFCLOSE(fp);
-        fp = XBADFILE;
-    }
-
-    ExpectNotNull(str = X509_STORE_new());
-    ExpectNotNull(ctx = X509_STORE_CTX_new());
-    ExpectNotNull(untrusted = sk_X509_new_null());
-
-    ExpectIntEQ(X509_STORE_set_flags(str, 0), 1);
-    if (loadCA) {
-        ExpectIntEQ(X509_STORE_load_locations(str, "./certs/ca-cert.pem", NULL),
-                1);
-    }
-    for (; *filenames; filenames++) {
-        ExpectIntEQ(test_X509_STORE_untrusted_load_cert_to_stack(*filenames,
-                untrusted), TEST_SUCCESS);
-    }
-
-    ExpectIntEQ(X509_STORE_CTX_init(ctx, str, cert, untrusted), 1);
-    ExpectIntEQ(X509_verify_cert(ctx), ret);
-    ExpectIntEQ(X509_STORE_CTX_get_error(ctx), err);
-
-    X509_free(cert);
-    X509_STORE_free(str);
-    X509_STORE_CTX_free(ctx);
-    sk_X509_pop_free(untrusted, NULL);
-
-    return EXPECT_RESULT();
-}
-#endif
-
-static int test_X509_STORE_untrusted(void)
-{
-    EXPECT_DECLS;
-#if defined(OPENSSL_EXTRA) && !defined(NO_RSA) && !defined(NO_FILESYSTEM)
-    const char* untrusted1[] = {
-        "./certs/intermediate/ca-int2-cert.pem",
-        NULL
-    };
-    const char* untrusted2[] = {
-        "./certs/intermediate/ca-int-cert.pem",
-        "./certs/intermediate/ca-int2-cert.pem",
-        NULL
-    };
-    const char* untrusted3[] = {
-        "./certs/intermediate/ca-int-cert.pem",
-        "./certs/intermediate/ca-int2-cert.pem",
-        "./certs/ca-cert.pem",
-        NULL
-    };
-    /* Adding unrelated certs that should be ignored */
-    const char* untrusted4[] = {
-        "./certs/client-ca.pem",
-        "./certs/intermediate/ca-int-cert.pem",
-        "./certs/server-cert.pem",
-        "./certs/intermediate/ca-int2-cert.pem",
-        NULL
-    };
-
-    /* Only immediate issuer in untrusted chain. Fails since can't build chain
-     * to loaded CA. */
-    ExpectIntEQ(test_X509_STORE_untrusted_certs(untrusted1, 0,
-            X509_V_ERR_UNABLE_TO_GET_ISSUER_CERT_LOCALLY, 1), TEST_SUCCESS);
-    /* Succeeds because path to loaded CA is available. */
-    ExpectIntEQ(test_X509_STORE_untrusted_certs(untrusted2, 1, 0, 1),
-            TEST_SUCCESS);
-    /* Root CA in untrusted chain is OK so long as CA has been loaded
-     * properly */
-    ExpectIntEQ(test_X509_STORE_untrusted_certs(untrusted3, 1, 0, 1),
-            TEST_SUCCESS);
-    /* Still needs properly loaded CA, while including it in untrusted
-     * list is not an error, it also doesn't count for verify */
-    ExpectIntEQ(test_X509_STORE_untrusted_certs(untrusted3, 0,
-                X509_V_ERR_UNABLE_TO_GET_ISSUER_CERT_LOCALLY, 0),
-            TEST_SUCCESS);
-    /* Succeeds because path to loaded CA is available. */
-    ExpectIntEQ(test_X509_STORE_untrusted_certs(untrusted4, 1, 0, 1),
-            TEST_SUCCESS);
-#endif
-    return EXPECT_RESULT();
-}
-
-static int test_wolfSSL_X509_STORE_set_flags(void)
-{
-    EXPECT_DECLS;
-#if defined(OPENSSL_EXTRA) && !defined(NO_CERTS) && \
-   !defined(NO_FILESYSTEM) && !defined(NO_RSA)
-    X509_STORE* store = NULL;
-    X509* x509 = NULL;
-
-    ExpectNotNull((store = wolfSSL_X509_STORE_new()));
-    ExpectNotNull((x509 = wolfSSL_X509_load_certificate_file(svrCertFile,
-        WOLFSSL_FILETYPE_PEM)));
-    ExpectIntEQ(X509_STORE_add_cert(store, x509), WOLFSSL_SUCCESS);
-
-#ifdef HAVE_CRL
-    ExpectIntEQ(X509_STORE_set_flags(store, WOLFSSL_CRL_CHECKALL),
-        WOLFSSL_SUCCESS);
-#else
-    ExpectIntEQ(X509_STORE_set_flags(store, WOLFSSL_CRL_CHECKALL),
-        WC_NO_ERR_TRACE(NOT_COMPILED_IN));
-#endif
-
-    wolfSSL_X509_free(x509);
-    wolfSSL_X509_STORE_free(store);
-#endif /* defined(OPENSSL_EXTRA) && !defined(NO_CERTS) &&
-        * !defined(NO_FILESYSTEM) && !defined(NO_RSA) */
-    return EXPECT_RESULT();
-}
-
-static int test_wolfSSL_X509_LOOKUP_load_file(void)
-{
-    EXPECT_DECLS;
-#if defined(OPENSSL_EXTRA) && defined(HAVE_CRL) && \
-   !defined(NO_FILESYSTEM) && !defined(NO_RSA) && defined(HAVE_ECC) && \
-   (!defined(NO_WOLFSSL_CLIENT) || !defined(WOLFSSL_NO_CLIENT_AUTH))
-    WOLFSSL_X509_STORE*  store = NULL;
-    WOLFSSL_X509_LOOKUP* lookup = NULL;
-
-    ExpectNotNull(store = wolfSSL_X509_STORE_new());
-    ExpectNotNull(lookup = X509_STORE_add_lookup(store, X509_LOOKUP_file()));
-    /* One RSA and one ECC certificate in file. */
-    ExpectIntEQ(wolfSSL_X509_LOOKUP_load_file(lookup, "certs/client-ca.pem",
-        X509_FILETYPE_PEM), 1);
-    ExpectIntEQ(wolfSSL_X509_LOOKUP_load_file(lookup, "certs/crl/crl2.pem",
-        X509_FILETYPE_PEM), 1);
-
-    if (store != NULL) {
-        ExpectIntEQ(wolfSSL_CertManagerVerify(store->cm, cliCertFile,
-            WOLFSSL_FILETYPE_PEM), 1);
-        ExpectIntEQ(wolfSSL_CertManagerVerify(store->cm, svrCertFile,
-            WOLFSSL_FILETYPE_PEM), WC_NO_ERR_TRACE(ASN_NO_SIGNER_E));
-    }
-    ExpectIntEQ(wolfSSL_X509_LOOKUP_load_file(lookup, "certs/ca-cert.pem",
-        X509_FILETYPE_PEM), 1);
-    if (store != NULL) {
-        ExpectIntEQ(wolfSSL_CertManagerVerify(store->cm, svrCertFile,
-            WOLFSSL_FILETYPE_PEM), 1);
-    }
-
-    wolfSSL_X509_STORE_free(store);
-#endif /* defined(OPENSSL_EXTRA) && defined(HAVE_CRL) &&
-        * !defined(NO_FILESYSTEM) && !defined(NO_RSA) */
-    return EXPECT_RESULT();
-}
-
-static int test_wolfSSL_X509_STORE_CTX_set_time(void)
-{
-    EXPECT_DECLS;
-#if defined(OPENSSL_EXTRA)
-    WOLFSSL_X509_STORE_CTX* ctx = NULL;
-    time_t c_time;
-
-    ExpectNotNull(ctx = wolfSSL_X509_STORE_CTX_new());
-    c_time = 365*24*60*60;
-    wolfSSL_X509_STORE_CTX_set_time(ctx, 0, c_time);
-    ExpectTrue((ctx->param->flags & WOLFSSL_USE_CHECK_TIME) ==
-        WOLFSSL_USE_CHECK_TIME);
-    ExpectTrue(ctx->param->check_time == c_time);
-    wolfSSL_X509_STORE_CTX_free(ctx);
-#endif /* OPENSSL_EXTRA */
     return EXPECT_RESULT();
 }
 
@@ -19166,35 +17244,6 @@ static int test_wolfSSL_get0_param(void)
     return EXPECT_RESULT();
 }
 
-static int test_wolfSSL_X509_VERIFY_PARAM_set1_host(void)
-{
-    EXPECT_DECLS;
-#if defined(OPENSSL_EXTRA)
-    const char host[] = "www.example.com";
-    WOLFSSL_X509_VERIFY_PARAM* pParam = NULL;
-
-    ExpectNotNull(pParam = (WOLFSSL_X509_VERIFY_PARAM*)XMALLOC(
-        sizeof(WOLFSSL_X509_VERIFY_PARAM), HEAP_HINT, DYNAMIC_TYPE_OPENSSL));
-    if (pParam != NULL) {
-        XMEMSET(pParam, 0, sizeof(WOLFSSL_X509_VERIFY_PARAM));
-
-        ExpectIntEQ(X509_VERIFY_PARAM_set1_host(NULL, host, sizeof(host)),
-            WOLFSSL_FAILURE);
-
-        X509_VERIFY_PARAM_set1_host(pParam, host, sizeof(host));
-
-        ExpectIntEQ(XMEMCMP(pParam->hostName, host, sizeof(host)), 0);
-
-        XMEMSET(pParam, 0, sizeof(WOLFSSL_X509_VERIFY_PARAM));
-
-        ExpectIntNE(XMEMCMP(pParam->hostName, host, sizeof(host)), 0);
-
-        XFREE(pParam, HEAP_HINT, DYNAMIC_TYPE_OPENSSL);
-    }
-#endif /* OPENSSL_EXTRA */
-    return EXPECT_RESULT();
-}
-
 static int test_wolfSSL_set1_host(void)
 {
     EXPECT_DECLS;
@@ -19238,110 +17287,6 @@ static int test_wolfSSL_set1_host(void)
     SSL_free(ssl);
     SSL_CTX_free(ctx);
 #endif
-    return EXPECT_RESULT();
-}
-
-static int test_wolfSSL_X509_VERIFY_PARAM_set1_ip(void)
-{
-    EXPECT_DECLS;
-#if defined(OPENSSL_EXTRA) && !defined(NO_FILESYSTEM)
-    unsigned char buf[16] = {0};
-    WOLFSSL_X509_VERIFY_PARAM* param = NULL;
-
-    ExpectNotNull(param = X509_VERIFY_PARAM_new());
-
-    ExpectIntEQ(X509_VERIFY_PARAM_set1_ip(NULL, NULL, 1), WOLFSSL_FAILURE);
-    ExpectIntEQ(X509_VERIFY_PARAM_set1_ip(param, NULL, 1), WOLFSSL_FAILURE);
-    ExpectIntEQ(X509_VERIFY_PARAM_set1_ip(NULL, buf, 1), WOLFSSL_FAILURE);
-    ExpectIntEQ(X509_VERIFY_PARAM_set1_ip(NULL, NULL, 16), WOLFSSL_FAILURE);
-    ExpectIntEQ(X509_VERIFY_PARAM_set1_ip(NULL, NULL, 4), WOLFSSL_FAILURE);
-    ExpectIntEQ(X509_VERIFY_PARAM_set1_ip(NULL, NULL, 0), WOLFSSL_FAILURE);
-    ExpectIntEQ(X509_VERIFY_PARAM_set1_ip(param, buf, 1), WOLFSSL_FAILURE);
-    ExpectIntEQ(X509_VERIFY_PARAM_set1_ip(param, NULL, 16), WOLFSSL_FAILURE);
-    ExpectIntEQ(X509_VERIFY_PARAM_set1_ip(param, NULL, 4), WOLFSSL_FAILURE);
-    ExpectIntEQ(X509_VERIFY_PARAM_set1_ip(NULL, buf, 16), WOLFSSL_FAILURE);
-    ExpectIntEQ(X509_VERIFY_PARAM_set1_ip(NULL, buf, 4), WOLFSSL_FAILURE);
-    ExpectIntEQ(X509_VERIFY_PARAM_set1_ip(NULL, buf, 0), WOLFSSL_FAILURE);
-
-    ExpectIntEQ(X509_VERIFY_PARAM_set1_ip(param, NULL, 0), WOLFSSL_SUCCESS);
-
-    /* test 127.0.0.1 */
-    buf[0] =0x7f; buf[1] = 0; buf[2] = 0; buf[3] = 1;
-    ExpectIntEQ(X509_VERIFY_PARAM_set1_ip(param, &buf[0], 4), SSL_SUCCESS);
-    ExpectIntEQ(XSTRNCMP(param->ipasc, "127.0.0.1", sizeof(param->ipasc)), 0);
-
-    /* test 2001:db8:3333:4444:5555:6666:7777:8888 */
-    buf[0]=32;buf[1]=1;buf[2]=13;buf[3]=184;
-    buf[4]=51;buf[5]=51;buf[6]=68;buf[7]=68;
-    buf[8]=85;buf[9]=85;buf[10]=102;buf[11]=102;
-    buf[12]=119;buf[13]=119;buf[14]=136;buf[15]=136;
-    ExpectIntEQ(X509_VERIFY_PARAM_set1_ip(param, &buf[0], 16), SSL_SUCCESS);
-    ExpectIntEQ(XSTRNCMP(param->ipasc,
-        "2001:db8:3333:4444:5555:6666:7777:8888", sizeof(param->ipasc)), 0);
-
-    /* test 2001:db8:: */
-    buf[0]=32;buf[1]=1;buf[2]=13;buf[3]=184;
-    buf[4]=0;buf[5]=0;buf[6]=0;buf[7]=0;
-    buf[8]=0;buf[9]=0;buf[10]=0;buf[11]=0;
-    buf[12]=0;buf[13]=0;buf[14]=0;buf[15]=0;
-    ExpectIntEQ(X509_VERIFY_PARAM_set1_ip(param, &buf[0], 16), SSL_SUCCESS);
-    ExpectIntEQ(XSTRNCMP(param->ipasc, "2001:db8::", sizeof(param->ipasc)), 0);
-
-    /* test ::1234:5678 */
-    buf[0]=0;buf[1]=0;buf[2]=0;buf[3]=0;
-    buf[4]=0;buf[5]=0;buf[6]=0;buf[7]=0;
-    buf[8]=0;buf[9]=0;buf[10]=0;buf[11]=0;
-    buf[12]=18;buf[13]=52;buf[14]=86;buf[15]=120;
-    ExpectIntEQ(X509_VERIFY_PARAM_set1_ip(param, &buf[0], 16), SSL_SUCCESS);
-    ExpectIntEQ(XSTRNCMP(param->ipasc, "::1234:5678", sizeof(param->ipasc)), 0);
-
-
-    /* test 2001:db8::1234:5678 */
-    buf[0]=32;buf[1]=1;buf[2]=13;buf[3]=184;
-    buf[4]=0;buf[5]=0;buf[6]=0;buf[7]=0;
-    buf[8]=0;buf[9]=0;buf[10]=0;buf[11]=0;
-    buf[12]=18;buf[13]=52;buf[14]=86;buf[15]=120;
-    ExpectIntEQ(X509_VERIFY_PARAM_set1_ip(param, &buf[0], 16), SSL_SUCCESS);
-    ExpectIntEQ(XSTRNCMP(param->ipasc, "2001:db8::1234:5678",
-                                                sizeof(param->ipasc)), 0);
-
-    /* test 2001:0db8:0001:0000:0000:0ab9:c0a8:0102*/
-    /*      2001:db8:1::ab9:c0a8:102 */
-    buf[0]=32;buf[1]=1;buf[2]=13;buf[3]=184;
-    buf[4]=0;buf[5]=1;buf[6]=0;buf[7]=0;
-    buf[8]=0;buf[9]=0;buf[10]=10;buf[11]=185;
-    buf[12]=192;buf[13]=168;buf[14]=1;buf[15]=2;
-    ExpectIntEQ(X509_VERIFY_PARAM_set1_ip(param, &buf[0], 16), SSL_SUCCESS);
-    ExpectIntEQ(XSTRNCMP(param->ipasc, "2001:db8:1::ab9:c0a8:102",
-                                                sizeof(param->ipasc)), 0);
-
-    XFREE(param, HEAP_HINT, DYNAMIC_TYPE_OPENSSL);
-#endif /* OPENSSL_EXTRA */
-    return EXPECT_RESULT();
-}
-
-static int test_wolfSSL_X509_STORE_CTX_get0_store(void)
-{
-    EXPECT_DECLS;
-#if defined(OPENSSL_EXTRA)
-    X509_STORE* store = NULL;
-    X509_STORE_CTX* ctx = NULL;
-    X509_STORE_CTX* ctx_no_init = NULL;
-
-    ExpectNotNull((store = X509_STORE_new()));
-    ExpectNotNull(ctx = X509_STORE_CTX_new());
-    ExpectNotNull(ctx_no_init = X509_STORE_CTX_new());
-    ExpectIntEQ(X509_STORE_CTX_init(ctx, store, NULL, NULL), SSL_SUCCESS);
-
-    ExpectNull(X509_STORE_CTX_get0_store(NULL));
-    /* should return NULL if ctx has not bee initialized */
-    ExpectNull(X509_STORE_CTX_get0_store(ctx_no_init));
-    ExpectNotNull(X509_STORE_CTX_get0_store(ctx));
-
-    wolfSSL_X509_STORE_CTX_free(ctx);
-    wolfSSL_X509_STORE_CTX_free(ctx_no_init);
-    X509_STORE_free(store);
-#endif /* OPENSSL_EXTRA */
     return EXPECT_RESULT();
 }
 
@@ -20189,323 +18134,6 @@ static int test_wolfSSL_CTX_set_srp_password(void)
     wolfSSL_CTX_free(ctx);
 #endif /* OPENSSL_EXTRA && WOLFCRYPT_HAVE_SRP */
        /* && !NO_SHA256 && !WC_NO_RNG && !NO_WOLFSSL_CLIENT */
-    return EXPECT_RESULT();
-}
-
-static int test_wolfSSL_X509_STORE(void)
-{
-    EXPECT_DECLS;
-#if defined(OPENSSL_EXTRA) && !defined(NO_RSA) && !defined(NO_TLS) && \
-    !defined(NO_FILESYSTEM)
-    X509_STORE *store = NULL;
-
-#ifdef HAVE_CRL
-    X509_STORE_CTX *storeCtx = NULL;
-    X509 *ca = NULL;
-    X509 *cert = NULL;
-    const char srvCert[] = "./certs/server-revoked-cert.pem";
-    const char caCert[] = "./certs/ca-cert.pem";
-#ifndef WOLFSSL_CRL_ALLOW_MISSING_CDP
-    X509_CRL *crl = NULL;
-    const char crlPem[] = "./certs/crl/crl.revoked";
-    XFILE fp = XBADFILE;
-#endif /* !WOLFSSL_CRL_ALLOW_MISSING_CDP */
-
-    ExpectNotNull(store = (X509_STORE *)X509_STORE_new());
-    ExpectNotNull((ca = wolfSSL_X509_load_certificate_file(caCert,
-                           SSL_FILETYPE_PEM)));
-    ExpectIntEQ(X509_STORE_add_cert(store, ca), SSL_SUCCESS);
-    ExpectNotNull((cert = wolfSSL_X509_load_certificate_file(srvCert,
-                    SSL_FILETYPE_PEM)));
-    ExpectNotNull((storeCtx = X509_STORE_CTX_new()));
-    ExpectIntEQ(X509_STORE_CTX_init(storeCtx, store, cert, NULL), SSL_SUCCESS);
-    ExpectIntEQ(X509_verify_cert(storeCtx), SSL_SUCCESS);
-    X509_STORE_free(store);
-    store = NULL;
-    X509_STORE_CTX_free(storeCtx);
-    storeCtx = NULL;
-    X509_free(cert);
-    cert = NULL;
-    X509_free(ca);
-    ca = NULL;
-
-#ifndef WOLFSSL_CRL_ALLOW_MISSING_CDP
-    /* should fail to verify now after adding in CRL */
-    ExpectNotNull(store = (X509_STORE *)X509_STORE_new());
-    ExpectNotNull((ca = wolfSSL_X509_load_certificate_file(caCert,
-                           SSL_FILETYPE_PEM)));
-    ExpectIntEQ(X509_STORE_add_cert(store, ca), SSL_SUCCESS);
-    ExpectTrue((fp = XFOPEN(crlPem, "rb")) != XBADFILE);
-    ExpectNotNull(crl = (X509_CRL *)PEM_read_X509_CRL(fp, (X509_CRL **)NULL,
-                NULL, NULL));
-    if (fp != XBADFILE)
-        XFCLOSE(fp);
-    ExpectIntEQ(X509_STORE_add_crl(store, crl), SSL_SUCCESS);
-    ExpectIntEQ(X509_STORE_set_flags(store, X509_V_FLAG_CRL_CHECK),SSL_SUCCESS);
-    ExpectNotNull((storeCtx = X509_STORE_CTX_new()));
-    ExpectNotNull((cert = wolfSSL_X509_load_certificate_file(srvCert,
-                    SSL_FILETYPE_PEM)));
-    ExpectIntEQ(X509_STORE_CTX_init(storeCtx, store, cert, NULL), SSL_SUCCESS);
-    ExpectIntNE(X509_verify_cert(storeCtx), SSL_SUCCESS);
-    ExpectIntEQ(X509_STORE_CTX_get_error(storeCtx),
-                WOLFSSL_X509_V_ERR_CERT_REVOKED);
-    X509_CRL_free(crl);
-    crl = NULL;
-    X509_STORE_free(store);
-    store = NULL;
-    X509_STORE_CTX_free(storeCtx);
-    storeCtx = NULL;
-    X509_free(cert);
-    cert = NULL;
-    X509_free(ca);
-    ca = NULL;
-#endif /* !WOLFSSL_CRL_ALLOW_MISSING_CDP */
-#endif /* HAVE_CRL */
-
-
-
-#if !defined(WOLFCRYPT_ONLY) && !defined(NO_FILESYSTEM)
-    {
-    #if !defined(NO_WOLFSSL_CLIENT) || !defined(NO_WOLFSSL_SERVER)
-        SSL_CTX* ctx = NULL;
-        SSL* ssl = NULL;
-        int i;
-        for (i = 0; i < 2; i++) {
-        #ifndef NO_WOLFSSL_SERVER
-            ExpectNotNull(ctx = SSL_CTX_new(wolfSSLv23_server_method()));
-        #else
-            ExpectNotNull(ctx = SSL_CTX_new(wolfSSLv23_client_method()));
-        #endif
-            ExpectNotNull(store = (X509_STORE *)X509_STORE_new());
-            SSL_CTX_set_cert_store(ctx, store);
-            ExpectNotNull(store = (X509_STORE *)X509_STORE_new());
-            SSL_CTX_set_cert_store(ctx, store);
-            ExpectNotNull(store = (X509_STORE *)X509_STORE_new());
-            ExpectIntEQ(SSL_CTX_use_certificate_file(ctx, svrCertFile,
-                    SSL_FILETYPE_PEM), SSL_SUCCESS);
-            ExpectIntEQ(SSL_CTX_use_PrivateKey_file(ctx, svrKeyFile,
-                    SSL_FILETYPE_PEM), SSL_SUCCESS);
-            ExpectNotNull(ssl = SSL_new(ctx));
-            if (i == 0) {
-                ExpectIntEQ(SSL_set0_verify_cert_store(ssl, store),
-                    SSL_SUCCESS);
-            }
-            else {
-                ExpectIntEQ(SSL_set1_verify_cert_store(ssl, store), SSL_SUCCESS);
-                #ifdef OPENSSL_ALL
-                ExpectIntEQ(SSL_CTX_set1_verify_cert_store(ctx, store), SSL_SUCCESS);
-                #endif
-            }
-            if (EXPECT_FAIL() || (i == 1)) {
-                X509_STORE_free(store);
-                store = NULL;
-            }
-            SSL_free(ssl);
-            ssl = NULL;
-            SSL_CTX_free(ctx);
-            ctx = NULL;
-        }
-    #endif /* !NO_WOLFSSL_CLIENT || !NO_WOLFSSL_SERVER */
-    }
-#endif
-#endif
-    return EXPECT_RESULT();
-}
-
-static int test_wolfSSL_X509_STORE_load_locations(void)
-{
-    EXPECT_DECLS;
-#if (defined(OPENSSL_ALL) || defined(WOLFSSL_APACHE_HTTPD)) && \
-    !defined(NO_FILESYSTEM) && !defined(NO_WOLFSSL_DIR) && !defined(NO_RSA) && \
-    !defined(NO_TLS)
-    SSL_CTX *ctx = NULL;
-    X509_STORE *store = NULL;
-
-    const char ca_file[] = "./certs/ca-cert.pem";
-    const char client_pem_file[] = "./certs/client-cert.pem";
-    const char client_der_file[] = "./certs/client-cert.der";
-    const char ecc_file[] = "./certs/ecc-key.pem";
-    const char certs_path[] = "./certs/";
-    const char bad_path[] = "./bad-path/";
-#ifdef HAVE_CRL
-    const char crl_path[] = "./certs/crl/";
-    const char crl_file[] = "./certs/crl/crl.pem";
-#endif
-
-#ifndef NO_WOLFSSL_SERVER
-    ExpectNotNull(ctx = SSL_CTX_new(SSLv23_server_method()));
-#else
-    ExpectNotNull(ctx = SSL_CTX_new(SSLv23_client_method()));
-#endif
-    ExpectNotNull(store = SSL_CTX_get_cert_store(ctx));
-    ExpectIntEQ(wolfSSL_CertManagerLoadCA(store->cm, ca_file, NULL),
-        WOLFSSL_SUCCESS);
-
-    /* Test bad arguments */
-    ExpectIntEQ(X509_STORE_load_locations(NULL, ca_file, NULL),
-        WC_NO_ERR_TRACE(WOLFSSL_FAILURE));
-    ExpectIntEQ(X509_STORE_load_locations(store, NULL, NULL), WC_NO_ERR_TRACE(WOLFSSL_FAILURE));
-    ExpectIntEQ(X509_STORE_load_locations(store, client_der_file, NULL),
-        WC_NO_ERR_TRACE(WOLFSSL_FAILURE));
-    ExpectIntEQ(X509_STORE_load_locations(store, ecc_file, NULL),
-        WC_NO_ERR_TRACE(WOLFSSL_FAILURE));
-    ExpectIntEQ(X509_STORE_load_locations(store, NULL, bad_path),
-        WC_NO_ERR_TRACE(WOLFSSL_FAILURE));
-
-#ifdef HAVE_CRL
-    /* Test with CRL */
-    ExpectIntEQ(X509_STORE_load_locations(store, crl_file, NULL),
-        WOLFSSL_SUCCESS);
-    ExpectIntEQ(X509_STORE_load_locations(store, NULL, crl_path),
-        WOLFSSL_SUCCESS);
-#endif
-
-    /* Test with CA */
-    ExpectIntEQ(X509_STORE_load_locations(store, ca_file, NULL),
-        WOLFSSL_SUCCESS);
-
-    /* Test with client_cert and certs path */
-    ExpectIntEQ(X509_STORE_load_locations(store, client_pem_file, NULL),
-        WOLFSSL_SUCCESS);
-    ExpectIntEQ(X509_STORE_load_locations(store, NULL, certs_path),
-        WOLFSSL_SUCCESS);
-
-#if defined(XGETENV) && !defined(NO_GETENV) && defined(_POSIX_C_SOURCE) && \
-    _POSIX_C_SOURCE >= 200112L
-    ExpectIntEQ(wolfSSL_CTX_UnloadCAs(ctx), WOLFSSL_SUCCESS);
-    /* Test with env vars */
-    ExpectIntEQ(setenv("SSL_CERT_FILE", client_pem_file, 1), 0);
-    ExpectIntEQ(setenv("SSL_CERT_DIR", certs_path, 1), 0);
-    ExpectIntEQ(X509_STORE_set_default_paths(store), WOLFSSL_SUCCESS);
-#endif
-
-#if defined(OPENSSL_EXTRA) || defined(DEBUG_WOLFSSL_VERBOSE)
-    /* Clear nodes */
-    ERR_clear_error();
-#endif
-
-    SSL_CTX_free(ctx);
-#endif
-    return EXPECT_RESULT();
-}
-
-static int test_X509_STORE_get0_objects(void)
-{
-    EXPECT_DECLS;
-#if defined(OPENSSL_ALL) && !defined(NO_FILESYSTEM) && !defined(NO_TLS) && \
-    !defined(NO_WOLFSSL_DIR) && !defined(NO_RSA)
-    X509_STORE *store = NULL;
-    X509_STORE *store_cpy = NULL;
-    SSL_CTX *ctx = NULL;
-    X509_OBJECT *obj = NULL;
-#ifdef HAVE_CRL
-    X509_OBJECT *objCopy = NULL;
-#endif
-    STACK_OF(X509_OBJECT) *objs = NULL;
-    STACK_OF(X509_OBJECT) *objsCopy = NULL;
-    int i;
-
-    /* Setup store */
-#ifndef NO_WOLFSSL_SERVER
-    ExpectNotNull(ctx = SSL_CTX_new(SSLv23_server_method()));
-#else
-    ExpectNotNull(ctx = SSL_CTX_new(SSLv23_client_method()));
-#endif
-    ExpectNotNull(store_cpy = X509_STORE_new());
-    ExpectNotNull(store = SSL_CTX_get_cert_store(ctx));
-    ExpectIntEQ(X509_STORE_load_locations(store, cliCertFile, NULL),
-        WOLFSSL_SUCCESS);
-    ExpectIntEQ(X509_STORE_load_locations(store, caCertFile, NULL),
-        WOLFSSL_SUCCESS);
-    ExpectIntEQ(X509_STORE_load_locations(store, svrCertFile, NULL),
-        WOLFSSL_SUCCESS);
-#ifdef HAVE_CRL
-    ExpectIntEQ(X509_STORE_load_locations(store, NULL, crlPemDir),
-        WOLFSSL_SUCCESS);
-#endif
-    /* Store ready */
-
-    /* Similar to HaProxy ssl_set_cert_crl_file use case */
-    ExpectNotNull(objs = X509_STORE_get0_objects(store));
-#ifdef HAVE_CRL
-#ifdef WOLFSSL_SIGNER_DER_CERT
-    ExpectIntEQ(sk_X509_OBJECT_num(objs), 4);
-#else
-    ExpectIntEQ(sk_X509_OBJECT_num(objs), 1);
-#endif
-#else
-#ifdef WOLFSSL_SIGNER_DER_CERT
-    ExpectIntEQ(sk_X509_OBJECT_num(objs), 3);
-#else
-    ExpectIntEQ(sk_X509_OBJECT_num(objs), 0);
-#endif
-#endif
-    ExpectIntEQ(sk_X509_OBJECT_num(NULL), 0);
-    ExpectNull(sk_X509_OBJECT_value(NULL, 0));
-    ExpectNull(sk_X509_OBJECT_value(NULL, 1));
-    ExpectNull(sk_X509_OBJECT_value(objs, sk_X509_OBJECT_num(objs)));
-    ExpectNull(sk_X509_OBJECT_value(objs, sk_X509_OBJECT_num(objs) + 1));
-#ifndef NO_WOLFSSL_STUB
-    ExpectNull(sk_X509_OBJECT_delete(objs, 0));
-#endif
-    ExpectNotNull(objsCopy = sk_X509_OBJECT_deep_copy(objs, NULL, NULL));
-    ExpectIntEQ(sk_X509_OBJECT_num(objs), sk_X509_OBJECT_num(objsCopy));
-    for (i = 0; i < sk_X509_OBJECT_num(objs) && EXPECT_SUCCESS(); i++) {
-        obj = (X509_OBJECT*)sk_X509_OBJECT_value(objs, i);
-    #ifdef HAVE_CRL
-        objCopy = (X509_OBJECT*)sk_X509_OBJECT_value(objsCopy, i);
-    #endif
-        switch (X509_OBJECT_get_type(obj)) {
-        case X509_LU_X509:
-        {
-            X509* x509 = NULL;
-            X509_NAME *subj_name = NULL;
-            ExpectNull(X509_OBJECT_get0_X509_CRL(NULL));
-            ExpectNull(X509_OBJECT_get0_X509_CRL(obj));
-            ExpectNotNull(x509 = X509_OBJECT_get0_X509(obj));
-            ExpectIntEQ(X509_STORE_add_cert(store_cpy, x509), WOLFSSL_SUCCESS);
-            ExpectNotNull(subj_name = X509_get_subject_name(x509));
-            ExpectPtrEq(obj, X509_OBJECT_retrieve_by_subject(objs, X509_LU_X509,
-                    subj_name));
-
-            break;
-        }
-        case X509_LU_CRL:
-#ifdef HAVE_CRL
-        {
-            X509_CRL* crl = NULL;
-            ExpectNull(X509_OBJECT_get0_X509(NULL));
-            ExpectNull(X509_OBJECT_get0_X509(obj));
-            ExpectNotNull(crl = X509_OBJECT_get0_X509_CRL(obj));
-            ExpectIntEQ(X509_STORE_add_crl(store_cpy, crl), WOLFSSL_SUCCESS);
-
-            ExpectNotNull(crl = X509_OBJECT_get0_X509_CRL(objCopy));
-            break;
-        }
-#endif
-        case X509_LU_NONE:
-        default:
-            Fail(("X509_OBJECT_get_type should return x509 or crl "
-                    "(when built with crl support)"),
-                    ("Unrecognized X509_OBJECT type or none"));
-        }
-    }
-
-    X509_STORE_free(store_cpy);
-    SSL_CTX_free(ctx);
-
-    wolfSSL_sk_X509_OBJECT_free(NULL);
-    objs = NULL;
-    wolfSSL_sk_pop_free(objsCopy, NULL);
-    objsCopy = NULL;
-    ExpectNotNull(objs = wolfSSL_sk_X509_OBJECT_new());
-    ExpectIntEQ(wolfSSL_sk_X509_OBJECT_push(NULL, NULL), WOLFSSL_FAILURE);
-    ExpectIntEQ(wolfSSL_sk_X509_OBJECT_push(objs, NULL), WOLFSSL_FAILURE);
-    ExpectIntEQ(wolfSSL_sk_X509_OBJECT_push(NULL, obj), WOLFSSL_FAILURE);
-    ExpectNotNull(objsCopy = sk_X509_OBJECT_deep_copy(objs, NULL, NULL));
-    wolfSSL_sk_X509_OBJECT_free(objsCopy);
-    wolfSSL_sk_X509_OBJECT_free(objs);
-#endif
     return EXPECT_RESULT();
 }
 
@@ -21537,468 +19165,6 @@ static int test_wolfSSL_a2i_IPADDRESS(void)
     return EXPECT_RESULT();
 }
 
-static int test_wolfSSL_X509_sign2(void)
-{
-    EXPECT_DECLS;
-    /* test requires WOLFSSL_AKID_NAME to match expected output */
-#if defined(OPENSSL_EXTRA) && !defined(NO_RSA) && !defined(NO_CERTS) && \
-    defined(WOLFSSL_CERT_GEN) && defined(WOLFSSL_ALT_NAMES) && \
-    defined(WOLFSSL_CERT_EXT) && defined(WOLFSSL_AKID_NAME) && \
-    (defined(WOLFSSL_QT) || defined(OPENSSL_ALL) || \
-    defined(WOLFSSL_IP_ALT_NAME))
-    WOLFSSL_X509 *x509 = NULL;
-    WOLFSSL_X509 *ca = NULL;
-    const unsigned char *der = NULL;
-    const unsigned char *pt = NULL;
-    WOLFSSL_EVP_PKEY  *priv = NULL;
-    WOLFSSL_X509_NAME *name = NULL;
-    int derSz;
-#ifndef NO_ASN_TIME
-    WOLFSSL_ASN1_TIME *notBefore = NULL;
-    WOLFSSL_ASN1_TIME *notAfter = NULL;
-
-    const int year = 365*24*60*60;
-    const int day  = 24*60*60;
-    const int hour = 60*60;
-    const int mini = 60;
-    time_t t;
-#endif
-
-    const unsigned char expected[] = {
-        0x30, 0x82, 0x05, 0x13, 0x30, 0x82, 0x03, 0xFB, 0xA0, 0x03, 0x02, 0x01,
-        0x02, 0x02, 0x14, 0x6B, 0x61, 0x49, 0x45, 0xFF, 0x4A, 0xD1, 0x54, 0x16,
-        0xB4, 0x35, 0x37, 0xC4, 0x98, 0x5D, 0xA9, 0xF6, 0x67, 0x60, 0x91, 0x30,
-        0x0D, 0x06, 0x09, 0x2A, 0x86, 0x48, 0x86, 0xF7, 0x0D, 0x01, 0x01, 0x0B,
-        0x05, 0x00, 0x30, 0x81, 0x94, 0x31, 0x0B, 0x30, 0x09, 0x06, 0x03, 0x55,
-        0x04, 0x06, 0x13, 0x02, 0x55, 0x53, 0x31, 0x10, 0x30, 0x0E, 0x06, 0x03,
-        0x55, 0x04, 0x08, 0x0C, 0x07, 0x4D, 0x6F, 0x6E, 0x74, 0x61, 0x6E, 0x61,
-        0x31, 0x10, 0x30, 0x0E, 0x06, 0x03, 0x55, 0x04, 0x07, 0x0C, 0x07, 0x42,
-        0x6F, 0x7A, 0x65, 0x6D, 0x61, 0x6E, 0x31, 0x11, 0x30, 0x0F, 0x06, 0x03,
-        0x55, 0x04, 0x0A, 0x0C, 0x08, 0x53, 0x61, 0x77, 0x74, 0x6F, 0x6F, 0x74,
-        0x68, 0x31, 0x13, 0x30, 0x11, 0x06, 0x03, 0x55, 0x04, 0x0B, 0x0C, 0x0A,
-        0x43, 0x6F, 0x6E, 0x73, 0x75, 0x6C, 0x74, 0x69, 0x6E, 0x67, 0x31, 0x18,
-        0x30, 0x16, 0x06, 0x03, 0x55, 0x04, 0x03, 0x0C, 0x0F, 0x77, 0x77, 0x77,
-        0x2E, 0x77, 0x6F, 0x6C, 0x66, 0x73, 0x73, 0x6C, 0x2E, 0x63, 0x6F, 0x6D,
-        0x31, 0x1F, 0x30, 0x1D, 0x06, 0x09, 0x2A, 0x86, 0x48, 0x86, 0xF7, 0x0D,
-        0x01, 0x09, 0x01, 0x16, 0x10, 0x69, 0x6E, 0x66, 0x6F, 0x40, 0x77, 0x6F,
-        0x6C, 0x66, 0x73, 0x73, 0x6C, 0x2E, 0x63, 0x6F, 0x6D, 0x30, 0x1E, 0x17,
-        0x0D, 0x30, 0x30, 0x30, 0x32, 0x31, 0x35, 0x32, 0x30, 0x33, 0x30, 0x30,
-        0x30, 0x5A, 0x17, 0x0D, 0x30, 0x31, 0x30, 0x32, 0x31, 0x34, 0x32, 0x30,
-        0x33, 0x30, 0x30, 0x30, 0x5A, 0x30, 0x81, 0x9E, 0x31, 0x0B, 0x30, 0x09,
-        0x06, 0x03, 0x55, 0x04, 0x06, 0x13, 0x02, 0x55, 0x53, 0x31, 0x10, 0x30,
-        0x0E, 0x06, 0x03, 0x55, 0x04, 0x08, 0x0C, 0x07, 0x4D, 0x6F, 0x6E, 0x74,
-        0x61, 0x6E, 0x61, 0x31, 0x10, 0x30, 0x0E, 0x06, 0x03, 0x55, 0x04, 0x07,
-        0x0C, 0x07, 0x42, 0x6F, 0x7A, 0x65, 0x6D, 0x61, 0x6E, 0x31, 0x15, 0x30,
-        0x13, 0x06, 0x03, 0x55, 0x04, 0x0A, 0x0C, 0x0C, 0x77, 0x6F, 0x6C, 0x66,
-        0x53, 0x53, 0x4C, 0x5F, 0x32, 0x30, 0x34, 0x38, 0x31, 0x19, 0x30, 0x17,
-        0x06, 0x03, 0x55, 0x04, 0x0B, 0x0C, 0x10, 0x50, 0x72, 0x6F, 0x67, 0x72,
-        0x61, 0x6D, 0x6D, 0x69, 0x6E, 0x67, 0x2D, 0x32, 0x30, 0x34, 0x38, 0x31,
-        0x18, 0x30, 0x16, 0x06, 0x03, 0x55, 0x04, 0x03, 0x0C, 0x0F, 0x77, 0x77,
-        0x77, 0x2E, 0x77, 0x6F, 0x6C, 0x66, 0x73, 0x73, 0x6C, 0x2E, 0x63, 0x6F,
-        0x6D, 0x31, 0x1F, 0x30, 0x1D, 0x06, 0x09, 0x2A, 0x86, 0x48, 0x86, 0xF7,
-        0x0D, 0x01, 0x09, 0x01, 0x16, 0x10, 0x69, 0x6E, 0x66, 0x6F, 0x40, 0x77,
-        0x6F, 0x6C, 0x66, 0x73, 0x73, 0x6C, 0x2E, 0x63, 0x6F, 0x6D, 0x30, 0x82,
-        0x01, 0x22, 0x30, 0x0D, 0x06, 0x09, 0x2A, 0x86, 0x48, 0x86, 0xF7, 0x0D,
-        0x01, 0x01, 0x01, 0x05, 0x00, 0x03, 0x82, 0x01, 0x0F, 0x00, 0x30, 0x82,
-        0x01, 0x0A, 0x02, 0x82, 0x01, 0x01, 0x00, 0xC3, 0x03, 0xD1, 0x2B, 0xFE,
-        0x39, 0xA4, 0x32, 0x45, 0x3B, 0x53, 0xC8, 0x84, 0x2B, 0x2A, 0x7C, 0x74,
-        0x9A, 0xBD, 0xAA, 0x2A, 0x52, 0x07, 0x47, 0xD6, 0xA6, 0x36, 0xB2, 0x07,
-        0x32, 0x8E, 0xD0, 0xBA, 0x69, 0x7B, 0xC6, 0xC3, 0x44, 0x9E, 0xD4, 0x81,
-        0x48, 0xFD, 0x2D, 0x68, 0xA2, 0x8B, 0x67, 0xBB, 0xA1, 0x75, 0xC8, 0x36,
-        0x2C, 0x4A, 0xD2, 0x1B, 0xF7, 0x8B, 0xBA, 0xCF, 0x0D, 0xF9, 0xEF, 0xEC,
-        0xF1, 0x81, 0x1E, 0x7B, 0x9B, 0x03, 0x47, 0x9A, 0xBF, 0x65, 0xCC, 0x7F,
-        0x65, 0x24, 0x69, 0xA6, 0xE8, 0x14, 0x89, 0x5B, 0xE4, 0x34, 0xF7, 0xC5,
-        0xB0, 0x14, 0x93, 0xF5, 0x67, 0x7B, 0x3A, 0x7A, 0x78, 0xE1, 0x01, 0x56,
-        0x56, 0x91, 0xA6, 0x13, 0x42, 0x8D, 0xD2, 0x3C, 0x40, 0x9C, 0x4C, 0xEF,
-        0xD1, 0x86, 0xDF, 0x37, 0x51, 0x1B, 0x0C, 0xA1, 0x3B, 0xF5, 0xF1, 0xA3,
-        0x4A, 0x35, 0xE4, 0xE1, 0xCE, 0x96, 0xDF, 0x1B, 0x7E, 0xBF, 0x4E, 0x97,
-        0xD0, 0x10, 0xE8, 0xA8, 0x08, 0x30, 0x81, 0xAF, 0x20, 0x0B, 0x43, 0x14,
-        0xC5, 0x74, 0x67, 0xB4, 0x32, 0x82, 0x6F, 0x8D, 0x86, 0xC2, 0x88, 0x40,
-        0x99, 0x36, 0x83, 0xBA, 0x1E, 0x40, 0x72, 0x22, 0x17, 0xD7, 0x52, 0x65,
-        0x24, 0x73, 0xB0, 0xCE, 0xEF, 0x19, 0xCD, 0xAE, 0xFF, 0x78, 0x6C, 0x7B,
-        0xC0, 0x12, 0x03, 0xD4, 0x4E, 0x72, 0x0D, 0x50, 0x6D, 0x3B, 0xA3, 0x3B,
-        0xA3, 0x99, 0x5E, 0x9D, 0xC8, 0xD9, 0x0C, 0x85, 0xB3, 0xD9, 0x8A, 0xD9,
-        0x54, 0x26, 0xDB, 0x6D, 0xFA, 0xAC, 0xBB, 0xFF, 0x25, 0x4C, 0xC4, 0xD1,
-        0x79, 0xF4, 0x71, 0xD3, 0x86, 0x40, 0x18, 0x13, 0xB0, 0x63, 0xB5, 0x72,
-        0x4E, 0x30, 0xC4, 0x97, 0x84, 0x86, 0x2D, 0x56, 0x2F, 0xD7, 0x15, 0xF7,
-        0x7F, 0xC0, 0xAE, 0xF5, 0xFC, 0x5B, 0xE5, 0xFB, 0xA1, 0xBA, 0xD3, 0x02,
-        0x03, 0x01, 0x00, 0x01, 0xA3, 0x82, 0x01, 0x4F, 0x30, 0x82, 0x01, 0x4B,
-        0x30, 0x0C, 0x06, 0x03, 0x55, 0x1D, 0x13, 0x04, 0x05, 0x30, 0x03, 0x01,
-        0x01, 0xFF, 0x30, 0x1C, 0x06, 0x03, 0x55, 0x1D, 0x11, 0x04, 0x15, 0x30,
-        0x13, 0x82, 0x0B, 0x65, 0x78, 0x61, 0x6D, 0x70, 0x6C, 0x65, 0x2E, 0x63,
-        0x6F, 0x6D, 0x87, 0x04, 0x7F, 0x00, 0x00, 0x01, 0x30, 0x1D, 0x06, 0x03,
-        0x55, 0x1D, 0x0E, 0x04, 0x16, 0x04, 0x14, 0x33, 0xD8, 0x45, 0x66, 0xD7,
-        0x68, 0x87, 0x18, 0x7E, 0x54, 0x0D, 0x70, 0x27, 0x91, 0xC7, 0x26, 0xD7,
-        0x85, 0x65, 0xC0, 0x30, 0x81, 0xDE, 0x06, 0x03, 0x55, 0x1D, 0x23, 0x04,
-        0x81, 0xD6, 0x30, 0x81, 0xD3, 0x80, 0x14, 0x33, 0xD8, 0x45, 0x66, 0xD7,
-        0x68, 0x87, 0x18, 0x7E, 0x54, 0x0D, 0x70, 0x27, 0x91, 0xC7, 0x26, 0xD7,
-        0x85, 0x65, 0xC0, 0xA1, 0x81, 0xA4, 0xA4, 0x81, 0xA1, 0x30, 0x81, 0x9E,
-        0x31, 0x0B, 0x30, 0x09, 0x06, 0x03, 0x55, 0x04, 0x06, 0x13, 0x02, 0x55,
-        0x53, 0x31, 0x10, 0x30, 0x0E, 0x06, 0x03, 0x55, 0x04, 0x08, 0x0C, 0x07,
-        0x4D, 0x6F, 0x6E, 0x74, 0x61, 0x6E, 0x61, 0x31, 0x10, 0x30, 0x0E, 0x06,
-        0x03, 0x55, 0x04, 0x07, 0x0C, 0x07, 0x42, 0x6F, 0x7A, 0x65, 0x6D, 0x61,
-        0x6E, 0x31, 0x15, 0x30, 0x13, 0x06, 0x03, 0x55, 0x04, 0x0A, 0x0C, 0x0C,
-        0x77, 0x6F, 0x6C, 0x66, 0x53, 0x53, 0x4C, 0x5F, 0x32, 0x30, 0x34, 0x38,
-        0x31, 0x19, 0x30, 0x17, 0x06, 0x03, 0x55, 0x04, 0x0B, 0x0C, 0x10, 0x50,
-        0x72, 0x6F, 0x67, 0x72, 0x61, 0x6D, 0x6D, 0x69, 0x6E, 0x67, 0x2D, 0x32,
-        0x30, 0x34, 0x38, 0x31, 0x18, 0x30, 0x16, 0x06, 0x03, 0x55, 0x04, 0x03,
-        0x0C, 0x0F, 0x77, 0x77, 0x77, 0x2E, 0x77, 0x6F, 0x6C, 0x66, 0x73, 0x73,
-        0x6C, 0x2E, 0x63, 0x6F, 0x6D, 0x31, 0x1F, 0x30, 0x1D, 0x06, 0x09, 0x2A,
-        0x86, 0x48, 0x86, 0xF7, 0x0D, 0x01, 0x09, 0x01, 0x16, 0x10, 0x69, 0x6E,
-        0x66, 0x6F, 0x40, 0x77, 0x6F, 0x6C, 0x66, 0x73, 0x73, 0x6C, 0x2E, 0x63,
-        0x6F, 0x6D, 0x82, 0x14, 0x6B, 0x61, 0x49, 0x45, 0xFF, 0x4A, 0xD1, 0x54,
-        0x16, 0xB4, 0x35, 0x37, 0xC4, 0x98, 0x5D, 0xA9, 0xF6, 0x67, 0x60, 0x91,
-        0x30, 0x1D, 0x06, 0x03, 0x55, 0x1D, 0x25, 0x04, 0x16, 0x30, 0x14, 0x06,
-        0x08, 0x2B, 0x06, 0x01, 0x05, 0x05, 0x07, 0x03, 0x01, 0x06, 0x08, 0x2B,
-        0x06, 0x01, 0x05, 0x05, 0x07, 0x03, 0x02, 0x30, 0x0D, 0x06, 0x09, 0x2A,
-        0x86, 0x48, 0x86, 0xF7, 0x0D, 0x01, 0x01, 0x0B, 0x05, 0x00, 0x03, 0x82,
-        0x01, 0x01, 0x00, 0x2F, 0x9F, 0x83, 0x05, 0x15, 0x1E, 0x5D, 0x7C, 0x22,
-        0x12, 0x20, 0xEE, 0x07, 0x35, 0x25, 0x39, 0xDD, 0x34, 0x06, 0xD3, 0x89,
-        0x31, 0x51, 0x8B, 0x9A, 0xE5, 0xE8, 0x60, 0x30, 0x07, 0x7A, 0xBB, 0x17,
-        0xB9, 0x54, 0x72, 0x83, 0xA2, 0x1F, 0x62, 0xE0, 0x18, 0xAC, 0x93, 0x5E,
-        0x63, 0xC7, 0xDD, 0x12, 0x58, 0x96, 0xC7, 0x90, 0x8B, 0x12, 0x50, 0xD2,
-        0x60, 0x0E, 0x24, 0x07, 0x53, 0x55, 0xD7, 0x8E, 0xC9, 0x56, 0x12, 0x28,
-        0xD8, 0xFD, 0x47, 0xE3, 0x13, 0xFB, 0x3C, 0xD6, 0x3D, 0x82, 0x09, 0x7E,
-        0x10, 0x19, 0xE1, 0xCD, 0xCC, 0x4C, 0x78, 0xDF, 0xE5, 0xFB, 0x2C, 0x8C,
-        0x88, 0xF7, 0x5B, 0x99, 0x93, 0xC6, 0xC7, 0x22, 0xA5, 0xFA, 0x76, 0x6C,
-        0xE9, 0xBC, 0x69, 0xBA, 0x02, 0x82, 0x18, 0xAF, 0x47, 0xD0, 0x9C, 0x5F,
-        0xED, 0xAE, 0x5A, 0x95, 0x59, 0x78, 0x86, 0x24, 0x22, 0xB6, 0x81, 0x03,
-        0x58, 0x9A, 0x14, 0x93, 0xDC, 0x24, 0x58, 0xF3, 0xD2, 0x6C, 0x8E, 0xD2,
-        0x6D, 0x8B, 0xE8, 0x4E, 0xC6, 0xA0, 0x2B, 0x0D, 0xDB, 0x1A, 0x76, 0x28,
-        0xA9, 0x8D, 0xFB, 0x51, 0xA6, 0xF0, 0x82, 0x30, 0xEE, 0x78, 0x1C, 0x71,
-        0xA8, 0x11, 0x8A, 0xA5, 0xC3, 0x91, 0xAB, 0x9A, 0x46, 0xFF, 0x8D, 0xCD,
-        0x82, 0x3F, 0x5D, 0xB6, 0x28, 0x46, 0x6D, 0x66, 0xE2, 0xEE, 0x1E, 0x82,
-        0x0D, 0x1A, 0x74, 0x87, 0xFB, 0xFD, 0x96, 0x26, 0x50, 0x09, 0xEC, 0xA7,
-        0x73, 0x89, 0x43, 0x3B, 0x42, 0x2D, 0xA9, 0x6B, 0x0F, 0x61, 0x81, 0x97,
-        0x11, 0x71, 0xF9, 0xDB, 0x9B, 0x69, 0x4B, 0x6E, 0xD3, 0x7D, 0xDA, 0xC6,
-        0x61, 0x9F, 0x39, 0x87, 0x53, 0x52, 0xA8, 0x4D, 0xAD, 0x80, 0x29, 0x6C,
-        0x19, 0xF0, 0x8D, 0xB1, 0x0D, 0x4E, 0xFB, 0x1B, 0xB7, 0xF1, 0x85, 0x49,
-        0x08, 0x2A, 0x94, 0xD0, 0x4E, 0x0B, 0x8F
-    };
-
-    pt = ca_key_der_2048;
-    ExpectNotNull(priv = wolfSSL_d2i_PrivateKey(EVP_PKEY_RSA, NULL, &pt,
-        sizeof_ca_key_der_2048));
-
-    pt = client_cert_der_2048;
-    ExpectNotNull(x509 = wolfSSL_d2i_X509(NULL, &pt,
-        sizeof_client_cert_der_2048));
-
-    pt = ca_cert_der_2048;
-    ExpectNotNull(ca = wolfSSL_d2i_X509(NULL, &pt, sizeof_ca_cert_der_2048));
-    ExpectNotNull(name = wolfSSL_X509_get_subject_name(ca));
-    ExpectIntEQ(wolfSSL_X509_set_issuer_name(x509, name), WOLFSSL_SUCCESS);
-
-#ifndef NO_ASN_TIME
-    t = (time_t)30 * year + 45 * day + 20 * hour + 30 * mini + 7 * day;
-    ExpectNotNull(notBefore = wolfSSL_ASN1_TIME_adj(NULL, t, 0, 0));
-    ExpectNotNull(notAfter = wolfSSL_ASN1_TIME_adj(NULL, t, 365, 0));
-    ExpectIntEQ(notAfter->length, 13);
-
-    ExpectTrue(wolfSSL_X509_set_notBefore(x509, notBefore));
-    ExpectTrue(wolfSSL_X509_set1_notBefore(x509, notBefore));
-    ExpectTrue(wolfSSL_X509_set_notAfter(x509, notAfter));
-    ExpectTrue(wolfSSL_X509_set1_notAfter(x509, notAfter));
-#endif
-
-    ExpectNull(wolfSSL_X509_notBefore(NULL));
-    ExpectNotNull(wolfSSL_X509_notBefore(x509));
-    ExpectNull(wolfSSL_X509_notAfter(NULL));
-    ExpectNotNull(wolfSSL_X509_notAfter(x509));
-
-    ExpectIntGT(wolfSSL_X509_sign(x509, priv, EVP_sha256()), 0);
-    ExpectNotNull((der = wolfSSL_X509_get_der(x509, &derSz)));
-
-    ExpectIntEQ(derSz, sizeof(expected));
-#ifndef NO_ASN_TIME
-    ExpectIntEQ(XMEMCMP(der, expected, derSz), 0);
-#endif
-
-    wolfSSL_X509_free(ca);
-    wolfSSL_X509_free(x509);
-    wolfSSL_EVP_PKEY_free(priv);
-#ifndef NO_ASN_TIME
-    wolfSSL_ASN1_TIME_free(notBefore);
-    wolfSSL_ASN1_TIME_free(notAfter);
-#endif
-#endif
-    return EXPECT_RESULT();
-}
-
-
-static int test_wolfSSL_X509_sign(void)
-{
-    EXPECT_DECLS;
-#if defined(OPENSSL_EXTRA) && !defined(NO_CERTS) && !defined(NO_ASN_TIME) && \
-    defined(WOLFSSL_CERT_GEN) && defined(WOLFSSL_CERT_REQ) && !defined(NO_RSA)
-    int ret = 0;
-    char *cn = NULL;
-    word32 cnSz = 0;
-    X509_NAME *name = NULL;
-    X509_NAME *emptyName = NULL;
-    X509 *x509 = NULL;
-    X509 *ca = NULL;
-    DecodedCert dCert;
-    EVP_PKEY *pub = NULL;
-    EVP_PKEY *priv = NULL;
-    EVP_MD_CTX *mctx = NULL;
-#if defined(USE_CERT_BUFFERS_1024)
-    const unsigned char* rsaPriv = client_key_der_1024;
-    const unsigned char* rsaPub = client_keypub_der_1024;
-    const unsigned char* certIssuer = client_cert_der_1024;
-    long clientKeySz = (long)sizeof_client_key_der_1024;
-    long clientPubKeySz = (long)sizeof_client_keypub_der_1024;
-    long certIssuerSz = (long)sizeof_client_cert_der_1024;
-#elif defined(USE_CERT_BUFFERS_2048)
-    const unsigned char* rsaPriv = client_key_der_2048;
-    const unsigned char* rsaPub = client_keypub_der_2048;
-    const unsigned char* certIssuer = client_cert_der_2048;
-    long clientKeySz = (long)sizeof_client_key_der_2048;
-    long clientPubKeySz = (long)sizeof_client_keypub_der_2048;
-    long certIssuerSz = (long)sizeof_client_cert_der_2048;
-#endif
-    byte sn[16];
-    int snSz = sizeof(sn);
-    int sigSz = 0;
-#ifndef NO_WOLFSSL_STUB
-    const WOLFSSL_ASN1_BIT_STRING* sig = NULL;
-    const WOLFSSL_X509_ALGOR* alg = NULL;
-#endif
-
-    /* Set X509_NAME fields */
-    ExpectNotNull(name = X509_NAME_new());
-    ExpectIntEQ(X509_NAME_add_entry_by_txt(name, "countryName", MBSTRING_UTF8,
-                                       (byte*)"US", 2, -1, 0), SSL_SUCCESS);
-    ExpectIntEQ(X509_NAME_add_entry_by_txt(name, "commonName", MBSTRING_UTF8,
-                             (byte*)"wolfssl.com", 11, -1, 0), SSL_SUCCESS);
-    ExpectIntEQ(X509_NAME_add_entry_by_txt(name, "emailAddress", MBSTRING_UTF8,
-                     (byte*)"support@wolfssl.com", 19, -1, 0), SSL_SUCCESS);
-
-    /* Get private and public keys */
-    ExpectNotNull(priv = wolfSSL_d2i_PrivateKey(EVP_PKEY_RSA, NULL, &rsaPriv,
-                                                                  clientKeySz));
-    ExpectNotNull(pub = wolfSSL_d2i_PUBKEY(NULL, &rsaPub, clientPubKeySz));
-    ExpectNotNull(x509 = X509_new());
-    ExpectIntEQ(X509_sign(x509, priv, EVP_sha256()), 0);
-    /* Set version 3 */
-    ExpectIntNE(X509_set_version(x509, 2L), 0);
-    /* Set subject name, add pubkey, and sign certificate */
-    ExpectIntEQ(X509_set_subject_name(x509, name), SSL_SUCCESS);
-    X509_NAME_free(name);
-    name = NULL;
-    ExpectIntEQ(X509_set_pubkey(x509, pub), SSL_SUCCESS);
-#ifdef WOLFSSL_ALT_NAMES
-    ExpectNull(wolfSSL_X509_get_next_altname(NULL));
-    ExpectNull(wolfSSL_X509_get_next_altname(x509));
-
-    /* Add some subject alt names */
-    ExpectIntNE(wolfSSL_X509_add_altname(NULL,
-                "ipsum", ASN_DNS_TYPE), SSL_SUCCESS);
-    ExpectIntEQ(wolfSSL_X509_add_altname(x509,
-                NULL, ASN_DNS_TYPE), SSL_SUCCESS);
-    ExpectIntEQ(wolfSSL_X509_add_altname(x509,
-                "sphygmomanometer",
-                ASN_DNS_TYPE), SSL_SUCCESS);
-    ExpectIntEQ(wolfSSL_X509_add_altname(x509,
-                "supercalifragilisticexpialidocious",
-                ASN_DNS_TYPE), SSL_SUCCESS);
-    ExpectIntEQ(wolfSSL_X509_add_altname(x509,
-                "Llanfairpwllgwyngyllgogerychwyrndrobwllllantysiliogogogoch",
-                ASN_DNS_TYPE), SSL_SUCCESS);
-#ifdef WOLFSSL_IP_ALT_NAME
-    {
-        unsigned char ip4_type[] = {127,128,0,255};
-        unsigned char ip6_type[] = {0xdd, 0xcc, 0xba, 0xab,
-                                    0xff, 0xee, 0x99, 0x88,
-                                    0x77, 0x66, 0x55, 0x44,
-                                    0x00, 0x33, 0x22, 0x11};
-        ExpectIntEQ(wolfSSL_X509_add_altname_ex(x509, (char*)ip4_type,
-                sizeof(ip4_type), ASN_IP_TYPE), SSL_SUCCESS);
-        ExpectIntEQ(wolfSSL_X509_add_altname_ex(x509, (char*)ip6_type,
-                sizeof(ip6_type), ASN_IP_TYPE), SSL_SUCCESS);
-    }
-#endif
-
-    {
-        int i;
-
-        if (x509 != NULL) {
-            x509->altNamesNext = x509->altNames;
-        }
-#ifdef WOLFSSL_IP_ALT_NAME
-        /* No names in IP address. */
-        ExpectNull(wolfSSL_X509_get_next_altname(x509));
-        ExpectNull(wolfSSL_X509_get_next_altname(x509));
-#endif
-        for (i = 0; i < 3; i++) {
-            ExpectNotNull(wolfSSL_X509_get_next_altname(x509));
-        }
-        ExpectNull(wolfSSL_X509_get_next_altname(x509));
-#ifdef WOLFSSL_MULTICIRCULATE_ALTNAMELIST
-        ExpectNotNull(wolfSSL_X509_get_next_altname(x509));
-#endif
-    }
-#endif /* WOLFSSL_ALT_NAMES */
-
-    {
-        ASN1_UTCTIME* infinite_past = NULL;
-        ExpectNotNull(infinite_past = ASN1_UTCTIME_set(NULL, 0));
-        ExpectIntEQ(X509_set1_notBefore(x509, infinite_past), 1);
-        ASN1_UTCTIME_free(infinite_past);
-    }
-
-    /* test valid sign case */
-    ExpectIntGT(ret = X509_sign(x509, priv, EVP_sha256()), 0);
-    /* test getting signature */
-#ifndef NO_WOLFSSL_STUB
-    wolfSSL_X509_get0_signature(&sig, &alg, x509);
-#endif
-    ExpectIntEQ(wolfSSL_X509_get_signature(x509, NULL, &sigSz),
-        WOLFSSL_SUCCESS);
-    ExpectIntGT(sigSz, 0);
-    ExpectIntEQ(wolfSSL_X509_get_signature(NULL, NULL, NULL),
-        WOLFSSL_FATAL_ERROR);
-    ExpectIntEQ(wolfSSL_X509_get_signature(x509, NULL, NULL),
-        WOLFSSL_FATAL_ERROR);
-    ExpectIntEQ(wolfSSL_X509_get_signature(NULL, NULL, &sigSz),
-        WOLFSSL_FATAL_ERROR);
-    sigSz = 0;
-    ExpectIntEQ(wolfSSL_X509_get_signature(x509, sn, &sigSz),
-        WOLFSSL_FATAL_ERROR);
-
-    /* test valid X509_sign_ctx case */
-    ExpectNotNull(mctx = EVP_MD_CTX_new());
-    ExpectIntEQ(EVP_DigestSignInit(mctx, NULL, EVP_sha256(), NULL, priv), 1);
-    ExpectIntGT(X509_sign_ctx(x509, mctx), 0);
-
-#if defined(OPENSSL_ALL) && defined(WOLFSSL_ALT_NAMES)
-    ExpectIntEQ(X509_get_ext_count(x509), 1);
-#endif
-#if defined(WOLFSSL_ALT_NAMES) && defined(WOLFSSL_IP_ALT_NAME)
-    ExpectIntEQ(wolfSSL_X509_check_ip_asc(x509, "127.128.0.255", 0), 1);
-    ExpectIntEQ(wolfSSL_X509_check_ip_asc(x509, "DDCC:BAAB:FFEE:9988:7766:5544:0033:2211", 0), 1);
-#endif
-
-    ExpectIntEQ(wolfSSL_X509_get_serial_number(x509, sn, &snSz),
-                WOLFSSL_SUCCESS);
-    DEBUG_WRITE_CERT_X509(x509, "signed.pem");
-
-    /* Variation in size depends on ASN.1 encoding when MSB is set.
-     * WOLFSSL_ASN_TEMPLATE code does not generate a serial number
-     * with the MSB set. See GenerateInteger in asn.c */
-#ifndef USE_CERT_BUFFERS_1024
-#ifndef WOLFSSL_ALT_NAMES
-    /* Valid case - size should be 781-786 with 16 byte serial number */
-    ExpectTrue((781 + snSz <= ret) && (ret <= 781 + 5 + snSz));
-#elif defined(WOLFSSL_IP_ALT_NAME)
-    /* Valid case - size should be 955-960 with 16 byte serial number */
-    ExpectTrue((939 + snSz <= ret) && (ret <= 939 + 5 + snSz));
-#else
-    /* Valid case - size should be 926-931 with 16 byte serial number */
-    ExpectTrue((910 + snSz <= ret) && (ret <= 910 + 5 + snSz));
-#endif
-#else
-#ifndef WOLFSSL_ALT_NAMES
-    /* Valid case - size should be 537-542 with 16 byte serial number */
-    ExpectTrue((521 + snSz <= ret) && (ret <= 521 + 5 + snSz));
-#elif defined(OPENSSL_ALL) || defined(WOLFSSL_IP_ALT_NAME)
-    /* Valid case - size should be 695-670 with 16 byte serial number */
-    ExpectTrue((679 + snSz <= ret) && (ret <= 679 + 5 + snSz));
-#else
-    /* Valid case - size should be 666-671 with 16 byte serial number */
-    ExpectTrue((650 + snSz <= ret) && (ret <= 650 + 5 + snSz));
-#endif
-#endif
-    /* check that issuer name is as expected after signature */
-    InitDecodedCert(&dCert, certIssuer, (word32)certIssuerSz, 0);
-    ExpectIntEQ(ParseCert(&dCert, CERT_TYPE, NO_VERIFY, NULL), 0);
-
-    ExpectNotNull(emptyName = X509_NAME_new());
-    ExpectNotNull(ca = d2i_X509(NULL, &certIssuer, (int)certIssuerSz));
-    ExpectIntEQ(wolfSSL_X509_get_isCA(NULL), 0);
-    ExpectIntEQ(wolfSSL_X509_get_isCA(ca), 1);
-    ExpectNotNull(name = X509_get_subject_name(ca));
-    ExpectIntEQ(X509_NAME_get_sz(NULL), WOLFSSL_FATAL_ERROR);
-    ExpectIntGT(cnSz = X509_NAME_get_sz(name), 0);
-    ExpectNotNull(cn = (char*)XMALLOC(cnSz, HEAP_HINT, DYNAMIC_TYPE_OPENSSL));
-    ExpectNull(X509_NAME_oneline(NULL, cn, (int)cnSz));
-    ExpectPtrEq(X509_NAME_oneline(name, cn, 0), cn);
-    ExpectPtrEq(X509_NAME_oneline(emptyName, cn, (int)cnSz), cn);
-    ExpectNull(X509_NAME_oneline(emptyName, NULL, 0));
-    ExpectPtrEq(X509_NAME_oneline(name, cn, (int)cnSz), cn);
-    ExpectIntEQ(0, XSTRNCMP(cn, dCert.subject, XSTRLEN(cn)));
-    XFREE(cn, HEAP_HINT, DYNAMIC_TYPE_OPENSSL);
-    cn = NULL;
-
-#if defined(XSNPRINTF)
-    ExpectNull(wolfSSL_X509_get_name_oneline(NULL, NULL, 0));
-    ExpectNotNull(cn = wolfSSL_X509_get_name_oneline(name, NULL, 0));
-    ExpectIntGT((int)(cnSz = (word32)XSTRLEN(cn) + 1), 0);
-    ExpectPtrEq(wolfSSL_X509_get_name_oneline(name, cn, (int)cnSz), cn);
-    ExpectNull(wolfSSL_X509_get_name_oneline(NULL, cn, (int)cnSz));
-    ExpectNull(wolfSSL_X509_get_name_oneline(name, cn, cnSz - 1));
-    ExpectPtrEq(wolfSSL_X509_get_name_oneline(name, cn, (int)cnSz), cn);
-    ExpectPtrEq(wolfSSL_X509_get_name_oneline(emptyName, cn, (int)cnSz), cn);
-    XFREE(cn, HEAP_HINT, DYNAMIC_TYPE_OPENSSL);
-    cn = NULL;
-#endif
-    X509_NAME_free(emptyName);
-
-#ifdef WOLFSSL_MULTI_ATTRIB
-    /* test adding multiple OU's to the signer */
-    ExpectNotNull(name = X509_get_subject_name(ca));
-    ExpectIntEQ(X509_NAME_add_entry_by_txt(name, "OU", MBSTRING_UTF8,
-                                       (byte*)"OU1", 3, -1, 0), SSL_SUCCESS);
-    ExpectIntEQ(X509_NAME_add_entry_by_txt(name, "OU", MBSTRING_UTF8,
-                                       (byte*)"OU2", 3, -1, 0), SSL_SUCCESS);
-    ExpectIntGT(X509_sign(ca, priv, EVP_sha256()), 0);
-#endif
-
-    ExpectNotNull(name = X509_get_subject_name(ca));
-    ExpectIntEQ(X509_set_issuer_name(x509, name), SSL_SUCCESS);
-
-    ExpectIntGT(X509_sign(x509, priv, EVP_sha256()), 0);
-    ExpectNotNull(name = X509_get_issuer_name(x509));
-    cnSz = X509_NAME_get_sz(name);
-    ExpectNotNull(cn = (char*)XMALLOC(cnSz, HEAP_HINT, DYNAMIC_TYPE_OPENSSL));
-    ExpectNotNull(cn = X509_NAME_oneline(name, cn, (int)cnSz));
-    /* compare and don't include the multi-attrib "/OU=OU1/OU=OU2" above */
-    ExpectIntEQ(0, XSTRNCMP(cn, dCert.issuer, XSTRLEN(dCert.issuer)));
-    XFREE(cn, HEAP_HINT, DYNAMIC_TYPE_OPENSSL);
-    cn = NULL;
-
-    FreeDecodedCert(&dCert);
-
-    /* Test invalid parameters */
-    ExpectIntEQ(X509_sign(NULL, priv, EVP_sha256()), 0);
-    ExpectIntEQ(X509_sign(x509, NULL, EVP_sha256()), 0);
-    ExpectIntEQ(X509_sign(x509, priv, NULL), 0);
-
-    ExpectIntEQ(X509_sign_ctx(NULL, mctx), 0);
-    EVP_MD_CTX_free(mctx);
-    mctx = NULL;
-    ExpectNotNull(mctx = EVP_MD_CTX_new());
-    ExpectIntEQ(X509_sign_ctx(x509, mctx), 0);
-    ExpectIntEQ(X509_sign_ctx(x509, NULL), 0);
-
-    /* test invalid version number */
-#if defined(OPENSSL_ALL)
-    ExpectIntNE(X509_set_version(x509, 6L), 0);
-    ExpectIntGT(X509_sign(x509, priv, EVP_sha256()), 0);
-
-    /* uses ParseCert which fails on bad version number */
-    ExpectIntEQ(X509_get_ext_count(x509), WC_NO_ERR_TRACE(WOLFSSL_FAILURE));
-#endif
-
-    EVP_MD_CTX_free(mctx);
-    EVP_PKEY_free(priv);
-    EVP_PKEY_free(pub);
-    X509_free(x509);
-    X509_free(ca);
-#endif
-    return EXPECT_RESULT();
-}
-
 static int test_wolfSSL_X509_ALGOR_get0(void)
 {
     EXPECT_DECLS;
@@ -22053,139 +19219,6 @@ static int test_wolfSSL_X509_ALGOR_get0(void)
     return EXPECT_RESULT();
 }
 
-
-static int test_wolfSSL_X509_VERIFY_PARAM(void)
-{
-    EXPECT_DECLS;
-#if defined(OPENSSL_EXTRA)
-    X509_VERIFY_PARAM *paramTo = NULL;
-    X509_VERIFY_PARAM *paramFrom = NULL;
-    char testIPv4[] = "127.0.0.1";
-    char testIPv6[] = "0001:0000:0000:0000:0000:0000:0000:0000/32";
-    char testhostName1[] = "foo.hoge.com";
-    char testhostName2[] = "foobar.hoge.com";
-
-    ExpectNotNull(paramTo = X509_VERIFY_PARAM_new());
-    ExpectNotNull(XMEMSET(paramTo, 0, sizeof(X509_VERIFY_PARAM)));
-
-    ExpectNotNull(paramFrom = X509_VERIFY_PARAM_new());
-    ExpectNotNull(XMEMSET(paramFrom, 0, sizeof(X509_VERIFY_PARAM)));
-
-    ExpectIntEQ(X509_VERIFY_PARAM_set1_host(paramFrom, testhostName1,
-        (int)XSTRLEN(testhostName1)), 1);
-    ExpectIntEQ(0, XSTRNCMP(paramFrom->hostName, testhostName1,
-        (int)XSTRLEN(testhostName1)));
-
-    X509_VERIFY_PARAM_set_hostflags(NULL, 0x00);
-
-    X509_VERIFY_PARAM_set_hostflags(paramFrom, 0x01);
-    ExpectIntEQ(0x01, paramFrom->hostFlags);
-
-    ExpectIntEQ(X509_VERIFY_PARAM_set1_ip_asc(NULL, testIPv4), 0);
-
-    ExpectIntEQ(X509_VERIFY_PARAM_set1_ip_asc(paramFrom, testIPv4), 1);
-    ExpectIntEQ(0, XSTRNCMP(paramFrom->ipasc, testIPv4, WOLFSSL_MAX_IPSTR));
-
-    ExpectIntEQ(X509_VERIFY_PARAM_set1_ip_asc(paramFrom, NULL), 1);
-
-    ExpectIntEQ(X509_VERIFY_PARAM_set1_ip_asc(paramFrom, testIPv6), 1);
-    ExpectIntEQ(0, XSTRNCMP(paramFrom->ipasc, testIPv6, WOLFSSL_MAX_IPSTR));
-
-    /* null pointer */
-    ExpectIntEQ(X509_VERIFY_PARAM_set1(NULL, paramFrom), 0);
-    /* in the case of "from" null, returns success */
-    ExpectIntEQ(X509_VERIFY_PARAM_set1(paramTo, NULL), 1);
-
-    ExpectIntEQ(X509_VERIFY_PARAM_set1(NULL, NULL), 0);
-
-    /* inherit flags test : VPARAM_DEFAULT */
-    ExpectIntEQ(X509_VERIFY_PARAM_set1(paramTo, paramFrom), 1);
-    ExpectIntEQ(0, XSTRNCMP(paramTo->hostName, testhostName1,
-                                    (int)XSTRLEN(testhostName1)));
-    ExpectIntEQ(0x01, paramTo->hostFlags);
-    ExpectIntEQ(0, XSTRNCMP(paramTo->ipasc, testIPv6, WOLFSSL_MAX_IPSTR));
-
-    /* inherit flags test : VPARAM OVERWRITE */
-    ExpectIntEQ(X509_VERIFY_PARAM_set1_host(paramTo, testhostName2,
-        (int)XSTRLEN(testhostName2)), 1);
-    ExpectIntEQ(X509_VERIFY_PARAM_set1_ip_asc(paramTo, testIPv4), 1);
-    X509_VERIFY_PARAM_set_hostflags(paramTo, 0x00);
-
-    if (paramTo != NULL) {
-        paramTo->inherit_flags = X509_VP_FLAG_OVERWRITE;
-    }
-
-    ExpectIntEQ(X509_VERIFY_PARAM_set1(paramTo, paramFrom), 1);
-    ExpectIntEQ(0, XSTRNCMP(paramTo->hostName, testhostName1,
-        (int)XSTRLEN(testhostName1)));
-    ExpectIntEQ(0x01, paramTo->hostFlags);
-    ExpectIntEQ(0, XSTRNCMP(paramTo->ipasc, testIPv6, WOLFSSL_MAX_IPSTR));
-
-    /* inherit flags test : VPARAM_RESET_FLAGS */
-    ExpectIntEQ(X509_VERIFY_PARAM_set1_host(paramTo, testhostName2,
-        (int)XSTRLEN(testhostName2)), 1);
-    ExpectIntEQ(X509_VERIFY_PARAM_set1_ip_asc(paramTo, testIPv4), 1);
-    X509_VERIFY_PARAM_set_hostflags(paramTo, 0x10);
-
-    if (paramTo != NULL) {
-        paramTo->inherit_flags = X509_VP_FLAG_RESET_FLAGS;
-    }
-
-    ExpectIntEQ(X509_VERIFY_PARAM_set1(paramTo, paramFrom), 1);
-    ExpectIntEQ(0, XSTRNCMP(paramTo->hostName, testhostName1,
-                                        (int)XSTRLEN(testhostName1)));
-    ExpectIntEQ(0x01, paramTo->hostFlags);
-    ExpectIntEQ(0, XSTRNCMP(paramTo->ipasc, testIPv6, WOLFSSL_MAX_IPSTR));
-
-    /* inherit flags test : VPARAM_LOCKED */
-    ExpectIntEQ(X509_VERIFY_PARAM_set1_host(paramTo, testhostName2,
-        (int)XSTRLEN(testhostName2)), 1);
-    ExpectIntEQ(X509_VERIFY_PARAM_set1_ip_asc(paramTo, testIPv4), 1);
-    X509_VERIFY_PARAM_set_hostflags(paramTo, 0x00);
-
-    if (paramTo != NULL) {
-        paramTo->inherit_flags = X509_VP_FLAG_LOCKED;
-    }
-
-    ExpectIntEQ(X509_VERIFY_PARAM_set1(paramTo, paramFrom), 1);
-    ExpectIntEQ(0, XSTRNCMP(paramTo->hostName, testhostName2,
-                                    (int)XSTRLEN(testhostName2)));
-    ExpectIntEQ(0x00, paramTo->hostFlags);
-    ExpectIntEQ(0, XSTRNCMP(paramTo->ipasc, testIPv4, WOLFSSL_MAX_IPSTR));
-
-    /* test for incorrect parameters */
-    ExpectIntEQ(X509_VERIFY_PARAM_set_flags(NULL, X509_V_FLAG_CRL_CHECK_ALL),
-        0);
-
-    ExpectIntEQ(X509_VERIFY_PARAM_set_flags(NULL, 0), 0);
-
-    /* inherit flags test : VPARAM_ONCE, not testable yet */
-
-    ExpectIntEQ(X509_VERIFY_PARAM_set_flags(paramTo, X509_V_FLAG_CRL_CHECK_ALL),
-        1);
-
-    ExpectIntEQ(X509_VERIFY_PARAM_get_flags(NULL), 0);
-    ExpectIntEQ(X509_VERIFY_PARAM_get_flags(paramTo),
-        X509_V_FLAG_CRL_CHECK_ALL);
-
-    ExpectIntEQ(X509_VERIFY_PARAM_clear_flags(NULL, X509_V_FLAG_CRL_CHECK_ALL),
-        WOLFSSL_FAILURE);
-    ExpectIntEQ(X509_VERIFY_PARAM_clear_flags(paramTo,
-        X509_V_FLAG_CRL_CHECK_ALL), 1);
-
-    ExpectIntEQ(X509_VERIFY_PARAM_get_flags(paramTo), 0);
-
-    ExpectNull(wolfSSL_X509_VERIFY_PARAM_lookup(NULL));
-    ExpectNull(wolfSSL_X509_VERIFY_PARAM_lookup(""));
-    ExpectNotNull(wolfSSL_X509_VERIFY_PARAM_lookup("ssl_client"));
-    ExpectNotNull(wolfSSL_X509_VERIFY_PARAM_lookup("ssl_server"));
-
-    X509_VERIFY_PARAM_free(paramTo);
-    X509_VERIFY_PARAM_free(paramFrom);
-    X509_VERIFY_PARAM_free(NULL); /* to confirm NULL parameter gives no harm */
-#endif
-    return EXPECT_RESULT();
-}
 
 #if defined(OPENSSL_EXTRA) && defined(HAVE_SSL_MEMIO_TESTS_DEPENDENCIES) && \
     !defined(WOLFSSL_HOSTNAME_VERIFY_ALT_NAME_ONLY)
@@ -23596,162 +20629,6 @@ static int test_wolfSSL_OBJ_txt2obj(void)
         ASN1_OBJECT_free(obj);
         obj = NULL;
     }
-#endif
-    return EXPECT_RESULT();
-}
-
-static int test_wolfSSL_PEM_write_bio_X509(void)
-{
-    EXPECT_DECLS;
-#if defined(OPENSSL_EXTRA) && defined(OPENSSL_ALL) && \
-    defined(WOLFSSL_AKID_NAME) && defined(WOLFSSL_CERT_EXT) && \
-    defined(WOLFSSL_CERT_GEN) && !defined(NO_BIO) && !defined(NO_RSA) && \
-    !defined(NO_FILESYSTEM)
-    /* This test contains the hard coded expected
-     * lengths. Update if necessary */
-    XFILE fp = XBADFILE;
-    WOLFSSL_EVP_PKEY *priv = NULL;
-
-    BIO* input = NULL;
-    BIO* output = NULL;
-    X509* x509a = NULL;
-    X509* x509b = NULL;
-    X509* empty = NULL;
-
-    ASN1_TIME* notBeforeA = NULL;
-    ASN1_TIME* notAfterA  = NULL;
-#ifndef NO_ASN_TIME
-    ASN1_TIME* notBeforeB = NULL;
-    ASN1_TIME* notAfterB  = NULL;
-#endif
-    int expectedLen;
-
-    ExpectTrue((fp = XFOPEN("certs/server-key.pem", "rb")) != XBADFILE);
-    ExpectNotNull(priv = wolfSSL_PEM_read_PrivateKey(fp, NULL, NULL, NULL));
-    if (fp != XBADFILE) {
-        XFCLOSE(fp);
-        fp = XBADFILE;
-    }
-
-    ExpectNotNull(input = BIO_new_file("certs/test/cert-ext-multiple.pem",
-        "rb"));
-    ExpectIntEQ(wolfSSL_BIO_get_len(input), 2000);
-
-    /* read PEM into X509 struct, get notBefore / notAfter to verify against */
-    ExpectNotNull(PEM_read_bio_X509(input, &x509a, NULL, NULL));
-    ExpectNotNull(notBeforeA = X509_get_notBefore(x509a));
-    ExpectNotNull(notAfterA = X509_get_notAfter(x509a));
-
-    /* write X509 back to PEM BIO; no need to sign as nothing changed. */
-    ExpectNotNull(output = BIO_new(wolfSSL_BIO_s_mem()));
-    ExpectNotNull(empty = wolfSSL_X509_new());
-    ExpectIntEQ(PEM_write_bio_X509(NULL, NULL), WOLFSSL_FAILURE);
-    ExpectIntEQ(PEM_write_bio_X509(output, NULL), WOLFSSL_FAILURE);
-    ExpectIntEQ(PEM_write_bio_X509(NULL, x509a), WOLFSSL_FAILURE);
-    ExpectIntEQ(PEM_write_bio_X509(output, empty), WOLFSSL_FAILURE);
-    ExpectIntEQ(PEM_write_bio_X509(output, x509a), WOLFSSL_SUCCESS);
-    /* compare length against expected */
-    expectedLen = 2000;
-    ExpectIntEQ(wolfSSL_BIO_get_len(output), expectedLen);
-    wolfSSL_X509_free(empty);
-
-#ifndef NO_ASN_TIME
-    /* read exported X509 PEM back into struct, sanity check on export,
-     * make sure notBefore/notAfter are the same and certs are identical. */
-    ExpectNotNull(PEM_read_bio_X509(output, &x509b, NULL, NULL));
-    ExpectNotNull(notBeforeB = X509_get_notBefore(x509b));
-    ExpectNotNull(notAfterB = X509_get_notAfter(x509b));
-    ExpectIntEQ(ASN1_TIME_compare(notBeforeA, notBeforeB), 0);
-    ExpectIntEQ(ASN1_TIME_compare(notAfterA, notAfterB), 0);
-    ExpectIntEQ(0, wolfSSL_X509_cmp(x509a, x509b));
-    X509_free(x509b);
-    x509b = NULL;
-#endif
-
-    /* Reset output buffer */
-    BIO_free(output);
-    output = NULL;
-    ExpectNotNull(output = BIO_new(wolfSSL_BIO_s_mem()));
-
-    /* Test forcing the AKID to be generated just from KeyIdentifier */
-    if (EXPECT_SUCCESS() && x509a->authKeyIdSrc != NULL) {
-        XMEMMOVE(x509a->authKeyIdSrc, x509a->authKeyId, x509a->authKeyIdSz);
-        x509a->authKeyId = x509a->authKeyIdSrc;
-        x509a->authKeyIdSrc = NULL;
-        x509a->authKeyIdSrcSz = 0;
-    }
-
-    /* Resign to re-generate the der */
-    ExpectIntGT(wolfSSL_X509_sign(x509a, priv, EVP_sha256()), 0);
-
-    ExpectIntEQ(PEM_write_bio_X509(output, x509a), WOLFSSL_SUCCESS);
-
-    /* Check that we generate a smaller output since the AKID will
-     * only contain the KeyIdentifier without any additional
-     * information */
-
-    /* Here we copy the validity struct from the original */
-    expectedLen = 1688;
-    ExpectIntEQ(wolfSSL_BIO_get_len(output), expectedLen);
-
-    /* Reset buffers and x509 */
-    BIO_free(input);
-    input = NULL;
-    BIO_free(output);
-    output = NULL;
-    X509_free(x509a);
-    x509a = NULL;
-
-    /* test CA and basicConstSet values are encoded when
-     * the cert is a CA */
-    ExpectNotNull(input = BIO_new_file("certs/server-cert.pem", "rb"));
-
-    /* read PEM into X509 struct */
-    ExpectNotNull(PEM_read_bio_X509(input, &x509a, NULL, NULL));
-
-    /* write X509 back to PEM BIO; no need to sign as nothing changed */
-    ExpectNotNull(output = BIO_new(wolfSSL_BIO_s_mem()));
-    ExpectIntEQ(PEM_write_bio_X509(output, x509a), WOLFSSL_SUCCESS);
-
-    /* read exported X509 PEM back into struct, ensure isCa and basicConstSet
-     * values are maintained and certs are identical.*/
-    ExpectNotNull(PEM_read_bio_X509(output, &x509b, NULL, NULL));
-    ExpectIntEQ(x509b->isCa, 1);
-    ExpectIntEQ(x509b->basicConstSet, 1);
-    ExpectIntEQ(0, wolfSSL_X509_cmp(x509a, x509b));
-
-    X509_free(x509a);
-    x509a = NULL;
-    X509_free(x509b);
-    x509b = NULL;
-    BIO_free(input);
-    input = NULL;
-    BIO_free(output);
-    output = NULL;
-
-    /* test CA and basicConstSet values are encoded when
-     * the cert is not CA */
-    ExpectNotNull(input = BIO_new_file("certs/client-uri-cert.pem", "rb"));
-
-    /* read PEM into X509 struct */
-    ExpectNotNull(PEM_read_bio_X509(input, &x509a, NULL, NULL));
-
-    /* write X509 back to PEM BIO; no need to sign as nothing changed */
-    ExpectNotNull(output = BIO_new(wolfSSL_BIO_s_mem()));
-    ExpectIntEQ(PEM_write_bio_X509(output, x509a), WOLFSSL_SUCCESS);
-
-    /* read exported X509 PEM back into struct, ensure isCa and
-     * basicConstSet values are maintained and certs are identical */
-    ExpectNotNull(PEM_read_bio_X509(output, &x509b, NULL, NULL));
-    ExpectIntEQ(x509b->isCa, 0);
-    ExpectIntEQ(x509b->basicConstSet, 1);
-    ExpectIntEQ(0, wolfSSL_X509_cmp(x509a, x509b));
-
-    wolfSSL_EVP_PKEY_free(priv);
-    X509_free(x509a);
-    X509_free(x509b);
-    BIO_free(input);
-    BIO_free(output);
 #endif
     return EXPECT_RESULT();
 }
@@ -26068,66 +22945,6 @@ static int test_wolfSSL_EVP_Cipher_extra(void)
     return EXPECT_RESULT();
 }
 
-static int test_wolfSSL_X509_ext_get_critical_by_NID(void)
-{
-    EXPECT_DECLS;
-#if defined(OPENSSL_EXTRA) && !defined(NO_CERTS)
-    WOLFSSL_X509* x509 = NULL;
-
-    ExpectNotNull(x509 = wolfSSL_X509_new());
-    ExpectIntEQ(wolfSSL_X509_ext_get_critical_by_NID(NULL,
-        WC_NID_basic_constraints), 0);
-    ExpectIntEQ(wolfSSL_X509_ext_get_critical_by_NID(x509,
-        WC_NID_basic_constraints), 0);
-    ExpectIntEQ(wolfSSL_X509_ext_get_critical_by_NID(x509,
-        WC_NID_subject_alt_name), 0);
-    ExpectIntEQ(wolfSSL_X509_ext_get_critical_by_NID(x509,
-        WC_NID_authority_key_identifier), 0);
-    ExpectIntEQ(wolfSSL_X509_ext_get_critical_by_NID(x509,
-        WC_NID_subject_key_identifier), 0);
-    ExpectIntEQ(wolfSSL_X509_ext_get_critical_by_NID(x509,
-        WC_NID_key_usage), 0);
-    ExpectIntEQ(wolfSSL_X509_ext_get_critical_by_NID(x509,
-        WC_NID_crl_distribution_points), 0);
-    ExpectIntEQ(wolfSSL_X509_ext_get_critical_by_NID(x509,
-        WC_NID_ext_key_usage), 0);
-#ifdef WOLFSSL_SEP
-    ExpectIntEQ(wolfSSL_X509_ext_get_critical_by_NID(x509,
-        WC_NID_certificate_policies), 0);
-#endif
-    ExpectIntEQ(wolfSSL_X509_ext_get_critical_by_NID(x509,
-        WC_NID_info_access), 0);
-    wolfSSL_X509_free(x509);
-#endif
-    return EXPECT_RESULT();
-}
-
-static int test_wolfSSL_X509_CRL_distribution_points(void)
-{
-    EXPECT_DECLS;
-#if defined(OPENSSL_EXTRA) && !defined(NO_CERTS) && !defined(NO_RSA) && \
-    !defined(NO_FILESYSTEM)
-    WOLFSSL_X509* x509 = NULL;
-    const char* file = "./certs/client-crl-dist.pem";
-
-    ExpectIntEQ(wolfSSL_X509_ext_isSet_by_NID(NULL,
-        WC_NID_crl_distribution_points), 0);
-
-    ExpectNotNull(x509 = wolfSSL_X509_new());
-    ExpectIntEQ(wolfSSL_X509_ext_isSet_by_NID(x509,
-        WC_NID_crl_distribution_points), 0);
-    wolfSSL_X509_free(x509);
-    x509 = NULL;
-
-    ExpectNotNull(x509 = wolfSSL_X509_load_certificate_file(file,
-         WOLFSSL_FILETYPE_PEM));
-    ExpectIntEQ(wolfSSL_X509_ext_isSet_by_NID(x509,
-        WC_NID_crl_distribution_points), 1);
-    wolfSSL_X509_free(x509);
-#endif
-    return EXPECT_RESULT();
-}
-
 static int test_wolfSSL_X509_SEP(void)
 {
     EXPECT_DECLS;
@@ -26247,207 +23064,6 @@ static int test_wolfSSL_OPENSSL_hexstr2buf(void)
         }
         OPENSSL_free(returnedBuf);
     }
-#endif
-    return EXPECT_RESULT();
-}
-
-static int test_wolfSSL_X509_CA_num(void)
-{
-    EXPECT_DECLS;
-#if defined(OPENSSL_EXTRA) && !defined(NO_CERTS) && !defined(NO_FILESYSTEM) && \
-    defined(HAVE_ECC) && !defined(NO_RSA)
-    WOLFSSL_X509_STORE *store = NULL;
-    WOLFSSL_X509 *x509_1 = NULL;
-    WOLFSSL_X509 *x509_2 = NULL;
-    int ca_num = 0;
-
-    ExpectNotNull(store = wolfSSL_X509_STORE_new());
-    ExpectNotNull(x509_1 = wolfSSL_X509_load_certificate_file(svrCertFile,
-        WOLFSSL_FILETYPE_PEM));
-    ExpectIntEQ(wolfSSL_X509_STORE_add_cert(store, x509_1), 1);
-    ExpectIntEQ(ca_num = wolfSSL_X509_CA_num(store), 1);
-
-    ExpectNotNull(x509_2 = wolfSSL_X509_load_certificate_file(eccCertFile,
-        WOLFSSL_FILETYPE_PEM));
-    ExpectIntEQ(wolfSSL_X509_STORE_add_cert(store, x509_2), 1);
-    ExpectIntEQ(ca_num = wolfSSL_X509_CA_num(store), 2);
-
-    wolfSSL_X509_free(x509_1);
-    wolfSSL_X509_free(x509_2);
-    wolfSSL_X509_STORE_free(store);
-#endif
-    return EXPECT_RESULT();
-}
-
-static int test_wolfSSL_make_cert(void)
-{
-    EXPECT_DECLS;
-#if !defined(NO_RSA) && !defined(NO_ASN_TIME) && defined(WOLFSSL_CERT_GEN) && \
-    defined(WOLFSSL_CERT_EXT)
-    int      ret = 0;
-    Cert     cert;
-    CertName name;
-    RsaKey   key;
-    WC_RNG   rng;
-    byte     der[FOURK_BUF];
-    word32   idx = 0;
-    const byte mySerial[8] = {1,2,3,4,5,6,7,8};
-
-#ifdef OPENSSL_EXTRA
-    const unsigned char* pt = NULL;
-    int                  certSz = 0;
-    X509*                x509 = NULL;
-    X509_NAME*           x509name = NULL;
-    X509_NAME_ENTRY*     entry = NULL;
-    ASN1_STRING*         entryValue = NULL;
-#endif
-
-    XMEMSET(&name, 0, sizeof(CertName));
-
-    /* set up cert name */
-    XMEMCPY(name.country, "US", sizeof("US"));
-    name.countryEnc = CTC_PRINTABLE;
-    XMEMCPY(name.state, "Oregon", sizeof("Oregon"));
-    name.stateEnc = CTC_UTF8;
-    XMEMCPY(name.locality, "Portland", sizeof("Portland"));
-    name.localityEnc = CTC_UTF8;
-    XMEMCPY(name.sur, "Test", sizeof("Test"));
-    name.surEnc = CTC_UTF8;
-    XMEMCPY(name.org, "wolfSSL", sizeof("wolfSSL"));
-    name.orgEnc = CTC_UTF8;
-    XMEMCPY(name.unit, "Development", sizeof("Development"));
-    name.unitEnc = CTC_UTF8;
-    XMEMCPY(name.commonName, "www.wolfssl.com", sizeof("www.wolfssl.com"));
-    name.commonNameEnc = CTC_UTF8;
-    XMEMCPY(name.serialDev, "wolfSSL12345", sizeof("wolfSSL12345"));
-    name.serialDevEnc = CTC_PRINTABLE;
-    XMEMCPY(name.userId, "TestUserID", sizeof("TestUserID"));
-    name.userIdEnc = CTC_PRINTABLE;
-#ifdef WOLFSSL_MULTI_ATTRIB
-    #if CTC_MAX_ATTRIB > 2
-    {
-        NameAttrib* n;
-        n = &name.name[0];
-        n->id   = ASN_DOMAIN_COMPONENT;
-        n->type = CTC_UTF8;
-        n->sz   = sizeof("com");
-        XMEMCPY(n->value, "com", sizeof("com"));
-
-        n = &name.name[1];
-        n->id   = ASN_DOMAIN_COMPONENT;
-        n->type = CTC_UTF8;
-        n->sz   = sizeof("wolfssl");
-        XMEMCPY(n->value, "wolfssl", sizeof("wolfssl"));
-    }
-    #endif
-#endif /* WOLFSSL_MULTI_ATTRIB */
-
-    ExpectIntEQ(wc_InitRsaKey(&key, HEAP_HINT), 0);
-#ifndef HAVE_FIPS
-    ExpectIntEQ(wc_InitRng_ex(&rng, HEAP_HINT, testDevId), 0);
-#else
-    ExpectIntEQ(wc_InitRng(&rng), 0);
-#endif
-
-    /* load test RSA key */
-    idx = 0;
-#if defined(USE_CERT_BUFFERS_1024)
-    ExpectIntEQ(wc_RsaPrivateKeyDecode(server_key_der_1024, &idx, &key,
-                sizeof_server_key_der_1024), 0);
-#elif defined(USE_CERT_BUFFERS_2048)
-    ExpectIntEQ(wc_RsaPrivateKeyDecode(server_key_der_2048, &idx, &key,
-                sizeof_server_key_der_2048), 0);
-#else
-    /* error case, no RSA key loaded, happens later */
-    (void)idx;
-#endif
-
-    XMEMSET(&cert, 0 , sizeof(Cert));
-    ExpectIntEQ(wc_InitCert(&cert), 0);
-
-    XMEMCPY(&cert.subject, &name, sizeof(CertName));
-    XMEMCPY(cert.serial, mySerial, sizeof(mySerial));
-    cert.serialSz = (int)sizeof(mySerial);
-    cert.isCA     = 1;
-#ifndef NO_SHA256
-    cert.sigType = CTC_SHA256wRSA;
-#else
-    cert.sigType = CTC_SHAwRSA;
-#endif
-
-    /* add SKID from the Public Key */
-    ExpectIntEQ(wc_SetSubjectKeyIdFromPublicKey(&cert, &key, NULL), 0);
-
-    /* add AKID from the Public Key */
-    ExpectIntEQ(wc_SetAuthKeyIdFromPublicKey(&cert, &key, NULL), 0);
-
-    ret = 0;
-    do {
-#if defined(WOLFSSL_ASYNC_CRYPT)
-        ret = wc_AsyncWait(ret, &key.asyncDev, WC_ASYNC_FLAG_CALL_AGAIN);
-#endif
-        if (ret >= 0) {
-            ret = wc_MakeSelfCert(&cert, der, FOURK_BUF, &key, &rng);
-        }
-    } while (ret == WC_NO_ERR_TRACE(WC_PENDING_E));
-    ExpectIntGT(ret, 0);
-
-#ifdef OPENSSL_EXTRA
-    /* der holds a certificate with DC's now check X509 parsing of it */
-    certSz = ret;
-    pt = der;
-    ExpectNotNull(x509 = d2i_X509(NULL, &pt, certSz));
-    ExpectNotNull(x509name = X509_get_subject_name(x509));
-#ifdef WOLFSSL_MULTI_ATTRIB
-    ExpectIntEQ((idx = X509_NAME_get_index_by_NID(x509name, NID_domainComponent,
-                    -1)), 5);
-    ExpectIntEQ((idx = X509_NAME_get_index_by_NID(x509name, NID_domainComponent,
-                    (int)idx)), 6);
-    ExpectIntEQ((idx = X509_NAME_get_index_by_NID(x509name, NID_domainComponent,
-                    (int)idx)), -1);
-#endif /* WOLFSSL_MULTI_ATTRIB */
-
-    /* compare DN at index 0 */
-    ExpectNotNull(entry = X509_NAME_get_entry(x509name, 0));
-    ExpectNotNull(entryValue = X509_NAME_ENTRY_get_data(entry));
-    ExpectIntEQ(ASN1_STRING_length(entryValue), 2);
-    ExpectStrEQ((const char*)ASN1_STRING_data(entryValue), "US");
-
-#ifndef WOLFSSL_MULTI_ATTRIB
-    /* compare Serial Number */
-    ExpectIntEQ((idx = X509_NAME_get_index_by_NID(x509name, NID_serialNumber,
-                    -1)), 7);
-    ExpectNotNull(entry = X509_NAME_get_entry(x509name, idx));
-    ExpectNotNull(entryValue = X509_NAME_ENTRY_get_data(entry));
-    ExpectIntEQ(ASN1_STRING_length(entryValue), XSTRLEN("wolfSSL12345"));
-    ExpectStrEQ((const char*)ASN1_STRING_data(entryValue), "wolfSSL12345");
-#endif
-
-#ifdef WOLFSSL_MULTI_ATTRIB
-    /* get first and second DC and compare result */
-    ExpectIntEQ((idx = X509_NAME_get_index_by_NID(x509name, NID_domainComponent,
-                    -1)), 5);
-    ExpectNotNull(entry = X509_NAME_get_entry(x509name, (int)idx));
-    ExpectNotNull(entryValue = X509_NAME_ENTRY_get_data(entry));
-    ExpectStrEQ((const char *)ASN1_STRING_data(entryValue), "com");
-
-    ExpectIntEQ((idx = X509_NAME_get_index_by_NID(x509name, NID_domainComponent,
-                   (int)idx)), 6);
-    ExpectNotNull(entry = X509_NAME_get_entry(x509name, (int)idx));
-    ExpectNotNull(entryValue = X509_NAME_ENTRY_get_data(entry));
-    ExpectStrEQ((const char *)ASN1_STRING_data(entryValue), "wolfssl");
-#endif /* WOLFSSL_MULTI_ATTRIB */
-
-    ExpectNull(X509_NAME_get_entry(NULL, 0));
-    /* try invalid index locations for regression test and sanity check */
-    ExpectNull(X509_NAME_get_entry(x509name, 11));
-    ExpectNull(X509_NAME_get_entry(x509name, 20));
-
-    X509_free(x509);
-#endif /* OPENSSL_EXTRA */
-
-    wc_FreeRsaKey(&key);
-    wc_FreeRng(&rng);
 #endif
     return EXPECT_RESULT();
 }
@@ -32214,197 +28830,6 @@ static int test_wolfSSL_SMIME_write_PKCS7(void)
 #endif /* HAVE_SMIME */
 #endif /* !NO_BIO */
 
-/* Test of X509 store use outside of SSL context w/ CRL lookup (ALWAYS
- * returns 0) */
-static int test_X509_STORE_No_SSL_CTX(void)
-{
-    EXPECT_DECLS;
-#if defined(OPENSSL_ALL) && defined(WOLFSSL_CERT_GEN) && \
-    (defined(WOLFSSL_CERT_REQ) || defined(WOLFSSL_CERT_EXT)) && \
-    !defined(NO_FILESYSTEM) && !defined(NO_WOLFSSL_DIR)  && \
-    (defined(OPENSSL_EXTRA) || defined(WOLFSSL_WPAS_SMALL)) && \
-    defined(HAVE_CRL) && !defined(NO_RSA)
-
-    X509_STORE *     store = NULL;
-    X509_STORE_CTX * storeCtx = NULL;
-    X509_CRL *       crl = NULL;
-    X509 *           ca = NULL;
-    X509 *           cert = NULL;
-    const char       cliCrlPem[] = "./certs/crl/cliCrl.pem";
-    const char       srvCert[] = "./certs/server-cert.pem";
-    const char       caCert[] = "./certs/ca-cert.pem";
-    const char       caDir[] = "./certs/crl/hash_pem";
-    XFILE            fp = XBADFILE;
-    X509_LOOKUP *    lookup = NULL;
-
-    ExpectNotNull(store = (X509_STORE *)X509_STORE_new());
-
-    /* Set up store with CA */
-    ExpectNotNull((ca = wolfSSL_X509_load_certificate_file(caCert,
-        SSL_FILETYPE_PEM)));
-    ExpectIntEQ(X509_STORE_add_cert(store, ca), SSL_SUCCESS);
-
-    /* Add CRL lookup directory to store
-     * NOTE: test uses ./certs/crl/hash_pem/0fdb2da4.r0, which is a copy
-     * of crl.pem */
-    ExpectNotNull((lookup = X509_STORE_add_lookup(store,
-        X509_LOOKUP_hash_dir())));
-    ExpectIntEQ(X509_LOOKUP_ctrl(lookup, X509_L_ADD_DIR, caDir,
-        X509_FILETYPE_PEM, NULL), SSL_SUCCESS);
-
-    ExpectIntEQ(X509_STORE_set_flags(store, X509_V_FLAG_CRL_CHECK),
-        SSL_SUCCESS);
-
-    /* Add CRL to store NOT containing the verified certificate, which
-     * forces use of the CRL lookup directory */
-    ExpectTrue((fp = XFOPEN(cliCrlPem, "rb")) != XBADFILE);
-    ExpectNotNull(crl = (X509_CRL *)PEM_read_X509_CRL(fp, (X509_CRL **)NULL,
-        NULL, NULL));
-    if (fp != XBADFILE)
-        XFCLOSE(fp);
-    ExpectIntEQ(X509_STORE_add_crl(store, crl), SSL_SUCCESS);
-
-    /* Create verification context outside of an SSL session */
-    ExpectNotNull((storeCtx = X509_STORE_CTX_new()));
-    ExpectNotNull((cert = wolfSSL_X509_load_certificate_file(srvCert,
-        SSL_FILETYPE_PEM)));
-    ExpectIntEQ(X509_STORE_CTX_init(storeCtx, store, cert, NULL), SSL_SUCCESS);
-
-    /* Perform verification, which should NOT indicate CRL missing due to the
-     * store CM's X509 store pointer being NULL */
-    ExpectIntNE(X509_verify_cert(storeCtx), WC_NO_ERR_TRACE(CRL_MISSING));
-
-    X509_CRL_free(crl);
-    X509_STORE_free(store);
-    X509_STORE_CTX_free(storeCtx);
-    X509_free(cert);
-    X509_free(ca);
-#endif
-    return EXPECT_RESULT();
-}
-
-/* Test of X509 store use outside of SSL context w/ CRL lookup, but
- * with X509_LOOKUP_add_dir and X509_FILETYPE_ASN1. */
-static int test_X509_LOOKUP_add_dir(void)
-{
-    EXPECT_DECLS;
-#if defined(OPENSSL_ALL) && defined(WOLFSSL_CERT_GEN) && \
-    (defined(WOLFSSL_CERT_REQ) || defined(WOLFSSL_CERT_EXT)) && \
-    !defined(NO_FILESYSTEM) && !defined(NO_WOLFSSL_DIR)  && \
-    (defined(OPENSSL_EXTRA) || defined(WOLFSSL_WPAS_SMALL)) && \
-    defined(HAVE_CRL) && !defined(NO_RSA)
-
-    X509_STORE *     store = NULL;
-    X509_STORE_CTX * storeCtx = NULL;
-    X509_CRL *       crl = NULL;
-    X509 *           ca = NULL;
-    X509 *           cert = NULL;
-    const char       cliCrlPem[] = "./certs/crl/cliCrl.pem";
-    const char       srvCert[] = "./certs/server-cert.pem";
-    const char       caCert[] = "./certs/ca-cert.pem";
-    const char       caDir[] = "./certs/crl/hash_der";
-    XFILE            fp = XBADFILE;
-    X509_LOOKUP *    lookup = NULL;
-
-    ExpectNotNull(store = (X509_STORE *)X509_STORE_new());
-
-    /* Set up store with CA */
-    ExpectNotNull((ca = wolfSSL_X509_load_certificate_file(caCert,
-        SSL_FILETYPE_PEM)));
-    ExpectIntEQ(X509_STORE_add_cert(store, ca), SSL_SUCCESS);
-
-    /* Add CRL lookup directory to store.
-     * Test uses ./certs/crl/hash_der/0fdb2da4.r0, which is a copy
-     * of crl.der */
-    ExpectNotNull((lookup = X509_STORE_add_lookup(store,
-        X509_LOOKUP_hash_dir())));
-
-    ExpectIntEQ(X509_LOOKUP_add_dir(lookup, caDir, X509_FILETYPE_ASN1),
-        SSL_SUCCESS);
-
-    ExpectIntEQ(X509_STORE_set_flags(store, X509_V_FLAG_CRL_CHECK),
-        SSL_SUCCESS);
-
-    /* Add CRL to store NOT containing the verified certificate, which
-     * forces use of the CRL lookup directory */
-    ExpectTrue((fp = XFOPEN(cliCrlPem, "rb")) != XBADFILE);
-    ExpectNotNull(crl = (X509_CRL *)PEM_read_X509_CRL(fp, (X509_CRL **)NULL,
-        NULL, NULL));
-    if (fp != XBADFILE) {
-        XFCLOSE(fp);
-        fp = XBADFILE;
-    }
-    ExpectIntEQ(X509_STORE_add_crl(store, crl), SSL_SUCCESS);
-
-    /* Create verification context outside of an SSL session */
-    ExpectNotNull((storeCtx = X509_STORE_CTX_new()));
-    ExpectNotNull((cert = wolfSSL_X509_load_certificate_file(srvCert,
-        SSL_FILETYPE_PEM)));
-    ExpectIntEQ(X509_STORE_CTX_init(storeCtx, store, cert, NULL), SSL_SUCCESS);
-
-    /* Perform verification, which should NOT return CRL missing */
-    ExpectIntNE(X509_verify_cert(storeCtx), WC_NO_ERR_TRACE(CRL_MISSING));
-
-    X509_CRL_free(crl);
-    crl = NULL;
-    X509_STORE_free(store);
-    store = NULL;
-    X509_STORE_CTX_free(storeCtx);
-    storeCtx = NULL;
-    X509_free(cert);
-    cert = NULL;
-    X509_free(ca);
-    ca = NULL;
-
-    /* Now repeat the same, but look for X509_FILETYPE_PEM.
-     * We should get CRL_MISSING at the end, because the lookup
-     * dir has only ASN1 CRLs. */
-
-    ExpectNotNull(store = (X509_STORE *)X509_STORE_new());
-
-    ExpectNotNull((ca = wolfSSL_X509_load_certificate_file(caCert,
-        SSL_FILETYPE_PEM)));
-    ExpectIntEQ(X509_STORE_add_cert(store, ca), SSL_SUCCESS);
-
-    ExpectNotNull((lookup = X509_STORE_add_lookup(store,
-        X509_LOOKUP_hash_dir())));
-
-    ExpectIntEQ(X509_LOOKUP_add_dir(lookup, caDir, X509_FILETYPE_PEM),
-        SSL_SUCCESS);
-
-    ExpectIntEQ(X509_STORE_set_flags(store, X509_V_FLAG_CRL_CHECK),
-        SSL_SUCCESS);
-
-    ExpectTrue((fp = XFOPEN(cliCrlPem, "rb")) != XBADFILE);
-    ExpectNotNull(crl = (X509_CRL *)PEM_read_X509_CRL(fp, (X509_CRL **)NULL,
-        NULL, NULL));
-    if (fp != XBADFILE) {
-        XFCLOSE(fp);
-        fp = XBADFILE;
-    }
-    ExpectIntEQ(X509_STORE_add_crl(store, crl), SSL_SUCCESS);
-
-    ExpectNotNull((storeCtx = X509_STORE_CTX_new()));
-    ExpectNotNull((cert = wolfSSL_X509_load_certificate_file(srvCert,
-        SSL_FILETYPE_PEM)));
-    ExpectIntEQ(X509_STORE_CTX_init(storeCtx, store, cert, NULL), SSL_SUCCESS);
-
-    /* Now we SHOULD get CRL_MISSING, because we looked for PEM
-     * in dir containing only ASN1/DER. */
-    ExpectIntEQ(X509_verify_cert(storeCtx), WC_NO_ERR_TRACE(WOLFSSL_FAILURE));
-    ExpectIntEQ(X509_STORE_CTX_get_error(storeCtx),
-            X509_V_ERR_UNABLE_TO_GET_CRL);
-
-    X509_CRL_free(crl);
-    X509_STORE_free(store);
-    X509_STORE_CTX_free(storeCtx);
-    X509_free(cert);
-    X509_free(ca);
-#endif
-    return EXPECT_RESULT();
-}
-
-
 
 /*----------------------------------------------------------------------------*
  | Certificate Failure Checks
@@ -32784,146 +29209,6 @@ static int test_wolfSSL_X509_CRL(void)
     return EXPECT_RESULT();
 }
 
-static int test_wolfSSL_X509_load_crl_file(void)
-{
-    EXPECT_DECLS;
-#if defined(OPENSSL_EXTRA) && defined(HAVE_CRL) && !defined(NO_FILESYSTEM) && \
-    !defined(NO_STDIO_FILESYSTEM) && !defined(NO_RSA) && !defined(NO_BIO) && \
-    !defined(WOLFSSL_CRL_ALLOW_MISSING_CDP)
-    int i;
-    char pem[][100] = {
-        "./certs/crl/crl.pem",
-        "./certs/crl/crl2.pem",
-        "./certs/crl/caEccCrl.pem",
-        "./certs/crl/eccCliCRL.pem",
-        "./certs/crl/eccSrvCRL.pem",
-    #ifdef WC_RSA_PSS
-        "./certs/crl/crl_rsapss.pem",
-    #endif
-        ""
-    };
-    char der[][100] = {
-        "./certs/crl/crl.der",
-        "./certs/crl/crl2.der",
-        ""
-    };
-    WOLFSSL_X509_STORE*  store = NULL;
-    WOLFSSL_X509_LOOKUP* lookup = NULL;
-
-    ExpectNotNull(store = wolfSSL_X509_STORE_new());
-    ExpectNotNull(lookup = X509_STORE_add_lookup(store, X509_LOOKUP_file()));
-
-    ExpectIntEQ(X509_LOOKUP_load_file(lookup, "certs/ca-cert.pem",
-        X509_FILETYPE_PEM), 1);
-#ifdef WC_RSA_PSS
-    ExpectIntEQ(X509_LOOKUP_load_file(lookup, "certs/rsapss/ca-rsapss.pem",
-        X509_FILETYPE_PEM), 1);
-#endif
-    ExpectIntEQ(X509_LOOKUP_load_file(lookup, "certs/server-revoked-cert.pem",
-        X509_FILETYPE_PEM), 1);
-    if (store) {
-        ExpectIntEQ(wolfSSL_CertManagerVerify(store->cm, svrCertFile,
-            WOLFSSL_FILETYPE_PEM), 1);
-        /* since store hasn't yet known the revoked cert*/
-        ExpectIntEQ(wolfSSL_CertManagerVerify(store->cm,
-            "certs/server-revoked-cert.pem", WOLFSSL_FILETYPE_PEM), 1);
-    }
-
-    ExpectIntEQ(X509_load_crl_file(lookup, pem[0], 0), 0);
-    for (i = 0; pem[i][0] != '\0'; i++) {
-        ExpectIntEQ(X509_load_crl_file(lookup, pem[i], WOLFSSL_FILETYPE_PEM),
-            1);
-    }
-
-    if (store) {
-        /* since store knows crl list */
-        ExpectIntEQ(wolfSSL_CertManagerVerify(store->cm,
-            "certs/server-revoked-cert.pem", WOLFSSL_FILETYPE_PEM),
-            WC_NO_ERR_TRACE(CRL_CERT_REVOKED));
-#ifdef WC_RSA_PSS
-        ExpectIntEQ(wolfSSL_CertManagerVerify(store->cm,
-            "certs/rsapss/server-rsapss-cert.pem", WOLFSSL_FILETYPE_PEM),
-            WC_NO_ERR_TRACE(ASN_NO_SIGNER_E));
-#endif
-    }
-    /* once feeing store */
-    X509_STORE_free(store);
-    store = NULL;
-
-    ExpectNotNull(store = wolfSSL_X509_STORE_new());
-    ExpectNotNull(lookup = X509_STORE_add_lookup(store, X509_LOOKUP_file()));
-
-    ExpectIntEQ(X509_LOOKUP_load_file(lookup, "certs/ca-cert.pem",
-        X509_FILETYPE_PEM), 1);
-    ExpectIntEQ(X509_LOOKUP_load_file(lookup, "certs/server-revoked-cert.pem",
-        X509_FILETYPE_PEM), 1);
-    if (store) {
-        ExpectIntEQ(wolfSSL_CertManagerVerify(store->cm, svrCertFile,
-            WOLFSSL_FILETYPE_PEM), 1);
-        /* since store hasn't yet known the revoked cert*/
-        ExpectIntEQ(wolfSSL_CertManagerVerify(store->cm,
-            "certs/server-revoked-cert.pem", WOLFSSL_FILETYPE_PEM), 1);
-    }
-
-    for (i = 0; der[i][0] != '\0'; i++) {
-        ExpectIntEQ(X509_load_crl_file(lookup, der[i], WOLFSSL_FILETYPE_ASN1),
-            1);
-    }
-
-    if (store) {
-        /* since store knows crl list */
-        ExpectIntEQ(wolfSSL_CertManagerVerify(store->cm,
-            "certs/server-revoked-cert.pem", WOLFSSL_FILETYPE_PEM),
-            WC_NO_ERR_TRACE(CRL_CERT_REVOKED));
-    }
-
-    /* test for incorrect parameter */
-    ExpectIntEQ(X509_load_crl_file(NULL, pem[0], 0), 0);
-    ExpectIntEQ(X509_load_crl_file(lookup, NULL, 0), 0);
-    ExpectIntEQ(X509_load_crl_file(NULL, NULL, 0), 0);
-
-    X509_STORE_free(store);
-    store = NULL;
-#endif
-    return EXPECT_RESULT();
-}
-
-static int test_wolfSSL_i2d_X509(void)
-{
-    EXPECT_DECLS;
-#if defined(OPENSSL_EXTRA) && defined(USE_CERT_BUFFERS_2048) && !defined(NO_RSA)
-    const unsigned char* cert_buf = server_cert_der_2048;
-    unsigned char* out = NULL;
-    unsigned char* tmp = NULL;
-    const unsigned char* nullPtr = NULL;
-    const unsigned char notCert[2] = { 0x30, 0x00 };
-    const unsigned char* notCertPtr = notCert;
-    X509* cert = NULL;
-
-    ExpectNull(d2i_X509(NULL, NULL, sizeof_server_cert_der_2048));
-    ExpectNull(d2i_X509(NULL, &nullPtr, sizeof_server_cert_der_2048));
-    ExpectNull(d2i_X509(NULL, &cert_buf, 0));
-    ExpectNull(d2i_X509(NULL, &notCertPtr, sizeof(notCert)));
-    ExpectNotNull(d2i_X509(&cert, &cert_buf, sizeof_server_cert_der_2048));
-    /* Pointer should be advanced */
-    ExpectPtrGT(cert_buf, server_cert_der_2048);
-    ExpectIntGT(i2d_X509(cert, &out), 0);
-    ExpectNotNull(out);
-    tmp = out;
-    ExpectIntGT(i2d_X509(cert, &tmp), 0);
-    ExpectPtrGT(tmp, out);
-#if defined(WOLFSSL_CERT_GEN) && !defined(NO_BIO) && !defined(NO_FILESYSTEM)
-    ExpectIntEQ(wolfSSL_PEM_write_X509(XBADFILE, NULL), 0);
-    ExpectIntEQ(wolfSSL_PEM_write_X509(XBADFILE, cert), 0);
-    ExpectIntEQ(wolfSSL_PEM_write_X509(stderr, cert), 1);
-#endif
-
-    XFREE(out, NULL, DYNAMIC_TYPE_OPENSSL);
-    X509_free(cert);
-#endif
-    return EXPECT_RESULT();
-}
-
 static int test_wolfSSL_d2i_X509_REQ(void)
 {
     EXPECT_DECLS;
@@ -33111,23 +29396,6 @@ static int test_wolfSSL_d2i_X509_REQ(void)
     }
 #endif /* !NO_DSA && !HAVE_SELFTEST */
 #endif /* WOLFSSL_CERT_REQ && (OPENSSL_ALL || OPENSSL_EXTRA) */
-    return EXPECT_RESULT();
-}
-
-static int test_wolfSSL_PEM_read_X509(void)
-{
-    EXPECT_DECLS;
-#if defined(OPENSSL_EXTRA) && defined(HAVE_CRL) && !defined(NO_FILESYSTEM) && \
-    !defined(NO_RSA)
-    X509 *x509 = NULL;
-    XFILE fp = XBADFILE;
-
-    ExpectTrue((fp = XFOPEN(svrCertFile, "rb")) != XBADFILE);
-    ExpectNotNull(x509 = (X509 *)PEM_read_X509(fp, (X509 **)NULL, NULL, NULL));
-    X509_free(x509);
-    if (fp != XBADFILE)
-        XFCLOSE(fp);
-#endif
     return EXPECT_RESULT();
 }
 
@@ -34721,301 +30989,6 @@ static int test_wolfSSL_EVP_PKEY_hkdf(void)
 #endif /* OPENSSL_EXTRA && HAVE_HKDF */
     return EXPECT_RESULT();
 }
-
-#ifndef NO_BIO
-static int test_wolfSSL_PEM_X509_INFO_read_bio(void)
-{
-    EXPECT_DECLS;
-#if defined(OPENSSL_ALL) && !defined(NO_FILESYSTEM) && !defined(NO_RSA)
-    BIO* bio = NULL;
-    X509_INFO* info = NULL;
-    STACK_OF(X509_INFO)* sk = NULL;
-    STACK_OF(X509_INFO)* sk2 = NULL;
-    char* subject = NULL;
-    char exp1[] = "/C=US/ST=Montana/L=Bozeman/O=Sawtooth/OU=Consulting/"
-                  "CN=www.wolfssl.com/emailAddress=info@wolfssl.com";
-    char exp2[] = "/C=US/ST=Montana/L=Bozeman/O=wolfSSL/OU=Support/"
-                  "CN=www.wolfssl.com/emailAddress=info@wolfssl.com";
-
-    ExpectNotNull(bio = BIO_new(BIO_s_file()));
-    ExpectIntGT(BIO_read_filename(bio, svrCertFile), 0);
-    ExpectNotNull(sk = PEM_X509_INFO_read_bio(bio, NULL, NULL, NULL));
-    ExpectIntEQ(sk_X509_INFO_num(sk), 2);
-
-    /* using dereference to maintain testing for Apache port*/
-    ExpectNull(sk_X509_INFO_pop(NULL));
-    ExpectNotNull(info = sk_X509_INFO_pop(sk));
-    ExpectNotNull(subject = X509_NAME_oneline(X509_get_subject_name(info->x509),
-        0, 0));
-
-    ExpectIntEQ(0, XSTRNCMP(subject, exp1, sizeof(exp1)));
-    XFREE(subject, 0, DYNAMIC_TYPE_OPENSSL);
-    subject = NULL;
-    X509_INFO_free(info);
-    info = NULL;
-
-    ExpectNotNull(info = sk_X509_INFO_pop(sk));
-    ExpectNotNull(subject = X509_NAME_oneline(X509_get_subject_name(info->x509),
-        0, 0));
-
-    ExpectIntEQ(0, XSTRNCMP(subject, exp2, sizeof(exp2)));
-    XFREE(subject, 0, DYNAMIC_TYPE_OPENSSL);
-    subject = NULL;
-    X509_INFO_free(info);
-    ExpectNull(info = sk_X509_INFO_pop(sk));
-
-    sk_X509_INFO_pop_free(sk, X509_INFO_free);
-    sk = NULL;
-    BIO_free(bio);
-    bio = NULL;
-
-    ExpectNotNull(sk = wolfSSL_sk_X509_INFO_new_null());
-    ExpectNotNull(bio = BIO_new(BIO_s_file()));
-    ExpectIntGT(BIO_read_filename(bio, svrCertFile), 0);
-    ExpectNotNull(sk2 = PEM_X509_INFO_read_bio(bio, sk, NULL, NULL));
-    ExpectPtrEq(sk, sk2);
-    if (sk2 != sk) {
-        sk_X509_INFO_pop_free(sk, X509_INFO_free);
-    }
-    sk = NULL;
-    BIO_free(bio);
-    sk_X509_INFO_pop_free(sk2, X509_INFO_free);
-
-    ExpectNotNull(sk = wolfSSL_sk_X509_INFO_new_null());
-    sk_X509_INFO_free(sk);
-#endif
-    return EXPECT_RESULT();
-}
-
-static int test_wolfSSL_PEM_X509_INFO_read(void)
-{
-    EXPECT_DECLS;
-#if defined(OPENSSL_ALL) && !defined(NO_FILESYSTEM) && !defined(NO_RSA)
-    XFILE fp = XBADFILE;
-    STACK_OF(X509_INFO)* sk = NULL;
-
-    ExpectTrue((fp = XFOPEN(svrCertFile, "rb")) != XBADFILE);
-    ExpectNull(wolfSSL_PEM_X509_INFO_read(XBADFILE, NULL, NULL, NULL));
-    ExpectNotNull(sk = wolfSSL_PEM_X509_INFO_read(fp, NULL, NULL, NULL));
-
-    sk_X509_INFO_pop_free(sk, X509_INFO_free);
-    if (fp != XBADFILE)
-        XFCLOSE(fp);
-#endif
-    return EXPECT_RESULT();
-}
-#endif /* !NO_BIO */
-
-static int test_wolfSSL_X509_STORE_get1_certs(void)
-{
-    EXPECT_DECLS;
-#if defined(OPENSSL_EXTRA) && defined(WOLFSSL_SIGNER_DER_CERT) && \
-    !defined(NO_FILESYSTEM) && !defined(NO_RSA)
-    X509_STORE_CTX *storeCtx = NULL;
-    X509_STORE *store = NULL;
-    X509 *caX509 = NULL;
-    X509 *svrX509 = NULL;
-    X509_NAME *subject = NULL;
-    WOLF_STACK_OF(WOLFSSL_X509) *certs = NULL;
-
-    ExpectNotNull(caX509 = X509_load_certificate_file(caCertFile,
-        SSL_FILETYPE_PEM));
-    ExpectNotNull((svrX509 = wolfSSL_X509_load_certificate_file(svrCertFile,
-        SSL_FILETYPE_PEM)));
-    ExpectNotNull(storeCtx = X509_STORE_CTX_new());
-    ExpectNotNull(store = X509_STORE_new());
-    ExpectNotNull(subject = X509_get_subject_name(caX509));
-
-    /* Errors */
-    ExpectNull(X509_STORE_get1_certs(storeCtx, subject));
-    ExpectNull(X509_STORE_get1_certs(NULL, subject));
-    ExpectNull(X509_STORE_get1_certs(storeCtx, NULL));
-
-    ExpectIntEQ(X509_STORE_add_cert(store, caX509), SSL_SUCCESS);
-    ExpectIntEQ(X509_STORE_CTX_init(storeCtx, store, caX509, NULL),
-        SSL_SUCCESS);
-
-    /* Should find the cert */
-    ExpectNotNull(certs = X509_STORE_get1_certs(storeCtx, subject));
-    ExpectIntEQ(1, wolfSSL_sk_X509_num(certs));
-
-    sk_X509_pop_free(certs, NULL);
-    certs = NULL;
-
-    /* Should not find the cert */
-    ExpectNotNull(subject = X509_get_subject_name(svrX509));
-    ExpectNotNull(certs = X509_STORE_get1_certs(storeCtx, subject));
-    ExpectIntEQ(0, wolfSSL_sk_X509_num(certs));
-
-    sk_X509_pop_free(certs, NULL);
-    certs = NULL;
-
-    X509_STORE_free(store);
-    X509_STORE_CTX_free(storeCtx);
-    X509_free(svrX509);
-    X509_free(caX509);
-#endif /* OPENSSL_EXTRA && WOLFSSL_SIGNER_DER_CERT && !NO_FILESYSTEM */
-    return EXPECT_RESULT();
-}
-
-#if defined(HAVE_SSL_MEMIO_TESTS_DEPENDENCIES) && \
-    defined(WOLFSSL_LOCAL_X509_STORE) && \
-    (defined(OPENSSL_ALL) || defined(WOLFSSL_QT)) && defined(HAVE_CRL)
-static int test_wolfSSL_X509_STORE_set_get_crl_provider(X509_STORE_CTX* ctx,
-        X509_CRL** crl_out, X509* cert) {
-    X509_CRL *crl = NULL;
-    XFILE fp = XBADFILE;
-    char* cert_issuer = X509_NAME_oneline(X509_get_issuer_name(cert), NULL, 0);
-    int ret = 0;
-
-    (void)ctx;
-
-    if (cert_issuer == NULL)
-        return 0;
-
-    if ((fp = XFOPEN("certs/crl/crl.pem", "rb")) != XBADFILE) {
-        PEM_read_X509_CRL(fp, &crl, NULL, NULL);
-        XFCLOSE(fp);
-        if (crl != NULL) {
-            char* crl_issuer = X509_NAME_oneline(
-                    X509_CRL_get_issuer(crl), NULL, 0);
-            if ((crl_issuer != NULL) &&
-                   (XSTRCMP(cert_issuer, crl_issuer) == 0)) {
-                *crl_out = X509_CRL_dup(crl);
-                if (*crl_out != NULL)
-                    ret = 1;
-            }
-            OPENSSL_free(crl_issuer);
-        }
-    }
-
-    X509_CRL_free(crl);
-    OPENSSL_free(cert_issuer);
-    return ret;
-}
-
-static int test_wolfSSL_X509_STORE_set_get_crl_provider2(X509_STORE_CTX* ctx,
-        X509_CRL** crl_out, X509* cert) {
-    (void)ctx;
-    (void)cert;
-    *crl_out = NULL;
-    return 1;
-}
-
-#ifndef NO_WOLFSSL_STUB
-static int test_wolfSSL_X509_STORE_set_get_crl_check(X509_STORE_CTX* ctx,
-        X509_CRL* crl) {
-    (void)ctx;
-    (void)crl;
-    return 1;
-}
-#endif
-
-static int test_wolfSSL_X509_STORE_set_get_crl_verify(int ok,
-        X509_STORE_CTX* ctx) {
-    int cert_error = X509_STORE_CTX_get_error(ctx);
-    X509_VERIFY_PARAM* param = X509_STORE_CTX_get0_param(ctx);
-    int flags = X509_VERIFY_PARAM_get_flags(param);
-    if ((flags & (X509_V_FLAG_CRL_CHECK | X509_V_FLAG_CRL_CHECK_ALL)) !=
-            (X509_V_FLAG_CRL_CHECK | X509_V_FLAG_CRL_CHECK_ALL)) {
-        /* Make sure the flags are set */
-        return 0;
-    }
-    /* Ignore CRL missing error */
-#ifndef OPENSSL_COMPATIBLE_DEFAULTS
-    if (cert_error == WC_NO_ERR_TRACE(CRL_MISSING))
-#else
-    if (cert_error == X509_V_ERR_UNABLE_TO_GET_CRL)
-#endif
-        return 1;
-    return ok;
-}
-
-static int test_wolfSSL_X509_STORE_set_get_crl_ctx_ready(WOLFSSL_CTX* ctx)
-{
-    EXPECT_DECLS;
-    X509_STORE* cert_store = NULL;
-
-    ExpectIntEQ(wolfSSL_CTX_EnableCRL(ctx, WOLFSSL_CRL_CHECKALL),
-            WOLFSSL_SUCCESS);
-    ExpectNotNull(cert_store = SSL_CTX_get_cert_store(ctx));
-    X509_STORE_set_get_crl(cert_store,
-            test_wolfSSL_X509_STORE_set_get_crl_provider);
-#ifndef NO_WOLFSSL_STUB
-    X509_STORE_set_check_crl(cert_store,
-            test_wolfSSL_X509_STORE_set_get_crl_check);
-#endif
-
-    return EXPECT_RESULT();
-}
-
-static int test_wolfSSL_X509_STORE_set_get_crl_ctx_ready2(WOLFSSL_CTX* ctx)
-{
-    EXPECT_DECLS;
-    X509_STORE* cert_store = NULL;
-    X509_VERIFY_PARAM* param = NULL;
-
-    SSL_CTX_set_verify(ctx, WOLFSSL_VERIFY_PEER, NULL);
-    ExpectIntEQ(wolfSSL_CTX_EnableCRL(ctx, WOLFSSL_CRL_CHECKALL),
-            WOLFSSL_SUCCESS);
-    ExpectNotNull(cert_store = SSL_CTX_get_cert_store(ctx));
-    X509_STORE_set_get_crl(cert_store,
-            test_wolfSSL_X509_STORE_set_get_crl_provider2);
-#ifndef NO_WOLFSSL_STUB
-    X509_STORE_set_check_crl(cert_store,
-            test_wolfSSL_X509_STORE_set_get_crl_check);
-#endif
-    X509_STORE_set_verify_cb(cert_store,
-            test_wolfSSL_X509_STORE_set_get_crl_verify);
-    ExpectNotNull(X509_STORE_get0_param(cert_store));
-    ExpectNotNull(param = X509_VERIFY_PARAM_new());
-    ExpectIntEQ(X509_VERIFY_PARAM_inherit(NULL, NULL) , WOLFSSL_SUCCESS);
-    ExpectIntEQ(X509_VERIFY_PARAM_inherit(param, NULL) , WOLFSSL_SUCCESS);
-    ExpectIntEQ(X509_VERIFY_PARAM_inherit(param,
-            X509_STORE_get0_param(cert_store)), WOLFSSL_SUCCESS);
-    ExpectIntEQ(X509_VERIFY_PARAM_inherit(param,
-            X509_STORE_get0_param(cert_store)), 1);
-    ExpectIntEQ(X509_VERIFY_PARAM_set_flags(
-        param, X509_V_FLAG_CRL_CHECK | X509_V_FLAG_CRL_CHECK_ALL), 1);
-    ExpectIntEQ(X509_STORE_set1_param(cert_store, param), 1);
-    ExpectIntEQ(X509_STORE_set_flags(cert_store,
-            X509_V_FLAG_CRL_CHECK | X509_V_FLAG_CRL_CHECK_ALL), 1);
-
-
-    X509_VERIFY_PARAM_free(param);
-    return EXPECT_RESULT();
-}
-#endif
-
-/* This test mimics the usage of the CRL provider in gRPC */
-static int test_wolfSSL_X509_STORE_set_get_crl(void)
-{
-    EXPECT_DECLS;
-#if defined(HAVE_SSL_MEMIO_TESTS_DEPENDENCIES) && \
-    defined(WOLFSSL_LOCAL_X509_STORE) && \
-    (defined(OPENSSL_ALL) || defined(WOLFSSL_QT)) && defined(HAVE_CRL)
-    test_ssl_cbf func_cb_client;
-    test_ssl_cbf func_cb_server;
-
-    XMEMSET(&func_cb_client, 0, sizeof(func_cb_client));
-    XMEMSET(&func_cb_server, 0, sizeof(func_cb_server));
-
-    func_cb_client.ctx_ready = test_wolfSSL_X509_STORE_set_get_crl_ctx_ready;
-
-    ExpectIntEQ(test_wolfSSL_client_server_nofail_memio(&func_cb_client,
-        &func_cb_server, NULL), TEST_SUCCESS);
-
-    XMEMSET(&func_cb_client, 0, sizeof(func_cb_client));
-    XMEMSET(&func_cb_server, 0, sizeof(func_cb_server));
-
-    func_cb_client.ctx_ready = test_wolfSSL_X509_STORE_set_get_crl_ctx_ready2;
-
-    ExpectIntEQ(test_wolfSSL_client_server_nofail_memio(&func_cb_client,
-        &func_cb_server, NULL), TEST_SUCCESS);
-#endif
-    return EXPECT_RESULT();
-}
-
 
 static int test_wolfSSL_dup_CA_list(void)
 {
@@ -45752,82 +41725,29 @@ TEST_CASE testCases[] = {
     TEST_DECL(test_wolfSSL_OtherName),
     TEST_DECL(test_wolfSSL_FPKI),
     TEST_DECL(test_wolfSSL_URI),
-    TEST_DECL(test_wolfSSL_TBS),
-
-    TEST_DECL(test_wolfSSL_get0_param),
-    TEST_DECL(test_wolfSSL_set1_host),
-
-    TEST_DECL(test_wolfSSL_X509_STORE_CTX),
-    TEST_DECL(test_wolfSSL_X509_STORE_CTX_ex),
-    TEST_DECL(test_X509_STORE_untrusted),
-#if defined(OPENSSL_ALL) && !defined(NO_RSA)
-    TEST_DECL(test_X509_STORE_InvalidCa),
-#endif
-    TEST_DECL(test_wolfSSL_X509_STORE_CTX_trusted_stack_cleanup),
-    TEST_DECL(test_wolfSSL_X509_STORE_CTX_get_issuer),
-    TEST_DECL(test_wolfSSL_X509_STORE_set_flags),
-    TEST_DECL(test_wolfSSL_X509_STORE_CTX_set_time),
-    TEST_DECL(test_wolfSSL_X509_STORE_CTX_get0_store),
-    TEST_DECL(test_wolfSSL_X509_STORE),
-    TEST_DECL(test_wolfSSL_X509_STORE_load_locations),
-    TEST_DECL(test_X509_STORE_get0_objects),
-    TEST_DECL(test_wolfSSL_X509_STORE_get1_certs),
-    TEST_DECL(test_wolfSSL_X509_STORE_set_get_crl),
-    TEST_DECL(test_wolfSSL_X509_CA_num),
-
-    TEST_DECL(test_wolfSSL_X509_LOOKUP_load_file),
-    TEST_DECL(test_wolfSSL_X509_LOOKUP_ctrl_file),
-    TEST_DECL(test_wolfSSL_X509_LOOKUP_ctrl_hash_dir),
-    TEST_DECL(test_wolfSSL_X509_load_crl_file),
-
-    TEST_DECL(test_wolfSSL_X509_VERIFY_PARAM),
-    TEST_DECL(test_wolfSSL_X509_VERIFY_PARAM_set1_ip),
-    TEST_DECL(test_wolfSSL_X509_VERIFY_PARAM_set1_host),
 
     /* X509 tests */
     TEST_OSSL_X509_DECLS,
+    TEST_OSSL_X509_NAME_DECLS,
+    TEST_OSSL_X509_EXT_DECLS,
+    TEST_OSSL_X509_PK_DECLS,
+    TEST_OSSL_X509_VFY_PARAMS_DECLS,
+    TEST_OSSL_X509_IO_DECLS,
+    TEST_OSSL_X509_CRYPTO_DECLS,
+    TEST_OSSL_X509_ACERT_DECLS,
+    TEST_OSSL_X509_INFO_DECLS,
 
     TEST_DECL(test_wolfSSL_X509_ALGOR_get0),
-
-    TEST_OSSL_X509_PK_DECLS,
-
-    TEST_DECL(test_wolfSSL_X509_ext_get_critical_by_NID),
-    TEST_DECL(test_wolfSSL_X509_CRL_distribution_points),
-
-    TEST_DECL(test_wolfSSL_i2d_X509),
-    TEST_DECL(test_wolfSSL_PEM_read_X509),
-    TEST_DECL(test_wolfSSL_PEM_write_bio_X509),
-
-    TEST_DECL(test_wolfSSL_X509_check_private_key),
-    TEST_DECL(test_wolfSSL_X509_verify),
-    TEST_DECL(test_wolfSSL_X509_sign),
-    TEST_DECL(test_wolfSSL_X509_sign2),
-    TEST_DECL(test_wolfSSL_make_cert),
-
     TEST_DECL(test_wolfSSL_X509_SEP),
     TEST_DECL(test_wolfSSL_X509_CRL),
-
-    /* X509 ACERT tests */
-    TEST_DECL(test_wolfSSL_X509_ACERT_verify),
-    TEST_DECL(test_wolfSSL_X509_ACERT_misc_api),
-    TEST_DECL(test_wolfSSL_X509_ACERT_buffer),
-    TEST_DECL(test_wolfSSL_X509_ACERT_new_and_sign),
-    TEST_DECL(test_wolfSSL_X509_ACERT_asn),
-
-#ifndef NO_BIO
-    TEST_DECL(test_wolfSSL_X509_INFO_multiple_info),
-    TEST_DECL(test_wolfSSL_X509_INFO),
-    TEST_DECL(test_wolfSSL_PEM_X509_INFO_read_bio),
-    TEST_DECL(test_wolfSSL_PEM_X509_INFO_read),
-#endif
-
 #ifndef NO_BIO
     TEST_DECL(test_wolfSSL_X509_print),
     TEST_DECL(test_wolfSSL_X509_CRL_print),
 #endif
 
-    TEST_OSSL_X509_NAME_DECLS,
-    TEST_OSSL_X509_EXT_DECLS,
+    /* X509 Store tests */
+    TEST_OSSL_X509_STORE_DECLS,
+    TEST_OSSL_X509_LOOKUP_DECLS,
 
     TEST_DECL(test_GENERAL_NAME_set0_othername),
     TEST_DECL(test_othername_and_SID_ext),
@@ -45841,10 +41761,6 @@ TEST_CASE testCases[] = {
     TEST_DECL(test_wolfSSL_d2i_X509_REQ),
     TEST_DECL(test_X509_REQ),
     TEST_DECL(test_wolfSSL_X509_REQ_print),
-
-    /* OpenSSL compatibility outside SSL context w/ CRL lookup directory */
-    TEST_DECL(test_X509_STORE_No_SSL_CTX),
-    TEST_DECL(test_X509_LOOKUP_add_dir),
 
     /* RAND compatibility API */
     TEST_DECL(test_wolfSSL_RAND_set_rand_method),
@@ -46031,6 +41947,9 @@ TEST_CASE testCases[] = {
 
     TEST_DECL(test_wolfSSL_tmp_dh),
     TEST_DECL(test_wolfSSL_ctrl),
+
+    TEST_DECL(test_wolfSSL_get0_param),
+    TEST_DECL(test_wolfSSL_set1_host),
 
 #if defined(OPENSSL_ALL) || (defined(OPENSSL_EXTRA) && \
     (defined(HAVE_STUNNEL) || defined(WOLFSSL_NGINX) || \
