+++ tests/api/include.am
@@ -83,6 +83,14 @@ tests_unit_test_SOURCES += tests/api/test_ossl_x509.c
 tests_unit_test_SOURCES += tests/api/test_ossl_x509_ext.c
 tests_unit_test_SOURCES += tests/api/test_ossl_x509_name.c
 tests_unit_test_SOURCES += tests/api/test_ossl_x509_pk.c
+tests_unit_test_SOURCES += tests/api/test_ossl_x509_vp.c
+tests_unit_test_SOURCES += tests/api/test_ossl_x509_io.c
+tests_unit_test_SOURCES += tests/api/test_ossl_x509_crypto.c
+tests_unit_test_SOURCES += tests/api/test_ossl_x509_acert.c
+tests_unit_test_SOURCES += tests/api/test_ossl_x509_info.c
+# OpenSSL X509 Store
+tests_unit_test_SOURCES += tests/api/test_ossl_x509_str.c
+tests_unit_test_SOURCES += tests/api/test_ossl_x509_lu.c
 # SSL PEM
 tests_unit_test_SOURCES += tests/api/test_ossl_pem.c
 # TLS 1.3 specific
@@ -158,6 +166,13 @@ EXTRA_DIST += tests/api/test_ossl_x509.h
 EXTRA_DIST += tests/api/test_ossl_x509_ext.h
 EXTRA_DIST += tests/api/test_ossl_x509_name.h
 EXTRA_DIST += tests/api/test_ossl_x509_pk.h
+EXTRA_DIST += tests/api/test_ossl_x509_vp.h
+EXTRA_DIST += tests/api/test_ossl_x509_io.h
+EXTRA_DIST += tests/api/test_ossl_x509_crypto.h
+EXTRA_DIST += tests/api/test_ossl_x509_acert.h
+EXTRA_DIST += tests/api/test_ossl_x509_info.h
+EXTRA_DIST += tests/api/test_ossl_x509_str.h
+EXTRA_DIST += tests/api/test_ossl_x509_lu.h
 EXTRA_DIST += tests/api/test_ossl_pem.h
 EXTRA_DIST += tests/api/test_tls13.h
