+++ tests/api/test_ossl_x509.h
@@ -41,6 +41,9 @@ int test_wolfSSL_X509_set_notAfter(void);
 int test_wolfSSL_X509_set_notBefore(void);
 int test_wolfSSL_X509_set_version(void);
 int test_wolfSSL_X509_get_serialNumber(void);
+int test_wolfSSL_get_tbs(void);
+int test_wolfSSL_X509_ext_get_critical_by_NID(void);
+int test_wolfSSL_X509_CRL_distribution_points(void);
 int test_wolfSSL_X509_check_ip_asc(void);
 int test_wolfSSL_X509_bad_altname(void);
 int test_wolfSSL_X509_name_match1(void);
@@ -52,33 +55,36 @@ int test_wolfSSL_X509_check_ca(void);
 int test_X509_get_signature_nid(void);
 int test_wolfSSL_X509_cmp(void);
 
-#define TEST_OSSL_X509_DECLS                                                \
-    TEST_DECL_GROUP("ossl_x509", test_x509_get_key_id),                     \
-    TEST_DECL_GROUP("ossl_x509", test_wolfSSL_X509_get_version),            \
-    TEST_DECL_GROUP("ossl_x509", test_wolfSSL_X509_cmp_time),               \
-    TEST_DECL_GROUP("ossl_x509", test_wolfSSL_X509_time_adj),               \
-    TEST_DECL_GROUP("ossl_x509", test_wolfSSL_X509_NID),                    \
-    TEST_DECL_GROUP("ossl_x509", test_wolfSSL_i2d_X509_NAME_canon),         \
-    TEST_DECL_GROUP("ossl_x509", test_wolfSSL_X509_subject_name_hash),      \
-    TEST_DECL_GROUP("ossl_x509", test_wolfSSL_X509_issuer_name_hash),       \
-    TEST_DECL_GROUP("ossl_x509", test_wolfSSL_X509_check_host),             \
-    TEST_DECL_GROUP("ossl_x509", test_wolfSSL_X509_check_email),            \
-    TEST_DECL_GROUP("ossl_x509", test_wolfSSL_X509),                        \
-    TEST_DECL_GROUP("ossl_x509", test_wolfSSL_X509_get0_tbs_sigalg),        \
-    TEST_DECL_GROUP("ossl_x509", test_wolfSSL_X509_set_name),               \
-    TEST_DECL_GROUP("ossl_x509", test_wolfSSL_X509_set_notAfter),           \
-    TEST_DECL_GROUP("ossl_x509", test_wolfSSL_X509_set_notBefore),          \
-    TEST_DECL_GROUP("ossl_x509", test_wolfSSL_X509_set_version),            \
-    TEST_DECL_GROUP("ossl_x509", test_wolfSSL_X509_get_serialNumber),       \
-    TEST_DECL_GROUP("ossl_x509", test_wolfSSL_X509_check_ip_asc),           \
-    TEST_DECL_GROUP("ossl_x509", test_wolfSSL_X509_bad_altname),            \
-    TEST_DECL_GROUP("ossl_x509", test_wolfSSL_X509_name_match1),            \
-    TEST_DECL_GROUP("ossl_x509", test_wolfSSL_X509_name_match2),            \
-    TEST_DECL_GROUP("ossl_x509", test_wolfSSL_X509_name_match3),            \
-    TEST_DECL_GROUP("ossl_x509", test_wolfSSL_X509_max_altnames),           \
-    TEST_DECL_GROUP("ossl_x509", test_wolfSSL_X509_max_name_constraints),   \
-    TEST_DECL_GROUP("ossl_x509", test_wolfSSL_X509_check_ca),               \
-    TEST_DECL_GROUP("ossl_x509", test_X509_get_signature_nid),              \
+#define TEST_OSSL_X509_DECLS                                                   \
+    TEST_DECL_GROUP("ossl_x509", test_x509_get_key_id),                        \
+    TEST_DECL_GROUP("ossl_x509", test_wolfSSL_X509_get_version),               \
+    TEST_DECL_GROUP("ossl_x509", test_wolfSSL_X509_cmp_time),                  \
+    TEST_DECL_GROUP("ossl_x509", test_wolfSSL_X509_time_adj),                  \
+    TEST_DECL_GROUP("ossl_x509", test_wolfSSL_X509_NID),                       \
+    TEST_DECL_GROUP("ossl_x509", test_wolfSSL_i2d_X509_NAME_canon),            \
+    TEST_DECL_GROUP("ossl_x509", test_wolfSSL_X509_subject_name_hash),         \
+    TEST_DECL_GROUP("ossl_x509", test_wolfSSL_X509_issuer_name_hash),          \
+    TEST_DECL_GROUP("ossl_x509", test_wolfSSL_X509_check_host),                \
+    TEST_DECL_GROUP("ossl_x509", test_wolfSSL_X509_check_email),               \
+    TEST_DECL_GROUP("ossl_x509", test_wolfSSL_X509),                           \
+    TEST_DECL_GROUP("ossl_x509", test_wolfSSL_X509_get0_tbs_sigalg),           \
+    TEST_DECL_GROUP("ossl_x509", test_wolfSSL_X509_set_name),                  \
+    TEST_DECL_GROUP("ossl_x509", test_wolfSSL_X509_set_notAfter),              \
+    TEST_DECL_GROUP("ossl_x509", test_wolfSSL_X509_set_notBefore),             \
+    TEST_DECL_GROUP("ossl_x509", test_wolfSSL_X509_set_version),               \
+    TEST_DECL_GROUP("ossl_x509", test_wolfSSL_X509_get_serialNumber),          \
+    TEST_DECL_GROUP("ossl_x509", test_wolfSSL_get_tbs),                        \
+    TEST_DECL_GROUP("ossl_x509", test_wolfSSL_X509_ext_get_critical_by_NID),   \
+    TEST_DECL_GROUP("ossl_x509", test_wolfSSL_X509_CRL_distribution_points),   \
+    TEST_DECL_GROUP("ossl_x509", test_wolfSSL_X509_check_ip_asc),              \
+    TEST_DECL_GROUP("ossl_x509", test_wolfSSL_X509_bad_altname),               \
+    TEST_DECL_GROUP("ossl_x509", test_wolfSSL_X509_name_match1),               \
+    TEST_DECL_GROUP("ossl_x509", test_wolfSSL_X509_name_match2),               \
+    TEST_DECL_GROUP("ossl_x509", test_wolfSSL_X509_name_match3),               \
+    TEST_DECL_GROUP("ossl_x509", test_wolfSSL_X509_max_altnames),              \
+    TEST_DECL_GROUP("ossl_x509", test_wolfSSL_X509_max_name_constraints),      \
+    TEST_DECL_GROUP("ossl_x509", test_wolfSSL_X509_check_ca),                  \
+    TEST_DECL_GROUP("ossl_x509", test_X509_get_signature_nid),                 \
     TEST_DECL_GROUP("ossl_x509", test_wolfSSL_X509_cmp)
 
 #endif /* WOLFCRYPT_TEST_OSSL_X509_H */
