+++ tests/api/test_ossl_x509_info.h
@@ -1,0 +1,38 @@
+/* test_ossl_x509_info.h
+ *
+ * Copyright (C) 2006-2025 wolfSSL Inc.
+ *
+ * This file is part of wolfSSL.
+ *
+ * wolfSSL is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * wolfSSL is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1335, USA
+ */
+
+#ifndef WOLFCRYPT_TEST_OSSL_X509_INFO_H
+#define WOLFCRYPT_TEST_OSSL_X509_INFO_H
+
+#include <tests/api/api_decl.h>
+
+int test_wolfSSL_X509_INFO_multiple_info(void);
+int test_wolfSSL_X509_INFO(void);
+int test_wolfSSL_PEM_X509_INFO_read_bio(void);
+int test_wolfSSL_PEM_X509_INFO_read(void);
+
+#define TEST_OSSL_X509_INFO_DECLS                                              \
+    TEST_DECL_GROUP("ossl_x509_info", test_wolfSSL_X509_INFO_multiple_info),   \
+    TEST_DECL_GROUP("ossl_x509_info", test_wolfSSL_X509_INFO),                 \
+    TEST_DECL_GROUP("ossl_x509_info", test_wolfSSL_PEM_X509_INFO_read_bio),    \
+    TEST_DECL_GROUP("ossl_x509_info", test_wolfSSL_PEM_X509_INFO_read)
+
+#endif /* WOLFCRYPT_TEST_OSSL_X509_INFO_H */
